Confidence in congress

conlegis

Source

General Social Survey https://gss.norc.org

Details

Question 165. I am going to name some institutions in this country. As far as the people running these institutions are concerned, would you say you have a great deal of confidence, only some confidence, or hardly any confidence at all in them? l. Congress

Overview

For further details see the official GSS documentation.

Counts by year:

year	iap	a great deal	don't know	hardly any	no answer	only some	skipped on web	Total

1972	1613	-	-	-	-	-	-	1613

1973	-	352	39	223	7	883	-	1504

1974	-	253	45	309	3	874	-	1484

1975	-	198	43	374	3	872	-	1490

1976	-	205	39	381	5	869	-	1499

1977	-	291	44	261	7	927	-	1530

1978	-	197	48	319	5	963	-	1532

1980	-	136	58	493	2	779	-	1468

1982	-	241	43	436	9	1131	-	1860

1983	-	158	47	371	7	1016	-	1599

1984	484	123	19	212	13	622	-	1473

1985	1534	-	-	-	-	-	-	1534

1986	-	238	37	298	5	892	-	1470

1987	-	284	65	333	3	1134	-	1819

1988	484	153	33	193	2	616	-	1481

1989	502	174	34	225	2	600	-	1537

1990	473	137	28	204	2	528	-	1372

1991	500	180	29	259	7	542	-	1517

1993	549	73	25	428	3	528	-	1606

1994	981	154	49	789	13	1006	-	2992

1996	979	146	61	820	5	893	-	2904

1998	921	200	62	571	7	1071	-	2832

2000	921	227	85	532	8	1044	-	2817

2002	1853	117	19	222	1	553	-	2765

2004	1936	115	12	251	1	497	-	2812

2006	2521	209	49	703	2	1026	-	4510

2008	658	126	25	526	2	686	-	2023

2010	667	125	26	587	4	635	-	2044

2012	636	87	39	639	4	569	-	1974

2014	851	93	41	900	2	651	-	2538

2016	911	112	30	1036	1	777	-	2867

2018	785	81	36	745	4	697	-	2348

2021	1360	141	4	1431	-	1089	7	4032

2022	1173	112	26	1315	6	904	8	3544

Total	23292	5438	1240	16386	145	25874	15	72390

Values

1 a great deal

2 only some

3 hardly any

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(q) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook