Courts dealing with criminals

courts

Source

General Social Survey https://gss.norc.org

Details

Question 90. In general, do you think the courts in this area deal too harshly or not harshly enough with criminals?

Overview

For further details see the official GSS documentation.

Counts by year:

year	about right	don't know	no answer	not harshly enough	too harshly	iap	not available in this year	Total

1972	265	173	4	1066	105	-	-	1613

1973	196	138	10	1092	68	-	-	1504

1974	72	51	8	580	42	731	-	1484

1975	144	104	7	1174	61	-	-	1490

1976	148	89	5	1210	47	-	-	1499

1977	123	84	3	1268	52	-	-	1530

1978	111	80	4	1297	40	-	-	1532

1980	112	87	3	1218	48	-	-	1468

1982	73	36	6	760	48	937	-	1860

1983	103	71	4	1363	58	-	-	1599

1984	161	63	13	1192	44	-	-	1473

1985	138	51	9	1284	52	-	-	1534

1986	117	53	3	1252	45	-	-	1470

1987	218	101	7	1417	76	-	-	1819

1988	145	71	4	1204	57	-	-	1481

1989	135	76	-	1285	41	-	-	1537

1990	124	65	5	1131	47	-	-	1372

1991	169	76	9	1202	61	-	-	1517

1993	156	88	8	1300	54	-	-	1606

1994	226	135	19	2534	78	-	-	2992

1996	310	197	16	2246	135	-	-	2904

1998	372	179	11	2095	175	-	-	2832

2000	436	240	19	1913	209	-	-	2817

2002	240	83	4	919	126	1393	-	2765

2004	266	90	3	861	120	1472	-	2812

2006	638	167	6	1902	279	1518	-	4510

2008	388	147	7	1259	222	-	-	2023

2010	341	157	10	1269	267	-	-	2044

2012	380	180	17	1128	269	-	-	1974

2014	484	207	20	1451	376	-	-	2538

2016	460	294	22	1578	513	-	-	2867

2018	357	276	16	1253	446	-	-	2348

2021	-	-	-	-	-	-	4032	4032

2022	298	241	14	933	308	1750	-	3544

Total	7906	4150	296	43636	4569	7801	4032	72390

Values

1 too harshly

2 not harshly enough

3 about right

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(q) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook