How many in family earned money

earnrs

Source

General Social Survey https://gss.norc.org

Details

Question 36. Just thinking about your family now -- those people in the househol who are related to you -- how many persons in the family, including yourself, earned any money last year from any job or employment?

Overview

For further details see the official GSS documentation.

Counts by year:

year	0	1	2	3	4	5	6	no answer	7	eight or more	iap	skipped on web	Total

1972	186	682	506	136	52	14	4	33	-	-	-	-	1613

1973	162	608	506	138	63	11	2	10	3	1	-	-	1504

1974	192	595	477	132	50	15	7	15	1	-	-	-	1484

1975	201	593	483	121	54	25	5	8	-	-	-	-	1490

1976	233	627	449	118	45	17	1	8	-	1	-	-	1499

1977	214	613	494	125	60	18	2	4	-	-	-	-	1530

1978	229	629	479	104	54	22	3	8	2	2	-	-	1532

1980	249	544	483	98	53	11	7	19	3	1	-	-	1468

1982	291	752	577	135	38	10	4	53	-	-	-	-	1860

1983	234	621	544	123	41	16	2	17	1	-	-	-	1599

1984	233	611	441	111	42	13	4	18	-	-	-	-	1473

1985	276	578	499	99	55	11	3	11	2	-	-	-	1534

1986	297	506	478	107	41	17	2	20	2	-	-	-	1470

1987	291	709	585	148	53	15	2	16	-	-	-	-	1819

1988	275	560	452	137	34	9	4	10	-	-	-	-	1481

1989	286	553	475	133	52	20	4	12	1	1	-	-	1537

1990	248	508	456	101	40	7	3	8	-	1	-	-	1372

1991	320	581	459	100	42	4	-	11	-	-	-	-	1517

1993	298	577	556	114	32	6	2	20	1	-	-	-	1606

1994	493	1190	981	189	78	6	-	55	-	-	-	-	2992

1996	432	1213	920	186	70	15	5	63	-	-	-	-	2904

1998	429	1149	949	187	74	13	1	23	2	5	-	-	2832

2000	471	1162	942	177	47	7	2	7	-	2	-	-	2817

2002	466	1097	918	165	69	19	8	19	1	3	-	-	2765

2004	394	1131	996	179	78	13	4	7	3	7	-	-	2812

2006	645	1890	1502	290	121	37	4	21	-	-	-	-	4510

2008	328	791	668	166	44	18	4	3	-	1	-	-	2023

2010	354	825	632	143	55	13	5	14	2	1	-	-	2044

2012	369	770	637	119	56	8	2	12	-	1	-	-	1974

2014	467	1023	797	171	49	13	3	14	1	-	-	-	2538

2016	571	1104	891	180	82	17	1	19	1	1	-	-	2867

2018	423	891	762	181	65	10	1	12	2	1	-	-	2348

2021	832	1421	1216	263	101	39	9	-	7	7	58	79	4032

2022	736	1392	971	215	105	34	10	29	3	6	6	37	3544

Total	12125	28496	23181	5091	1995	523	120	599	38	42	64	116	72390

Values

8 eight or more

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(q) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook