People helpful or looking out for selves

helpful

Source

General Social Survey https://gss.norc.org

Details

Question 161. Would you say that most of the time people try to be helpful, or that they are mostly just looking out for themselves?

Overview

For further details see the official GSS documentation.

Counts by year:

year	depends	don't know	iap	looking out for themselves	no answer	try to be helpful	not available in this year	Total

1972	100	22	1	740	1	749	-	1613

1973	56	4	-	739	3	702	-	1504

1974	-	-	1484	-	-	-	-	1484

1975	99	8	-	545	2	836	-	1490

1976	91	5	-	756	1	646	-	1499

1977	-	-	1530	-	-	-	-	1530

1978	80	4	-	536	5	907	-	1532

1980	64	9	-	678	3	714	-	1468

1982	-	-	1860	-	-	-	-	1860

1983	68	11	-	603	-	917	-	1599

1984	54	3	-	650	3	763	-	1473

1985	-	-	1534	-	-	-	-	1534

1986	76	7	-	565	5	817	-	1470

1987	81	11	-	935	5	787	-	1819

1988	46	6	484	448	3	494	-	1481

1989	52	5	502	459	1	518	-	1537

1990	55	7	473	372	1	464	-	1372

1991	72	8	500	436	5	496	-	1517

1993	60	14	549	431	1	551	-	1606

1994	128	9	981	935	7	932	-	2992

1996	138	10	979	940	4	833	-	2904

1998	176	7	921	815	2	911	-	2832

2000	166	10	921	851	3	866	-	2817

2002	88	2	1853	391	-	431	-	2765

2004	76	3	1936	359	-	438	-	2812

2006	194	4	2521	924	-	867	-	4510

2008	118	2	658	624	-	621	-	2023

2010	135	5	667	588	2	647	-	2044

2012	98	8	636	612	2	618	-	1974

2014	145	5	851	742	-	795	-	2538

2016	137	7	911	892	3	917	-	2867

2018	142	9	785	694	1	717	-	2348

2021	-	-	-	-	-	-	4032	4032

2022	149	9	2325	549	-	512	-	3544

Total	2944	214	25862	18809	63	20466	4032	72390

Values

1 try to be helpful

2 looking out for themselves

3 depends

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(q) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook