Should govt do more or less?

helpnot

Source

General Social Survey https://gss.norc.org

Details

Question 310. Please look at card BD. Some people think that the government in Washington is trying to do too many things that should be left to individuals and private busienss; they are at point 5 on this card. Others disagree and think that the government should do even more to solve our country's problems; they are at point 1. a. Where would you place yourself on this scale, or haven't you made up your mind on this?

Overview

For further details see the official GSS documentation.

Counts by year:

year	iap	2	4	agree with both	don't know	government does too much	government should do more	no answer	skipped on web	Total

1972	1613	-	-	-	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	-	-	-	1504

1974	1484	-	-	-	-	-	-	-	-	1484

1975	-	161	173	433	104	245	371	3	-	1490

1976	1499	-	-	-	-	-	-	-	-	1499

1977	1530	-	-	-	-	-	-	-	-	1530

1978	1532	-	-	-	-	-	-	-	-	1532

1980	1468	-	-	-	-	-	-	-	-	1468

1982	1860	-	-	-	-	-	-	-	-	1860

1983	-	175	280	574	110	258	198	4	-	1599

1984	-	184	241	537	79	204	204	24	-	1473

1985	1534	-	-	-	-	-	-	-	-	1534

1986	-	185	228	604	81	189	180	3	-	1470

1987	-	260	241	657	104	220	328	9	-	1819

1988	484	138	141	392	60	128	137	1	-	1481

1989	502	147	159	390	72	131	134	2	-	1537

1990	473	135	133	364	54	95	118	-	-	1372

1991	500	149	148	395	70	111	139	5	-	1517

1993	549	156	162	401	39	155	138	6	-	1606

1994	981	271	358	709	98	325	235	15	-	2992

1996	979	239	324	722	100	304	229	7	-	2904

1998	921	231	339	749	105	282	202	3	-	2832

2000	921	226	331	709	95	313	213	9	-	2817

2002	1853	93	123	431	27	108	129	1	-	2765

2004	1936	86	123	381	22	126	137	1	-	2812

2006	2521	199	288	859	51	250	339	3	-	4510

2008	658	174	175	539	53	194	226	4	-	2023

2010	667	165	189	498	58	257	205	5	-	2044

2012	636	147	171	530	45	242	199	4	-	1974

2014	851	174	258	624	65	342	223	1	-	2538

2016	913	210	273	754	92	292	329	4	-	2867

2018	785	219	207	644	68	190	231	4	-	2348

2021	1377	386	319	977	10	388	539	-	36	4032

2022	1174	260	255	906	71	335	520	7	16	3544

Total	33705	4770	5639	14779	1733	5684	5903	125	52	72390

Values

1 government should do more

3 agree with both

5 government does too much

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(q) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook