In what month and year was your last hiv test

hivtest1

Source

General Social Survey https://gss.norc.org

Details

Question 1562a. Not including blood donations, in what month and year was your last HIV test?

Values

198001 jan 1980

198002 feb 1980

198003 mar 1980

198004 apr 1980

198005 may 1980

198006 jun 1980

198007 jul 1980

198008 aug 1980

198009 sep 1980

198010 oct 1980

198011 nov 1980

198012 dec 1980

198098 1980, don't know when

198099 1980, no month given

198101 jan 1981

198102 feb 1981

198103 mar 1981

198104 apr 1981

198105 may 1981

198106 jun 1981

198107 jul 1981

198108 aug 1981

198109 sep 1981

198110 oct 1981

198111 nov 1981

198112 dec 1981

198198 1981, don't know when

198199 1981, no month given

198201 jan 1982

198202 feb 1982

198203 mar 1982

198204 apr 1982

198205 may 1982

198206 jun 1982

198207 jul 1982

198208 aug 1982

198209 sep 1982

198210 oct 1982

198211 nov 1982

198212 dec 1982

198298 1982, don't know when

198299 1982, no month given

198301 jan 1983

198302 feb 1983

198303 mar 1983

198304 apr 1983

198305 may 1983

198306 jun 1983

198307 jul 1983

198308 aug 1983

198309 sep 1983

198310 oct 1983

198311 nov 1983

198312 dec 1983

198398 1983, don't know when

198399 1983, no month given

198401 jan 1984

198402 feb 1984

198403 mar 1984

198404 apr 1984

198405 may 1984

198406 jun 1984

198407 jul 1984

198408 aug 1984

198409 sep 1984

198410 oct 1984

198411 nov 1984

198412 dec 1984

198498 1984, don't know when

198499 1984, no month given

198501 jan 1985

198502 feb 1985

198503 mar 1985

198504 apr 1985

198505 may 1985

198506 jun 1985

198507 jul 1985

198508 aug 1985

198509 sep 1985

198510 oct 1985

198511 nov 1985

198512 dec 1985

198598 1985, don't know when

198599 1985, no month given

198601 jan 1986

198602 feb 1986

198603 mar 1986

198604 apr 1986

198605 may 1986

198606 jun 1986

198607 jul 1986

198608 aug 1986

198609 sep 1986

198610 oct 1986

198611 nov 1986

198612 dec 1986

198698 1986, don't know when

198699 1986, no month given

198701 jan 1987

198702 feb 1987

198703 mar 1987

198704 apr 1987

198705 may 1987

198706 jun 1987

198707 jul 1987

198708 aug 1987

198709 sep 1987

198710 oct 1987

198711 nov 1987

198712 dec 1987

198798 1987, don't know when

198799 1987, no month given

198801 jan 1988

198802 feb 1988

198803 mar 1988

198804 apr 1988

198805 may 1988

198806 jun 1988

198807 jul 1988

198808 aug 1988

198809 sep 1988

198810 oct 1988

198811 nov 1988

198812 dec 1988

198898 1988, don't know when

198899 1988, no month given

198901 jan 1989

198902 feb 1989

198903 mar 1989

198904 apr 1989

198905 may 1989

198906 jun 1989

198907 jul 1989

198908 aug 1989

198909 sep 1989

198910 oct 1989

198911 nov 1989

198912 dec 1989

198998 1989, don't know when

198999 1989, no month given

199001 jan 1990

199002 feb 1990

199003 mar 1990

199004 apr 1990

199005 may 1990

199006 jun 1990

199007 jul 1990

199008 aug 1990

199009 sep 1990

199010 oct 1990

199011 nov 1990

199012 dec 1990

199098 1990, don't know when

199099 1990, no month given

199101 jan 1991

199102 feb 1991

199103 mar 1991

199104 apr 1991

199105 may 1991

199106 jun 1991

199107 jul 1991

199108 aug 1991

199109 sep 1991

199110 oct 1991

199111 nov 1991

199112 dec 1991

199198 1991, don't know when

199199 1991, no month given

199201 jan 1992

199202 feb 1992

199203 mar 1992

199204 apr 1992

199205 may 1992

199206 jun 1992

199207 jul 1992

199208 aug 1992

199209 sep 1992

199210 oct 1992

199211 nov 1992

199212 dec 1992

199298 1992, don't know when

199299 1992, no month given

199301 jan 1993

199302 feb 1993

199303 mar 1993

199304 apr 1993

199305 may 1993

199306 jun 1993

199307 jul 1993

199308 aug 1993

199309 sep 1993

199310 oct 1993

199311 nov 1993

199312 dec 1993

199398 1993, don't know when

199399 1993, no month given

199401 jan 1994

199402 feb 1994

199403 mar 1994

199404 apr 1994

199405 may 1994

199406 jun 1994

199407 jul 1994

199408 aug 1994

199409 sep 1994

199410 oct 1994

199411 nov 1994

199412 dec 1994

199498 1994, don't know when

199499 1994, no month given

199501 jan 1995

199502 feb 1995

199503 mar 1995

199504 apr 1995

199505 may 1995

199506 jun 1995

199507 jul 1995

199508 aug 1995

199509 sep 1995

199510 oct 1995

199511 nov 1995

199512 dec 1995

199598 1995, don't know when

199599 1995, no month given

199601 jan 1996

199602 feb 1996

199603 mar 1996

199604 apr 1996

199605 may 1996

199606 jun 1996

199607 jul 1996

199608 aug 1996

199609 sep 1996

199610 oct 1996

199611 nov 1996

199612 dec 1996

199698 1996, don't know when

199699 1996, no month given

199701 jan 1997

199702 feb 1997

199703 mar 1997

199704 apr 1997

199705 may 1997

199706 jun 1997

199707 jul 1997

199708 aug 1997

199709 sep 1997

199710 oct 1997

199711 nov 1997

199712 dec 1997

199798 1997, don't know when

199799 1997, no month given

199801 jan 1998

199802 feb 1998

199803 mar 1998

199804 apr 1998

199805 may 1998

199806 jun 1998

199807 jul 1998

199808 aug 1998

199809 sep 1998

199810 oct 1998

199811 nov 1998

199812 dec 1998

199898 1998, don't know when

199899 1998, no month given

199901 jan 1999

199902 feb 1999

199903 mar 1999

199904 apr 1999

199905 may 1999

199906 jun 1999

199907 jul 1999

199908 aug 1999

199909 sep 1999

199910 oct 1999

199911 nov 1999

199912 dec 1999

199998 1999, don't know when

199999 1999, no month given

200001 jan 2000

200002 feb 2000

200003 mar 2000

200004 apr 2000

200005 may 2000

200006 jun 2000

200007 jul 2000

200008 aug 2000

200009 sep 2000

200010 oct 2000

200011 nov 2000

200012 dec 2000

200098 2000, don't know when

200099 2000, no month given

200101 jan 2001

200102 feb 2001

200103 mar 2001

200104 apr 2001

200105 may 2001

200106 jun 2001

200107 jul 2001

200108 aug 2001

200109 sep 2001

200110 oct 2001

200111 nov 2001

200112 dec 2001

200198 2001, don't know when

200199 2001, no month given

200201 jan 2002

200202 feb 2002

200203 mar 2002

200204 apr 2002

200205 may 2002

200206 jun 2002

200207 jul 2002

200208 aug 2002

200209 sep 2002

200210 oct 2002

200211 nov 2002

200212 dec 2002

200298 2002, don't know when

200299 2002, no month given

200301 jan 2003

200302 feb 2003

200303 mar 2003

200304 apr 2003

200305 may 2003

200306 jun 2003

200307 jul 2003

200308 aug 2003

200309 sep 2003

200310 oct 2003

200311 nov 2003

200312 dec 2003

200398 2003, don't know when

200399 2003, no month given

200401 jan 2004

200402 feb 2004

200403 mar 2004

200404 apr 2004

200405 may 2004

200406 jun 2004

200407 jul 2004

200408 aug 2004

200409 sep 2004

200410 oct 2004

200411 nov 2004

200412 dec 2004

200498 2004, don't know when

200499 2004, no month given

200501 jan 2005

200502 feb 2005

200503 mar 2005

200504 apr 2005

200505 may 2005

200506 jun 2005

200507 jul 2005

200508 aug 2005

200509 sep 2005

200510 oct 2005

200511 nov 2005

200512 dec 2005

200598 2005, don't know when

200599 2005, no month given

200601 jan 2006

200602 feb 2006

200603 mar 2006

200604 apr 2006

200605 may 2006

200606 jun 2006

200607 jul 2006

200608 aug 2006

200609 sep 2006

200610 oct 2006

200611 nov 2006

200612 dec 2006

200698 2006, don't know when

200699 2006, no month given

200701 jan 2007

200702 feb 2007

200703 mar 2007

200704 apr 2007

200705 may 2007

200706 jun 2007

200707 jul 2007

200708 aug 2007

200709 sep 2007

200710 oct 2007

200711 nov 2007

200712 dec 2007

200798 2007, don't know when

200799 2007, no month given

200801 jan 2008

200802 feb 2008

200803 mar 2008

200804 apr 2008

200805 may 2008

200806 jun 2008

200807 jul 2008

200808 aug 2008

200809 sep 2008

200810 oct 2008

200811 nov 2008

200812 dec 2008

200898 2008, don't know when

200899 2008, no month given

200901 jan 2009

200902 feb 2009

200903 mar 2009

200904 apr 2009

200905 may 2009

200906 jun 2009

200907 jul 2009

200908 aug 2009

200909 sep 2009

200910 oct 2009

200911 nov 2009

200912 dec 2009

200998 2009, don't know when

200999 2009, no month given

201001 jan 2010

201002 feb 2010

201003 mar 2010

201004 apr 2010

201005 may 2010

201006 jun 2010

201007 jul 2010

201008 aug 2010

201009 sep 2010

201010 oct 2010

201011 nov 2010

201012 dec 2010

201098 2010, don't know when

201099 2010, no month given

201101 jan 2011

201102 feb 2011

201103 mar 2011

201104 apr 2011

201105 may 2011

201106 jun 2011

201107 jul 2011

201108 aug 2011

201109 sep 2011

201110 oct 2011

201111 nov 2011

201112 dec 2011

201198 2011, don't know when

201199 2011, no month given

201201 jan 2012

201202 feb 2012

201203 mar 2012

201204 apr 2012

201205 may 2012

201206 jun 2012

201207 jul 2012

201208 aug 2012

201209 sep 2012

201210 oct 2012

201211 nov 2012

201212 dec 2012

201298 2012, don't know when

201299 2012, no month given

201301 jan 2013

201302 feb 2013

201303 mar 2013

201304 apr 2013

201305 may 2013

201306 jun 2013

201307 jul 2013

201308 aug 2013

201309 sep 2013

201310 oct 2013

201311 nov 2013

201312 dec 2013

201398 2013, don't know when

201399 2013, no month given

201401 jan 2014

201402 feb 2014

201403 mar 2014

201404 apr 2014

201405 may 2014

201406 jun 2014

201407 jul 2014

201408 aug 2014

201409 sep 2014

201410 oct 2014

201411 nov 2014

201412 dec 2014

201498 2014, don't know when

201499 2014, no month given

201501 jan 2015

201502 feb 2015

201503 mar 2015

201504 apr 2015

201505 may 2015

201506 jun 2015

201507 jul 2015

201508 aug 2015

201509 sep 2015

201510 oct 2015

201511 nov 2015

201512 dec 2015

201598 2015, don't know when

201599 2015, no month given

201601 jan 2016

201602 feb 2016

201603 mar 2016

201604 apr 2016

201605 may 2016

201606 jun 2016

201607 jul 2016

201608 aug 2016

201609 sep 2016

201610 oct 2016

201611 nov 2016

201612 dec 2016

201698 2016, don't know when

201699 2016, no month given

201701 jan 2017

201702 feb 2017

201703 mar 2017

201704 apr 2017

201705 may 2017

201706 jun 2017

201707 jul 2017

201708 aug 2017

201709 sep 2017

201710 oct 2017

201711 nov 2017

201712 dec 2017

201798 2017, don't know when

201799 2017, no month given

201801 jan 2018

201802 feb 2018

201803 mar 2018

201804 apr 2018

201805 may 2018

201806 jun 2018

201807 jul 2018

201808 aug 2018

201809 sep 2018

201810 oct 2018

201811 nov 2018

201812 dec 2018

201898 2018, don't know when

201899 2018, no month given

201901 jan 2019

201902 feb 2019

201903 mar 2019

201904 apr 2019

201905 may 2019

201906 jun 2019

201907 jul 2019

201908 aug 2019

201909 sep 2019

201910 oct 2019

201911 nov 2019

201912 dec 2019

201998 2019, don't know when

201999 2019, no month given

202001 jan 2020

202002 feb 2020

202003 mar 2020

202004 apr 2020

202005 may 2020

202006 jun 2020

202007 jul 2020

202008 aug 2020

202009 sep 2020

202010 oct 2020

202011 nov 2020

202012 dec 2020

202098 2020, don't know when

202099 2020, no month given

202101 jan 2021

202102 feb 2021

202103 mar 2021

202104 apr 2021

202198 2021, don't know month

202199 2021, no month given

999802 feb, don't know year

999806 june, don't know year

999898 don't know year & month

999899 don't know year & no answer month

999910 oct, no answer year

999998 no answer year & don't know month

999999 no answer year & month

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(q) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook