R's industry code (1970)

industry

Source

General Social Survey https://gss.norc.org

Details

Question 2c. What kind of place (do/did) you work for? d. What (do/did) they (make/do)?

Overview

For further details see the official GSS documentation.

Counts by year:

year	NA(m)	NA(n)	accounting, auditing, and bookkeeping services	advertising	agricultural chemicals	agricultural production	agricultural services, except horticultural	air transportation	aircraft and parts	alcoholic beverages	apparel and accessories	apparel and accessories stores, except shoe stores	automobile repair and related services	automobile services, except repair	bakery products	banking	beauty shops	beverage industries	blast furnaces, steel works, rolling and finishing mills	bowling alleys, billiard and pool parlors	business management and consulting services	business services, n.e.c.	canning and preserving fruits, vegetables, and sea foods	cement concrete, gypsum, and plaster products	coal mining	colleges and universities	computer programming services	confectionery and related products	construction and material handling machines	convalescent institutions	credit agencies	crude petroleum and natural gas extractions	cutlery, hand tools, and other hardware	cycles and miscellaneous transportation equipment	dairy product stores	dairy products	department and mail order establishments	detective and protective services	direct selling establishments	dressmaking shops	drug stores	drugs and medicines	drugs, chemicals, and allied products	dry goods and apparel	eating and drinking places	educational services, n.e.c.	electric light and power	electric-gas utilities	electrical goods	electrical machinery, equipment, and supplies, n.e.c.	electrical repair shops	electronic computing equipment	elementary and secondary schools	employment and temporary help agencies	engineering and architectural services	engines and turbines	fabricated structural metal products	farm and garden supply stores	farm machinery and equipment	farm products-raw materials	federal public administration	food and related products	food stores, n.e.c.	footwear, except rubber	forestry	furniture and home furnishing stores	furniture fixtures	gasoline service stations	general building contractors	general contractors, except building	glass and glass products	grain-mill products	grocery stores	hardware and farm equipment stores	hardware, plumbing, and heating supplies	health services, n.e.c.	horticultural services	hospitals	hotels and motels	household appliances	household appliances, tv, and radio stores	industrial chemicals	insurance	knitting mills	laundering, cleaning, and other garment services	legal services	libraries	limited price variety stores	local public administration	lodging places, except hotels and motels	logging	lumber and building material retailing	lumber and construction materials	machinery equipment and supplies	machinery, except electrical, n.e.c.	meat products	metal mining	metals and minerals, n.e.c.	metalworking machinery	misc. general merchandise stores	miscellaneous chemicals	miscellaneous entertainment and recreation services	miscellaneous fabricated metal products	miscellaneous fabricated textile products	miscellaneous food preparation and kindred products	miscellaneous manufacturing industries	miscellaneous nonmetallic mineral and stone products	miscellaneous paper and pulp products	miscellaneous personal services	miscellaneous plastic products	miscellaneous professional and related services	miscellaneous repair services	miscellaneous retail stores	miscellaneous vehicle dealers	mobile dwellings and campers	motor vehicle dealers	motor vehicles and equipment	motor vehicles and motor vehicle equipment	museums, art galleries, and zoos	newspaper publishing and printing	nonmetallic mining and quarrying, except fuel	nonprofit membership organizations	not specified chemicals and allied products	not specified construction	not specified electrical machinery, equipment, and supplies	not specified food industries	not specified machinery	not specified manufacturing industries	not specified metal industries	not specified retail trade	not specified wholesale trade	office and accounting machines	offices of dentists	offices of physicians	optical and health services supplies	ordinance	other primary iron and steel industries	other primary nonferrous industries	paints, varnishes, and related products	paper and its products	paperboard containers and boxes	petroleum products	petroleum refining	photographic equipment and supplies	pipe lines, except natural gas	postal service	pottery and related products	primary aluminum industries	printing, publishing, and allied industries, except newspapers	private households	pulp, paper and paperboard mills	radio broadcasting and television	radio, t.v., and communication equipment	railroad locomotives and equipment	railroads and railway express	real estate, incl. real estate-insurance-law offices	religious organizations	retail bakeries	retail florists	rubber products	sanitary services	sawmills, planing mills, and mill work	scientific and controlling instruments	scrap and waste materials	security, commodity brokerage, and investment companies	services incidental to transportation	services to dwellings and other buildings	ship and boat building and repairing	soaps and cosmetics	special trade contractors	state public administration	street railways and bus lines	structural clay products	tanned, curried, and finished leather	telegraph and miscellaneous communication services	telephone (wire and radio)	theaters and motion pictures	tire, battery, and accessory dealers	trucking service	warehousing and storage	water supply	water transportation	welfare services	wholesalers, n.e.c.	yarn, thread, and fabric mills	commercial research, development, and testing labs	fisheries	floor coverings, except hard surface	fuel and ice dealers	gas and steam supply systems	jewelry stores	leather products, except footwear	liquor stores	metal stamping	miscellaneous textile mill products	offices of chiropractors	offices of health practitioners,n.e.c.	plastics, synthetics and resins, except fibers	residential welfare facilities	screw machine products	taxicab service	tobacco manufactures	watches, clocks, and clockwork-operated devices	miscellaneous wood products	other and not specified utilities	vending machine operators	barber shops	dyeing and finishing textiles, except wool and knit goods	shoe stores	synthetic fibers	not specified educational services	not specified electrical and hardware products	shoe repair shops	miscellaneous petroleum and coal products	not specified professional equipment	NA(y)	NA(i)	Total

1972	162	31	2	5	3	53	4	5	14	1	25	11	2	4	2	28	10	5	15	1	8	6	5	1	2	23	1	1	2	8	4	5	5	1	1	1	24	1	3	2	3	3	2	1	52	8	9	1	1	13	1	7	98	1	3	4	11	2	1	5	48	13	3	5	1	5	7	11	37	16	1	4	20	3	2	10	2	53	10	3	5	4	30	4	11	4	3	5	25	3	1	7	1	8	8	8	1	1	3	1	1	5	11	2	4	5	2	4	3	4	5	1	5	2	3	8	4	22	1	9	1	7	1	5	11	1	1	3	1	2	1	2	7	7	4	6	2	3	2	3	6	3	4	1	1	17	2	1	10	46	5	2	6	1	19	17	8	2	2	13	6	9	2	1	4	3	4	1	5	34	8	2	3	1	4	24	3	6	14	3	3	8	7	7	14	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1613

1973	147	19	3	2	-	38	5	4	9	1	31	15	6	-	4	27	14	3	11	1	2	6	8	3	9	27	-	1	5	11	6	2	4	-	-	-	36	3	4	1	3	4	1	1	51	6	9	1	2	9	2	7	90	-	5	1	6	-	5	2	37	10	1	2	-	6	10	7	26	17	7	1	21	1	2	7	1	59	17	3	7	4	44	6	10	13	2	-	27	2	1	4	2	5	12	2	2	-	1	3	-	7	6	4	5	7	1	3	1	4	3	1	5	-	5	10	3	20	-	6	1	2	6	3	1	1	3	1	-	1	-	4	5	10	4	2	4	1	2	2	6	1	2	5	-	12	2	3	14	31	3	1	9	-	15	15	6	3	1	6	5	4	-	2	2	-	5	6	4	32	7	-	3	-	3	18	7	1	13	1	2	4	13	1	8	1	1	1	2	1	1	1	1	5	1	1	2	1	2	1	4	2	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1504

1974	128	12	1	2	-	36	-	5	9	-	26	15	2	-	4	21	8	5	-	1	-	2	2	1	8	20	1	1	5	16	5	4	5	-	2	5	24	-	4	-	5	3	-	1	39	6	5	3	2	23	1	7	113	3	2	2	2	-	4	1	59	7	1	2	-	3	7	3	36	10	7	1	30	3	-	9	1	55	13	8	1	7	26	3	9	8	4	3	27	5	2	2	-	2	11	2	6	2	2	6	1	15	3	1	12	16	-	4	23	1	3	3	7	3	1	5	2	32	1	4	1	-	-	2	4	2	1	2	-	1	-	5	5	14	1	-	8	-	2	1	4	2	3	6	3	10	-	5	17	32	4	-	6	2	21	8	8	1	1	8	7	4	3	-	5	2	3	2	2	32	24	6	-	-	-	22	1	1	9	1	3	-	3	6	12	1	-	1	-	1	-	-	2	-	5	-	-	2	-	-	2	3	-	3	3	1	-	-	-	-	-	-	-	-	-	-	-	1484

1975	134	12	4	3	-	33	2	2	6	1	42	8	8	1	-	20	11	4	11	-	6	10	4	6	9	14	2	-	2	19	7	3	2	1	-	2	45	-	2	-	9	4	-	1	68	1	6	-	-	17	2	5	79	1	7	1	4	4	5	2	46	5	1	2	2	7	10	6	32	9	4	2	22	1	1	4	1	68	12	5	2	7	30	13	9	7	3	2	19	1	2	3	1	4	19	9	3	-	1	3	-	9	3	4	-	9	3	-	-	3	11	6	10	-	4	10	3	22	1	8	1	7	-	1	2	1	2	-	1	-	1	2	5	8	2	1	5	2	1	1	2	-	1	3	-	19	-	2	8	38	2	1	8	1	10	10	14	2	-	9	5	4	3	3	5	1	5	3	3	36	12	6	-	-	1	14	5	3	15	4	3	1	5	5	5	1	1	-	3	3	1	2	1	2	4	-	1	2	2	-	-	1	-	2	1	-	2	2	2	1	-	-	-	-	-	-	-	1490

1976	145	2	3	3	1	48	5	6	10	1	37	10	9	1	1	19	6	3	13	-	1	8	4	-	16	27	1	-	7	19	1	6	1	-	1	6	32	1	11	1	6	2	4	-	58	-	11	2	2	15	2	12	81	2	5	1	8	-	4	-	34	9	1	2	2	5	10	6	18	8	4	1	31	3	2	8	2	53	12	1	3	2	31	1	11	10	-	2	40	4	2	4	2	2	14	6	6	-	5	3	-	4	2	4	3	9	3	5	3	4	4	5	10	1	-	8	1	24	-	10	-	8	5	4	-	1	1	-	-	2	2	5	2	12	7	5	3	4	1	-	6	1	2	3	-	9	1	3	17	30	4	3	4	-	13	13	11	4	1	10	3	4	4	1	1	1	2	3	1	21	6	2	1	1	-	23	4	3	11	4	2	5	13	2	9	2	2	-	3	3	1	2	2	-	1	-	-	-	1	-	-	3	2	-	1	1	2	2	3	3	-	-	-	-	-	-	-	1499

1977	113	1	5	2	1	36	4	4	11	-	27	15	10	1	4	25	13	6	13	-	3	6	13	2	10	20	2	3	9	14	7	8	4	1	1	6	35	3	2	-	5	4	2	-	64	2	9	-	1	19	1	1	101	3	6	-	6	-	5	1	45	5	1	3	1	4	13	4	27	20	4	3	28	1	-	6	2	63	12	4	2	2	19	8	6	4	2	6	19	2	3	3	5	3	17	5	7	-	5	5	-	7	3	6	1	11	-	6	4	-	-	7	12	3	2	12	3	21	-	6	1	6	1	11	1	-	2	2	1	1	2	2	6	8	4	3	4	-	1	-	6	3	3	-	-	12	3	5	13	28	5	6	8	-	7	17	13	2	4	6	4	7	4	1	3	3	1	5	3	33	13	1	2	-	-	20	8	2	24	3	5	-	6	2	4	2	3	1	2	3	-	-	2	3	-	-	-	4	3	-	1	4	-	2	-	1	2	-	2	-	1	-	-	-	-	-	-	1530

1978	127	1	3	3	1	26	2	1	14	2	39	12	9	1	4	34	14	1	18	-	-	8	8	2	14	23	2	-	7	12	5	6	3	2	-	3	31	2	6	2	9	8	2	1	73	-	10	1	1	17	2	3	90	2	3	2	4	-	3	2	47	6	3	2	5	3	8	6	20	10	3	2	27	3	2	6	6	60	14	6	-	2	29	4	10	9	1	6	34	1	-	1	5	3	8	4	8	-	7	3	1	4	9	2	2	8	2	1	3	6	3	6	11	-	-	5	-	17	1	6	1	7	1	1	5	-	1	1	-	-	1	1	7	12	9	3	9	1	2	1	1	6	6	3	-	4	2	4	18	30	4	1	5	1	10	18	9	1	1	8	3	7	1	-	8	2	4	4	1	25	16	5	-	-	1	27	5	1	7	3	2	4	10	5	12	2	-	-	2	2	3	2	1	2	1	-	1	2	-	2	3	-	5	3	1	2	2	-	1	2	1	1	1	-	-	-	-	1532

1980	119	4	6	2	-	35	7	7	8	2	28	11	6	1	3	19	8	2	13	2	5	2	2	2	18	14	4	1	4	16	3	7	2	-	-	3	24	1	8	-	9	3	1	2	67	12	2	3	3	13	1	9	76	2	4	1	15	1	-	2	37	12	3	4	1	5	4	6	13	9	-	1	21	4	2	5	3	56	6	3	5	2	28	5	8	11	4	2	32	3	3	2	1	2	17	3	1	-	2	1	2	10	14	2	2	13	-	7	9	5	11	3	12	3	1	11	1	23	-	10	1	9	6	5	5	-	1	1	-	3	4	2	4	11	2	2	5	5	-	1	4	1	5	4	-	8	1	5	15	29	1	9	3	2	6	15	8	1	-	4	1	4	1	-	7	4	1	5	2	39	28	3	1	-	1	14	5	-	18	1	1	2	7	4	6	2	-	1	-	3	2	-	2	-	7	-	-	3	-	-	5	1	-	1	1	-	1	1	2	1	-	1	-	-	-	-	-	1468

1982	122	18	3	3	-	36	2	3	17	-	40	17	12	3	2	30	19	3	10	4	5	25	11	-	8	30	3	4	4	30	5	6	3	-	-	2	27	6	2	-	8	2	3	5	66	9	12	2	3	13	-	3	116	3	7	5	5	2	3	3	35	13	2	1	4	8	9	8	19	19	3	2	27	2	-	13	2	83	11	4	3	3	32	3	7	12	1	4	24	11	4	6	-	7	19	10	7	-	5	1	2	13	7	-	12	11	1	7	11	3	2	8	12	3	3	9	2	21	3	5	5	10	6	7	1	1	3	6	4	4	4	1	6	15	3	4	10	-	2	1	2	3	3	10	1	15	1	2	18	65	1	3	6	1	13	16	8	2	2	3	9	2	2	1	7	-	9	6	3	53	15	4	1	-	2	23	3	5	23	-	3	2	21	6	8	2	-	-	2	1	3	2	-	-	2	3	-	-	-	2	3	2	3	1	-	1	3	-	4	1	-	-	3	2	-	-	-	1860

1983	103	10	8	6	2	41	7	5	5	1	25	12	12	5	1	25	9	2	22	1	1	12	1	3	7	38	5	-	6	17	4	6	-	1	3	4	24	2	4	1	5	1	4	2	80	6	3	3	3	12	2	7	96	2	8	5	16	1	4	-	48	14	2	6	-	5	7	6	26	2	4	-	29	5	2	14	2	79	11	3	5	4	32	4	7	9	1	2	50	-	1	2	-	6	10	13	8	-	2	6	1	13	9	1	13	5	4	1	6	1	4	7	12	-	-	5	2	11	3	3	1	6	2	-	4	2	-	6	2	2	1	3	12	11	10	1	2	1	4	2	4	2	4	-	-	6	-	3	10	24	4	4	7	-	6	21	7	2	4	3	3	1	2	2	10	1	2	1	2	49	24	1	-	-	2	21	2	4	10	1	-	-	-	5	8	-	-	-	-	1	2	1	2	1	3	3	4	-	-	1	-	2	4	4	3	-	2	-	-	1	1	-	-	-	1	-	-	1599

1984	88	6	8	7	-	23	7	7	13	-	13	10	10	2	6	25	12	1	6	-	7	14	7	1	3	23	1	-	-	17	9	5	4	-	-	5	25	-	2	-	2	8	1	-	76	7	9	2	3	12	-	9	96	3	4	-	3	3	-	3	39	10	6	7	-	6	7	7	36	3	3	1	30	3	2	23	1	63	14	5	-	4	17	4	17	8	-	2	33	1	3	2	1	2	5	6	1	-	-	1	3	8	12	-	7	42	2	5	14	2	3	1	23	-	1	9	4	20	-	6	3	5	2	2	1	1	1	1	-	1	-	3	5	6	5	-	16	3	1	1	-	-	4	4	-	6	1	2	10	49	-	-	6	1	6	20	2	-	3	4	-	1	2	1	6	4	4	2	3	14	18	3	1	-	13	6	3	1	13	-	-	3	3	10	18	2	-	-	-	1	1	1	4	-	2	-	1	2	-	-	2	-	1	2	1	3	1	-	1	1	-	-	-	-	-	-	-	1473

1985	81	-	4	3	-	39	5	4	8	1	29	12	7	3	3	28	14	4	10	2	8	14	6	2	-	35	5	-	1	18	5	6	2	1	-	3	32	4	3	-	3	3	4	1	79	3	10	1	2	10	1	9	85	1	9	1	4	2	1	-	36	13	5	1	2	5	7	5	36	14	6	-	12	4	2	17	6	75	12	-	8	4	44	-	5	8	2	3	31	5	1	3	2	9	13	6	2	-	4	5	1	9	5	1	-	8	2	3	7	5	6	5	14	-	2	9	1	13	2	4	-	5	4	3	4	-	1	2	1	1	4	2	9	12	4	2	3	3	-	-	2	3	5	2	-	10	1	-	20	28	2	6	3	1	12	26	10	1	3	4	4	2	3	6	3	4	3	2	2	46	18	2	-	-	2	27	5	4	14	2	1	3	5	4	3	-	-	1	1	6	2	-	1	1	1	-	-	3	1	1	2	-	1	1	3	-	3	-	1	-	-	-	-	-	-	-	-	1534

1986	100	5	10	4	-	36	6	4	16	-	22	7	12	2	-	31	16	6	9	2	2	19	2	2	1	29	7	-	3	34	1	7	3	1	-	1	32	4	3	1	6	1	-	-	59	5	6	3	2	9	1	10	82	5	7	1	4	-	2	2	36	8	2	3	3	3	8	7	24	8	3	2	26	2	4	11	5	68	13	4	3	2	25	1	13	6	3	6	35	2	2	1	3	4	8	3	2	1	4	3	1	13	3	1	4	10	1	3	3	2	2	5	14	-	-	7	1	24	-	6	1	5	2	2	5	2	4	-	2	2	-	4	12	10	3	6	1	1	4	2	1	-	4	2	-	8	2	-	16	25	4	2	5	-	6	23	4	3	1	1	-	5	3	4	6	1	-	1	6	32	21	4	-	-	2	10	1	4	13	5	-	2	4	9	10	1	-	1	1	2	4	1	2	-	1	-	2	3	2	-	2	2	-	2	2	-	4	1	-	-	4	-	-	3	-	-	-	1470

1987	95	8	4	3	-	40	8	3	19	-	29	9	6	5	6	31	19	4	12	-	6	15	-	3	3	40	9	2	2	41	7	3	4	1	-	1	34	1	6	-	6	4	-	-	78	7	9	1	3	10	-	8	96	5	10	2	5	2	1	1	47	11	1	6	-	10	4	9	26	2	6	-	39	4	-	23	6	81	23	1	6	-	32	3	9	15	3	4	39	4	1	3	2	4	5	7	1	1	4	6	2	13	8	2	3	15	3	3	10	-	9	12	23	-	-	6	1	28	1	10	2	2	6	13	4	-	1	7	4	3	2	2	6	14	1	1	2	3	-	1	4	-	2	-	-	13	-	1	15	59	3	10	5	1	12	26	6	3	2	3	3	6	2	2	6	5	2	4	2	62	21	7	2	-	5	23	2	7	14	2	2	1	8	12	10	1	2	1	-	2	1	1	-	-	9	-	1	3	4	1	2	-	-	4	5	-	2	-	4	1	1	-	-	1	-	-	-	1819

1988	85	3	7	5	2	24	2	7	9	2	25	15	13	2	3	22	10	9	1	2	12	11	1	4	3	29	3	2	3	18	10	-	1	-	-	1	32	3	1	-	7	4	-	-	61	1	9	3	-	9	1	3	77	5	8	1	10	-	3	-	41	10	5	-	-	5	11	6	25	4	2	-	26	4	3	15	2	73	11	3	3	4	22	2	5	8	1	4	33	2	4	1	2	3	11	5	-	-	3	11	1	14	1	2	2	11	2	2	8	5	6	4	20	3	-	9	3	16	1	4	3	9	3	3	4	2	-	1	-	3	1	1	4	12	7	-	5	3	1	3	1	1	2	2	-	13	1	-	12	38	4	4	6	-	14	15	10	3	2	4	1	2	3	2	3	2	13	2	1	50	17	6	-	-	1	24	7	2	11	6	1	2	6	2	3	3	2	-	-	1	2	-	1	3	8	1	1	8	4	-	-	1	1	1	3	-	1	1	2	1	1	1	1	1	-	-	-	1481

1989	92	7	4	1	1	23	6	11	6	-	16	14	8	5	1	28	6	6	2	1	7	13	3	-	5	49	6	-	4	20	7	3	1	-	-	1	24	3	3	-	7	4	5	3	66	9	10	4	2	7	6	5	86	7	6	4	9	1	-	-	52	8	-	4	-	8	5	5	20	10	3	1	32	5	1	12	5	84	12	5	2	4	32	5	10	19	3	3	21	9	2	5	3	-	10	9	1	-	7	7	2	9	1	2	3	17	1	4	6	2	3	4	9	-	1	13	1	13	-	5	-	10	1	13	5	1	1	4	1	11	2	1	7	6	7	3	4	-	1	2	5	-	4	1	-	8	1	3	19	17	3	3	2	-	2	23	6	4	1	1	1	4	1	-	7	8	8	2	-	36	15	6	1	-	-	19	5	5	16	2	2	2	6	6	8	2	1	-	-	1	-	-	2	2	1	-	5	4	4	1	1	-	1	4	2	2	2	3	-	2	-	5	-	1	-	-	-	1537

1990	71	7	7	2	-	27	5	4	11	1	19	11	10	2	3	24	9	-	8	1	6	17	1	2	-	35	4	1	3	14	4	4	1	-	-	1	21	1	2	-	6	3	2	-	61	8	4	3	-	11	3	4	97	8	6	1	9	1	-	1	48	12	2	5	3	1	5	2	13	16	5	2	19	1	1	13	3	55	9	-	-	4	28	8	6	12	3	1	18	1	2	2	3	6	4	8	-	-	4	2	1	10	7	1	2	10	-	3	4	4	7	6	18	-	-	9	-	21	-	4	1	6	3	7	4	1	3	-	1	9	2	-	5	6	3	1	2	-	-	2	3	1	4	-	-	9	2	2	14	25	2	7	2	-	4	24	5	3	4	4	2	2	-	1	5	4	8	1	1	53	17	3	1	-	3	15	3	-	19	3	-	2	17	5	8	5	-	1	2	2	1	1	-	-	2	3	-	3	1	1	1	-	-	1	1	-	-	-	1	-	-	1	-	-	-	-	-	1372

1991	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1517	-	1517

1993	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1606	-	1606

1994	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2992	-	2992

1996	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2904	-	2904

1998	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2832	-	2832

2000	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2817	-	2817

2002	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2765	-	2765

2004	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2812	-	2812

2006	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	4510	-	4510

2008	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2023	-	2023

2010	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2044	-	2044

2012	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1974	-	1974

2014	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2538	-	2538

2016	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2867	-	2867

2018	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2348	-	2348

2021	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	4032	-	4032

2022	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	3544	3544

Total	1912	146	82	56	11	594	77	82	185	13	473	204	142	38	47	437	198	64	174	18	79	188	78	34	116	476	56	16	67	324	90	81	45	9	8	45	502	35	66	8	99	61	31	18	1098	90	133	33	30	219	26	109	1559	53	100	32	121	19	41	25	735	166	39	55	24	89	132	104	434	177	65	23	440	49	26	196	50	1128	212	58	55	59	501	74	153	163	36	55	507	56	34	51	33	70	191	106	56	5	59	67	19	163	104	35	75	207	27	61	115	51	82	84	217	18	23	145	32	348	14	106	23	104	49	82	61	16	26	37	18	46	27	40	107	174	76	40	85	30	24	23	57	27	58	46	5	179	20	41	246	594	51	62	91	11	176	307	135	37	32	91	57	68	36	27	88	45	74	50	41	647	280	61	16	2	40	330	69	49	244	41	30	41	134	91	146	27	12	8	18	33	24	14	23	19	48	11	18	40	24	10	28	21	19	31	27	11	27	10	23	14	9	9	5	8	1	42581	3544	72390

Values

17 agricultural production

18 agricultural services, except horticultural

19 horticultural services

27 forestry

28 fisheries

47 metal mining

48 coal mining

49 crude petroleum and natural gas extractions

57 nonmetallic mining and quarrying, except fuel

67 general building contractors

68 general contractors, except building

69 special trade contractors

77 not specified construction

107 logging

108 sawmills, planing mills, and mill work

109 miscellaneous wood products

118 furniture fixtures

119 glass and glass products

127 cement concrete, gypsum, and plaster products

128 structural clay products

137 pottery and related products

138 miscellaneous nonmetallic mineral and stone products

139 blast furnaces, steel works, rolling and finishing mills

147 other primary iron and steel industries

148 primary aluminum industries

149 other primary nonferrous industries

157 cutlery, hand tools, and other hardware

158 fabricated structural metal products

159 screw machine products

167 metal stamping

168 miscellaneous fabricated metal products

169 not specified metal industries

177 engines and turbines

178 farm machinery and equipment

179 construction and material handling machines

187 metalworking machinery

188 office and accounting machines

189 electronic computing equipment

197 machinery, except electrical, n.e.c.

198 not specified machinery

199 household appliances

207 radio, t.v., and communication equipment

208 electrical machinery, equipment, and supplies, n.e.c.

209 not specified electrical machinery, equipment, and supplies

219 motor vehicles and motor vehicle equipment

227 aircraft and parts

228 ship and boat building and repairing

229 railroad locomotives and equipment

237 mobile dwellings and campers

238 cycles and miscellaneous transportation equipment

239 scientific and controlling instruments

247 optical and health services supplies

248 photographic equipment and supplies

249 watches, clocks, and clockwork-operated devices

257 not specified professional equipment

258 ordinance

259 miscellaneous manufacturing industries

268 meat products

269 dairy products

278 canning and preserving fruits, vegetables, and sea foods

279 grain-mill products

287 bakery products

288 confectionery and related products

289 beverage industries

297 miscellaneous food preparation and kindred products

298 not specified food industries

299 tobacco manufactures

307 knitting mills

308 dyeing and finishing textiles, except wool and knit goods

309 floor coverings, except hard surface

317 yarn, thread, and fabric mills

318 miscellaneous textile mill products

319 apparel and accessories

327 miscellaneous fabricated textile products

328 pulp, paper and paperboard mills

329 miscellaneous paper and pulp products

337 paperboard containers and boxes

338 newspaper publishing and printing

339 printing, publishing, and allied industries, except newspapers

347 industrial chemicals

348 plastics, synthetics and resins, except fibers

349 synthetic fibers

357 drugs and medicines

358 soaps and cosmetics

359 paints, varnishes, and related products

367 agricultural chemicals

368 miscellaneous chemicals

369 not specified chemicals and allied products

377 petroleum refining

378 miscellaneous petroleum and coal products

379 rubber products

387 miscellaneous plastic products

388 tanned, curried, and finished leather

389 footwear, except rubber

397 leather products, except footwear

398 not specified manufacturing industries

407 railroads and railway express

408 street railways and bus lines

409 taxicab service

417 trucking service

418 warehousing and storage

419 water transportation

427 air transportation

428 pipe lines, except natural gas

429 services incidental to transportation

447 radio broadcasting and television

448 telephone (wire and radio)

449 telegraph and miscellaneous communication services

467 electric light and power

468 electric-gas utilities

469 gas and steam supply systems

477 water supply

478 sanitary services

479 other and not specified utilities

507 motor vehicles and equipment

508 drugs, chemicals, and allied products

509 dry goods and apparel

527 food and related products

528 farm products-raw materials

529 electrical goods

537 hardware, plumbing, and heating supplies

538 not specified electrical and hardware products

539 machinery equipment and supplies

557 metals and minerals, n.e.c.

558 petroleum products

559 scrap and waste materials

567 alcoholic beverages

568 paper and its products

569 lumber and construction materials

587 wholesalers, n.e.c.

588 not specified wholesale trade

607 lumber and building material retailing

608 hardware and farm equipment stores

609 department and mail order establishments

617 limited price variety stores

618 vending machine operators

619 direct selling establishments

627 misc. general merchandise stores

628 grocery stores

629 dairy product stores

637 retail bakeries

638 food stores, n.e.c.

639 motor vehicle dealers

647 tire, battery, and accessory dealers

648 gasoline service stations

649 miscellaneous vehicle dealers

657 apparel and accessories stores, except shoe stores

658 shoe stores

667 furniture and home furnishing stores

668 household appliances, tv, and radio stores

669 eating and drinking places

677 drug stores

678 liquor stores

679 farm and garden supply stores

687 jewelry stores

688 fuel and ice dealers

689 retail florists

697 miscellaneous retail stores

698 not specified retail trade

707 banking

708 credit agencies

709 security, commodity brokerage, and investment companies

717 insurance

718 real estate, incl. real estate-insurance-law offices

727 advertising

728 services to dwellings and other buildings

729 commercial research, development, and testing labs

737 employment and temporary help agencies

738 business management and consulting services

739 computer programming services

747 detective and protective services

748 business services, n.e.c.

749 automobile services, except repair

757 automobile repair and related services

758 electrical repair shops

759 miscellaneous repair services

769 private households

777 hotels and motels

778 lodging places, except hotels and motels

779 laundering, cleaning, and other garment services

787 beauty shops

788 barber shops

789 shoe repair shops

797 dressmaking shops

798 miscellaneous personal services

807 theaters and motion pictures

808 bowling alleys, billiard and pool parlors

809 miscellaneous entertainment and recreation services

828 offices of physicians

829 offices of dentists

837 offices of chiropractors

838 hospitals

839 convalescent institutions

847 offices of health practitioners,n.e.c.

848 health services, n.e.c.

849 legal services

857 elementary and secondary schools

858 colleges and universities

859 libraries

867 educational services, n.e.c.

868 not specified educational services

869 museums, art galleries, and zoos

877 religious organizations

878 welfare services

879 residential welfare facilities

887 nonprofit membership organizations

888 engineering and architectural services

889 accounting, auditing, and bookkeeping services

897 miscellaneous professional and related services

907 postal service

917 federal public administration

927 state public administration

937 local public administration