Contrast between judge and lover

judgeluv

Source

General Social Survey https://gss.norc.org

Details

Question 114. There are many different ways of picturing God. We'd like to know the kinds of images you are most likely to associate with God. Here is a card with sets of contrasting images. On a scale of 1-7 where would you place your image of God between the two contrasting images? c. 1. Judge 7. Lover

Overview

For further details see the official GSS documentation.

Counts by year:

year	iap	2	3	4	5	6	don't know	judge	lover	no answer	Total

1972	1613	-	-	-	-	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	-	-	-	-	1504

1974	1484	-	-	-	-	-	-	-	-	-	1484

1975	1490	-	-	-	-	-	-	-	-	-	1490

1976	1499	-	-	-	-	-	-	-	-	-	1499

1977	1530	-	-	-	-	-	-	-	-	-	1530

1978	1532	-	-	-	-	-	-	-	-	-	1532

1980	1468	-	-	-	-	-	-	-	-	-	1468

1982	1860	-	-	-	-	-	-	-	-	-	1860

1983	1599	-	-	-	-	-	-	-	-	-	1599

1984	484	102	84	218	41	22	1	373	83	65	1473

1985	-	118	113	363	81	51	20	583	148	57	1534

1986	-	133	97	316	53	53	13	620	129	56	1470

1987	-	154	122	449	60	53	20	706	169	86	1819

1988	-	143	99	401	48	54	33	528	113	62	1481

1989	531	90	86	279	31	30	6	367	73	44	1537

1990	444	81	72	229	34	24	9	345	84	50	1372

1991	158	140	116	320	75	55	-	385	59	209	1517

1993	526	103	76	250	54	46	9	396	87	59	1606

1994	1015	197	137	485	77	61	31	737	172	80	2992

1996	1913	85	77	268	32	31	12	351	95	40	2904

1998	1387	135	108	420	56	59	22	482	106	57	2832

2000	2817	-	-	-	-	-	-	-	-	-	2817

2002	2765	-	-	-	-	-	-	-	-	-	2765

2004	2812	-	-	-	-	-	-	-	-	-	2812

2006	4510	-	-	-	-	-	-	-	-	-	4510

2008	658	94	128	355	57	53	111	448	82	37	2023

2010	2044	-	-	-	-	-	-	-	-	-	2044

2012	1974	-	-	-	-	-	-	-	-	-	1974

2014	2538	-	-	-	-	-	-	-	-	-	2538

2016	2867	-	-	-	-	-	-	-	-	-	2867

2018	2348	-	-	-	-	-	-	-	-	-	2348

2021	4032	-	-	-	-	-	-	-	-	-	4032

2022	3544	-	-	-	-	-	-	-	-	-	3544

Total	54946	1575	1315	4353	699	592	287	6321	1400	902	72390

Values

1 judge

7 lover

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(q) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook