R child census industry code (2010)

kdind10

Source

General Social Survey https://gss.norc.org

Details

Question None

Overview

For further details see the official GSS documentation.

Counts by year:

year	iap	accounting, tax preparation, bookkeeping, and payroll services	active duty military personnel, all other	administration of economic programs and space research	administration of environmental quality and housing programs	administration of human resource programs	advertising and related services	aerospace product and parts manufacturing	air transportation	aircraft and parts manufacturing	aluminum production and processing	animal production	apparel, fabrics, and notions, merchant wholesalers	architectural, engineering, and related services	auto parts, accessories, and tire stores	automobile dealers	automotive equipment rental and leasing	automotive repair and maintenance	banking and related activities	beauty salons	beverage manufacturing	book stores and news dealers	building material and supplies dealers	bus service and urban transit	business support services	car washes	child day care services	civic, social, advocacy organizations, and grantmaking and giving services	clothing stores	colleges and universities, including junior colleges	commercial and service industry machinery manufacturing	communications, and audio and video equipment manufacturing	community food and housing, and emergency services	computer and peripheral equipment manufacturing	computer systems design and related services	construction	couriers and messengers	crop production	cut and sew apparel manufacturing	dairy product manufacturing	department stores and discount stores	don't know	drycleaning and laundry services	electric power generation, transmission, and distribution	electrical lighting and electrical equipment manufacturing, and other electrical component manufacturing, n.e.c.	electronic and precision equipment repair and maintenance	electronic component and product manufacturing, n.e.c.	elementary and secondary schools	employment services	executive offices and legislative bodies	fabric mills, except knitting mills	fishing, hunting, and trapping	funeral homes, and cemeteries and crematories	furniture and related product manufacturing	gasoline stations	glass and glass product manufacturing	groceries and related products, merchant wholesalers	grocery stores	hardware, plumbing and heating equipment, and supplies, merchant wholesalers	health and personal care, except drug, stores	home health care services	hospitals	household appliance stores	independent artists, performing arts, spectator sports, and related industries	individual and family services	industrial and miscellaneous chemicals	insurance carriers and related activities	investigation and security services	iron and steel mills and steel product manufacturing	jewelry, luggage, and leather goods stores	justice, public order, and safety activities	knitting fabric mills, and apparel knitting mills	landscaping services	legal services	libraries and archives	lumber and other construction materials, merchant wholesalers	machine shops; turned product; screw, nut, and bolt manufacturing	machinery, equipment, and supplies, merchant wholesalers	management, scientific, and technical consulting services	metal forgings and stampings	miscellaneous durable goods, merchant wholesalers	miscellaneous general merchandise stores	miscellaneous retail stores	motion pictures and video industries	motor vehicles and motor vehicle equipment manufacturing	motor vehicles, parts and supplies, merchant wholesalers	museums, art galleries, historical sites, and similar institutions	music stores	national security and international affairs	natural gas distribution	navigational, measuring, electromedical, and control instruments manufacturing	newspaper publishers	no answer	non-depository credit and related activities	not specified manufacturing industries	nursing care facilities	offices of dentists	offices of other health practitioners	offices of physicians	oil and gas extraction	other amusement, gambling, and recreation industries	other general government and support	other health care services	other information services	other personal services	other professional, scientific, and technical services	other schools and instruction, and educational support services	outpatient care centers	paper and paper products, merchant wholesalers	periodical, book, and directory publishers	personal and household goods repair and maintenance	petroleum refining	pharmacies and drug stores	plastics product manufacturing	postal service	printing and related support activities	private households	professional and commercial equipment and supplies, merchant wholesalers	pulp, paper, and paperboard mills	radio and television broadcasting and cable subscription programming	radio, tv, and computer stores	real estate	recreational vehicle parks and camps, and rooming and boarding houses	recyclable material, merchant wholesalers	religious organizations	residential care facilities, without nursing	resin, synthetic rubber and fibers, and filaments manufacturing	restaurants and other food services	retail bakeries	retail florists	savings institutions, including credit unions	scientific research and development services	seafood and other miscellaneous foods, n.e.c.	securities, commodities, funds, trusts, and other financial investments	services incidental to transportation	services to buildings and dwellings	sewage treatment facilities	ship and boat building	shoe stores	soap, cleaning compound, and cosmetics manufacturing	specialized design services	sporting goods, camera, and hobby and toy stores	structural clay product manufacturing	structural metals, and boiler, tank, and shipping container manufacturing	support activities for agriculture and forestry	support activities for mining	textile product mills, except carpet and rug	travel arrangements and reservation services	traveler accommodation	truck transportation	u.s. air force	u.s. armed forces/n.s.	u.s. army	u.s. coast guard	u.s. marine corps	u.s. navy	uncodeable	used merchandise stores	veterinary services	video tape and disk rental	warehousing and storage	waste management and remediation services	water transportation	wholesale electronic markets, agents and brokers	wired telecommunications carriers	Total

1972	1613	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1504

1974	1484	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1484

1975	1490	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1490

1976	1499	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1499

1977	1530	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1530

1978	1532	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1532

1980	1468	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1468

1982	1860	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1860

1983	1599	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1599

1984	1473	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1473

1985	1534	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1534

1986	1470	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1470

1987	1819	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1819

1988	1481	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1481

1989	1537	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1537

1990	1372	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1372

1991	1517	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1517

1993	1606	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1606

1994	2386	5	2	2	1	1	2	1	6	2	1	5	3	3	2	3	2	4	14	5	1	1	3	1	1	1	5	3	3	16	1	3	1	5	7	27	3	5	5	1	10	4	2	1	3	1	2	27	1	8	2	2	2	4	2	1	8	14	1	1	1	22	1	4	3	3	12	1	2	1	15	1	3	9	2	1	1	1	1	2	1	2	3	4	4	1	1	2	6	2	1	1	35	3	4	7	3	1	8	1	5	2	3	1	1	1	2	2	2	3	2	2	4	1	4	1	3	1	1	4	1	9	1	1	3	3	1	39	2	2	2	2	1	3	2	2	1	2	1	1	5	1	1	1	1	3	1	3	8	5	2	1	2	1	1	3	2	1	1	1	4	1	1	1	2	2992

1996	2904	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2904

1998	2832	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2832

2000	2817	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2817

2002	2765	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2765

2004	2812	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2812

2006	4510	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	4510

2008	2023	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2023

2010	2044	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2044

2012	1974	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1974

2014	2538	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2538

2016	2867	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2867

2018	2348	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2348

2021	4032	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	4032

2022	3544	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	3544

Total	71784	5	2	2	1	1	2	1	6	2	1	5	3	3	2	3	2	4	14	5	1	1	3	1	1	1	5	3	3	16	1	3	1	5	7	27	3	5	5	1	10	4	2	1	3	1	2	27	1	8	2	2	2	4	2	1	8	14	1	1	1	22	1	4	3	3	12	1	2	1	15	1	3	9	2	1	1	1	1	2	1	2	3	4	4	1	1	2	6	2	1	1	35	3	4	7	3	1	8	1	5	2	3	1	1	1	2	2	2	3	2	2	4	1	4	1	3	1	1	4	1	9	1	1	3	3	1	39	2	2	2	2	1	3	2	2	1	2	1	1	5	1	1	1	1	3	1	3	8	5	2	1	2	1	1	3	2	1	1	1	4	1	1	1	2	72390

Values

170 crop production

180 animal production

190 forestry, except logging

270 logging

280 fishing, hunting, and trapping

290 support activities for agriculture and forestry

370 oil and gas extraction

380 coal mining

390 metal ore mining

470 nonmetallic mineral mining and quarrying

480 not specified type of mining

490 support activities for mining

570 electric power generation, transmission, and distribution

580 natural gas distribution

590 electric and gas, and other combinations

670 water, steam, air-conditioning, and irrigation systems

680 sewage treatment facilities

690 not specified utilities

770 construction

1070 animal food, grain, and oilseed milling

1080 sugar and confectionery products

1090 fruit and vegetable preserving and specialty food manufacturing

1170 dairy product manufacturing

1180 animal slaughtering and processing

1190 retail bakeries

1270 bakeries, except retail

1280 seafood and other miscellaneous foods, n.e.c.

1290 not specified food industries

1370 beverage manufacturing

1390 tobacco manufacturing

1470 fiber, yarn, and thread mills

1480 fabric mills, except knitting mills

1490 textile and fabric finishing and coating mills

1570 carpet and rug mills

1590 textile product mills, except carpet and rug

1670 knitting fabric mills, and apparel knitting mills

1680 cut and sew apparel manufacturing

1690 apparel accessories and other apparel manufacturing

1770 footwear manufacturing

1790 leather tanning and finishing and other allied products manufacturing

1870 pulp, paper, and paperboard mills

1880 paperboard containers and boxes

1890 miscellaneous paper and pulp products

1990 printing and related support activities

2070 petroleum refining

2090 miscellaneous petroleum and coal products

2170 resin, synthetic rubber and fibers, and filaments manufacturing

2180 agricultural chemical manufacturing

2190 pharmaceutical and medicine manufacturing

2270 paint, coating, and adhesive manufacturing

2280 soap, cleaning compound, and cosmetics manufacturing

2290 industrial and miscellaneous chemicals

2370 plastics product manufacturing

2380 tire manufacturing

2390 rubber product, except tire, manufacturing

2470 pottery, ceramics, and plumbing fixture manufacturing

2480 structural clay product manufacturing

2490 glass and glass product manufacturing

2570 cement, concrete, lime, and gypsum product manufacturing

2590 miscellaneous nonmetallic mineral product manufacturing

2670 iron and steel mills and steel product manufacturing

2680 aluminum production and processing

2690 nonferrous metal (except aluminum) production and processing

2770 foundries

2780 metal forgings and stampings

2790 cutlery and hand tool manufacturing

2870 structural metals, and boiler, tank, and shipping container manufacturing

2880 machine shops; turned product; screw, nut, and bolt manufacturing

2890 coating, engraving, heat treating and allied activities

2970 ordnance

2980 miscellaneous fabricated metal products manufacturing

2990 not specified metal industries

3070 agricultural implement manufacturing

3080 construction, and mining and oil and gas field machinery manufacturing

3090 commercial and service industry machinery manufacturing

3170 metalworking machinery manufacturing

3180 engines, turbines, and power transmission equipment manufacturing

3190 machinery manufacturing, n.e.c.

3290 not specified machinery manufacturing

3360 computer and peripheral equipment manufacturing

3370 communications, and audio and video equipment manufacturing

3380 navigational, measuring, electromedical, and control instruments manufacturing

3390 electronic component and product manufacturing, n.e.c.

3470 household appliance manufacturing

3490 electrical lighting and electrical equipment manufacturing, and other electrical component manufacturing, n.e.c.

3570 motor vehicles and motor vehicle equipment manufacturing

3580 aircraft and parts manufacturing

3590 aerospace product and parts manufacturing

3670 railroad rolling stock manufacturing

3680 ship and boat building

3690 other transportation equipment manufacturing

3770 sawmills and wood preservation

3780 veneer, plywood, and engineered wood products

3790 prefabricated wood buildings and mobile homes

3870 miscellaneous wood products

3890 furniture and related product manufacturing

3960 medical equipment and supplies manufacturing

3970 sporting and athletic goods, and doll, toy and game manufacturing

3980 miscellaneous manufacturing, n.e.c.

3990 not specified manufacturing industries

4070 motor vehicles, parts and supplies, merchant wholesalers

4080 furniture and home furnishing, merchant wholesalers

4090 lumber and other construction materials, merchant wholesalers

4170 professional and commercial equipment and supplies, merchant wholesalers

4180 metals and minerals, except petroleum, merchant wholesalers

4190 electrical and electronic goods, merchant wholesalers

4260 hardware, plumbing and heating equipment, and supplies, merchant wholesalers

4270 machinery, equipment, and supplies, merchant wholesalers

4280 recyclable material, merchant wholesalers

4290 miscellaneous durable goods, merchant wholesalers

4370 paper and paper products, merchant wholesalers

4380 drugs, sundries, and chemical and allied products, merchant wholesalers

4390 apparel, fabrics, and notions, merchant wholesalers

4470 groceries and related products, merchant wholesalers

4480 farm product raw materials, merchant wholesalers

4490 petroleum and petroleum products, merchant wholesalers

4560 alcoholic beverages, merchant wholesalers

4570 farm supplies, merchant wholesalers

4580 miscellaneous nondurable goods, merchant wholesalers

4585 wholesale electronic markets, agents and brokers

4590 not specified wholesale trade

4670 automobile dealers

4680 other motor vehicle dealers

4690 auto parts, accessories, and tire stores

4770 furniture and home furnishings stores

4780 household appliance stores

4790 radio, tv, and computer stores

4870 building material and supplies dealers

4880 hardware stores

4890 lawn and garden equipment and supplies stores

4970 grocery stores

4980 specialty food stores

4990 beer, wine, and liquor stores

5070 pharmacies and drug stores

5080 health and personal care, except drug, stores

5090 gasoline stations

5170 clothing stores

5180 shoe stores

5190 jewelry, luggage, and leather goods stores

5270 sporting goods, camera, and hobby and toy stores

5280 sewing, needlework, and piece goods stores

5290 music stores

5370 book stores and news dealers

5380 department stores and discount stores

5390 miscellaneous general merchandise stores

5470 retail florists

5480 office supplies and stationery stores

5490 used merchandise stores

5570 gift, novelty, and souvenir shops

5580 miscellaneous retail stores

5590 electronic shopping

5591 electronic auctions

5592 mail order houses

5670 vending machine operators

5680 fuel dealers

5690 other direct selling establishments

5790 not specified retail trade

6070 air transportation

6080 rail transportation

6090 water transportation

6170 truck transportation

6180 bus service and urban transit

6190 taxi and limousine service

6270 pipeline transportation

6280 scenic and sightseeing transportation

6290 services incidental to transportation

6370 postal service

6380 couriers and messengers

6390 warehousing and storage

6470 newspaper publishers

6480 periodical, book, and directory publishers

6490 software publishers

6570 motion pictures and video industries

6590 sound recording industries

6670 radio and television broadcasting and cable subscription programming

6672 internet publishing and broadcasting and web search portals

6680 wired telecommunications carriers

6690 other telecommunications services

6695 data processing, hosting, and related services

6770 libraries and archives

6780 other information services

6870 banking and related activities

6880 savings institutions, including credit unions

6890 non-depository credit and related activities

6970 securities, commodities, funds, trusts, and other financial investments

6990 insurance carriers and related activities

7070 real estate

7080 automotive equipment rental and leasing

7170 video tape and disk rental

7180 other consumer goods rental

7190 commercial, industrial, and other intangible assets rental and leasing

7270 legal services

7280 accounting, tax preparation, bookkeeping, and payroll services

7290 architectural, engineering, and related services

7370 specialized design services

7380 computer systems design and related services

7390 management, scientific, and technical consulting services

7460 scientific research and development services

7470 advertising and related services

7480 veterinary services

7490 other professional, scientific, and technical services

7570 management of companies and enterprises

7580 employment services

7590 business support services

7670 travel arrangements and reservation services

7680 investigation and security services

7690 services to buildings and dwellings

7770 landscaping services

7780 other administrative and other support services

7790 waste management and remediation services

7860 elementary and secondary schools

7870 colleges and universities, including junior colleges

7880 business, technical, and trade schools and training

7890 other schools and instruction, and educational support services

7970 offices of physicians

7980 offices of dentists

7990 offices of chiropractors

8070 offices of optometrists

8080 offices of other health practitioners

8090 outpatient care centers

8170 home health care services

8180 other health care services

8190 hospitals

8270 nursing care facilities

8290 residential care facilities, without nursing

8370 individual and family services

8380 community food and housing, and emergency services

8390 vocational rehabilitation services

8470 child day care services

8560 independent artists, performing arts, spectator sports, and related industries

8570 museums, art galleries, historical sites, and similar institutions

8580 bowling centers

8590 other amusement, gambling, and recreation industries

8660 traveler accommodation

8670 recreational vehicle parks and camps, and rooming and boarding houses

8680 restaurants and other food services

8690 drinking places, alcoholic beverages

8770 automotive repair and maintenance

8780 car washes

8790 electronic and precision equipment repair and maintenance

8870 commercial and industrial machinery and equipment repair and maintenance

8880 personal and household goods repair and maintenance

8890 footwear and leather goods repair

8970 barber shops

8980 beauty salons

8990 nail salons and other personal care services

9070 drycleaning and laundry services

9080 funeral homes, and cemeteries and crematories

9090 other personal services

9160 religious organizations

9170 civic, social, advocacy organizations, and grantmaking and giving services

9180 labor unions

9190 business, professional, political, and similar organizations

9290 private households

9370 executive offices and legislative bodies

9380 public finance activities

9390 other general government and support

9470 justice, public order, and safety activities

9480 administration of human resource programs

9490 administration of environmental quality and housing programs

9570 administration of economic programs and space research

9590 national security and international affairs

9670 u.s. army

9680 u.s. air force

9690 u.s. navy

9770 u.s. marine corps

9780 u.s. coast guard

9790 u.s. armed forces/n.s.

9870 active duty military personnel, all other

9997 uncodeable

9998 dk

9999 na

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(q) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook