Mothers census occupation code (2010)

maocc10

Source

General Social Survey https://gss.norc.org

Details

Question None

Values

10 chief executives

20 general and operations managers

30 legislators

40 advertising and promotions managers

50 marketing and sales managers

60 public relations and fundraising managers

100 administrative services managers

110 computer and information systems managers

120 financial managers

135 compensation and benefits managers

136 human resources managers

137 training and development managers

140 industrial production managers

150 purchasing managers

160 transportation, storage, and distribution managers

205 farmers, ranchers, and other agricultural managers

220 construction managers

230 education administrators

300 architectural and engineering managers

310 food service managers

325 funeral service managers

330 gaming managers

340 lodging managers

350 medical and health services managers

360 natural sciences managers

400 postmasters and mail superintendents

410 property, real estate, and community association managers

420 social and community service managers

425 emergency management directors

430 managers, all other

500 agents and business managers of artists, performers,

510 buyers and purchasing agents, farm products

520 wholesale and retail buyers, except farm products

530 purchasing agents, except wholesale, retail, and farm products

540 claims adjusters, appraisers, examiners, and investigators

565 compliance officers

600 cost estimators

630 human resources workers

640 compensation, benefits, and job analysis specialists

650 training and development specialists

700 logisticians

710 management analysts

725 meeting, convention, and event planners

726 fundraisers

735 market research analysts and marketing specialists

740 business operations specialists, all other

800 accountants and auditors

810 appraisers and assessors of real estate

820 budget analysts

830 credit analysts

840 financial analysts

850 personal financial advisors

860 insurance underwriters

900 financial examiners

910 credit counselors and loan officers

930 tax examiners and collectors, and revenue agents

940 tax preparers

950 financial specialists, all other

1005 computer and information research scientists

1006 computer systems analysts

1007 information security analysts

1010 computer programmers

1020 software developers, applications and systems software

1030 web developers

1050 computer support specialists

1060 database administrators

1105 network and computer systems administrators

1106 computer network architects

1107 computer occupations, all other

1200 actuaries

1210 mathematicians

1220 operations research analysts

1230 statisticians

1240 miscellaneous mathematical science occupations

1300 architects, except naval

1310 surveyors, cartographers, and photogrammetrists

1320 aerospace engineers

1330 agricultural engineers

1340 biomedical engineers

1350 chemical engineers

1360 civil engineers

1400 computer hardware engineers

1410 electrical and electronics engineers

1420 environmental engineers

1430 industrial engineers, including health and safety

1440 marine engineers and naval architects

1450 materials engineers

1460 mechanical engineers

1500 mining and geological engineers, including mining safety engineers

1510 nuclear engineers

1520 petroleum engineers

1530 engineers, all other

1540 drafters

1550 engineering technicians, except drafters

1560 surveying and mapping technicians

1600 agricultural and food scientists

1610 biological scientists

1640 conservation scientists and foresters

1650 medical scientists

1660 life scientists, all other

1700 astronomers and physicists

1710 atmospheric and space scientists

1720 chemists and materials scientists

1740 environmental scientists and geoscientists

1760 physical scientists, all other

1800 economists

1815 survey researchers

1820 psychologists

1830 sociologists

1840 urban and regional planners

1860 miscellaneous social scientists and related workers

1900 agricultural and food science technicians

1910 biological technicians

1920 chemical technicians

1930 geological and petroleum technicians

1940 nuclear technicians

1950 social science research assistants

1965 miscellaneous life, physical, and social science technicians

2000 counselors

2010 social workers

2015 probation officers and correctional treatment specialists

2016 social and human service assistants

2025 miscellaneous community and social service specialists,

2040 clergy

2050 directors, religious activities and education

2060 religious workers, all other

2100 lawyers

2105 judicial law clerks

2110 judges, magistrates, and other judicial workers

2145 paralegals and legal assistants

2160 miscellaneous legal support workers

2200 postsecondary teachers

2300 preschool and kindergarten teachers

2310 elementary and middle school teachers

2320 secondary school teachers

2330 special education teachers

2340 other teachers and instructors

2400 archivists, curators, and museum technicians

2430 librarians

2440 library technicians

2540 teacher assistants

2550 other education, training, and library workers

2600 artists and related workers

2630 designers

2700 actors

2710 producers and directors

2720 athletes, coaches, umpires, and related workers

2740 dancers and choreographers

2750 musicians, singers, and related workers

2760 entertainers and performers, sports and related workers, all other

2800 announcers

2810 news analysts, reporters and correspondents

2825 public relations specialists

2830 editors

2840 technical writers

2850 writers and authors

2860 miscellaneous media and communication workers

2900 broadcast and sound engineering technicians and radio

2910 photographers

2920 television, video, and motion picture camera operators and

2960 media and communication equipment workers, all other

3000 chiropractors

3010 dentists

3030 dietitians and nutritionists

3040 optometrists

3050 pharmacists

3060 physicians and surgeons

3110 physician assistants

3120 podiatrists

3140 audiologists

3150 occupational therapists

3160 physical therapists

3200 radiation therapists

3210 recreational therapists

3220 respiratory therapists

3230 speech-language pathologists

3235 exercise physiologists

3245 therapists, all other

3250 veterinarians

3255 registered nurses

3256 nurse anesthetists

3257 nurse midwives

3258 nurse practitioners

3260 health diagnosing and treating practitioners, all other

3300 clinical laboratory technologists and technicians

3310 dental hygienists

3320 diagnostic related technologists and technicians

3400 emergency medical technicians and paramedics

3420 health practitioner support technologists and technicians

3500 licensed practical and licensed vocational nurses

3510 medical records and health information technicians

3520 opticians, dispensing

3535 miscellaneous health technologists and technicians

3540 other healthcare practitioners and technical occupations

3600 nursing, psychiatric, and home health aides

3610 occupational therapy assistants and aides

3620 physical therapist assistants and aides

3630 massage therapists

3640 dental assistants

3645 medical assistants

3646 medical transcriptionists

3647 pharmacy aides

3648 veterinary assistants and laboratory animal caretakers

3649 phlebotomists

3655 miscellaneous healthcare support occupations, including medical equipment preparers

3700 first-line supervisors of correctional officers

3710 first-line supervisors of police and detectives

3720 first-line supervisors of fire fighting and prevention workers

3730 first-line supervisors of protective service workers, all other

3740 firefighters

3750 fire inspectors

3800 bailiffs, correctional officers, and jailers

3820 detectives and criminal investigators

3830 fish and game wardens

3840 parking enforcement workers

3850 police and sheriff's patrol officers

3860 transit and railroad police

3900 animal control workers

3910 private detectives and investigators

3930 security guards and gaming surveillance officers

3940 crossing guards

3945 transportation security screeners

3955 lifeguards and other recreational, and all other protective service workers

4000 chefs and head cooks

4010 first-line supervisors of food preparation and serving workers

4020 cooks

4030 food preparation workers

4040 bartenders

4050 combined food preparation and serving workers, including fast food

4060 counter attendants, cafeteria, food concession, and coffee shop

4110 waiters and waitresses

4120 food servers, nonrestaurant

4130 dining room and cafeteria attendants and bartender helpers

4140 dishwashers

4150 hosts and hostesses, restaurant, lounge, and coffee shop

4160 food preparation and serving related workers, all other

4200 first-line supervisors of housekeeping and janitorial workers

4210 first-line supervisors of landscaping, lawn service, and groundskeeping workers

4220 janitors and building cleaners

4230 maids and housekeeping cleaners

4240 pest control workers

4250 grounds maintenance workers

4300 first-line supervisors of gaming workers

4320 first-line supervisors of personal service workers

4340 animal trainers

4350 nonfarm animal caretakers

4400 gaming services workers

4410 motion picture projectionists

4420 ushers, lobby attendants, and ticket takers

4430 miscellaneous entertainment attendants and related workers

4460 embalmers and funeral attendants

4465 morticians, undertakers, and funeral directors

4500 barbers

4510 hairdressers, hairstylists, and cosmetologists

4520 miscellaneous personal appearance workers

4530 baggage porters, bellhops, and concierges

4540 tour and travel guides

4600 childcare workers

4610 personal care aides

4620 recreation and fitness workers

4640 residential advisors

4650 personal care and service workers, all other

4700 first-line supervisors of retail sales workers

4710 first-line supervisors of non-retail sales workers

4720 cashiers

4740 counter and rental clerks

4750 parts salespersons

4760 retail salespersons

4800 advertising sales agents

4810 insurance sales agents

4820 securities, commodities, and financial services sales agents

4830 travel agents

4840 sales representatives, services, all other

4850 sales representatives, wholesale and manufacturing

4900 models, demonstrators, and product promoters

4920 real estate brokers and sales agents

4930 sales engineers

4940 telemarketers

4950 door-to-door sales workers, news and street vendors, and related workers

4965 sales and related workers, all other

5000 first-line supervisors of office and administrative support workers

5010 switchboard operators, including answering service

5020 telephone operators

5030 communications equipment operators, all other

5100 bill and account collectors

5110 billing and posting clerks

5120 bookkeeping, accounting, and auditing clerks

5130 gaming cage workers

5140 payroll and timekeeping clerks

5150 procurement clerks

5160 tellers

5165 financial clerks, all other

5200 brokerage clerks

5210 correspondence clerks

5220 court, municipal, and license clerks

5230 credit authorizers, checkers, and clerks

5240 customer service representatives

5250 eligibility interviewers, government programs

5260 file clerks

5300 hotel, motel, and resort desk clerks

5310 interviewers, except eligibility and loan

5320 library assistants, clerical

5330 loan interviewers and clerks

5340 new accounts clerks

5350 order clerks

5360 human resources assistants, except payroll and timekeeping

5400 receptionists and information clerks

5410 reservation and transportation ticket agents and travel clerks

5420 information and record clerks, all other

5500 cargo and freight agents

5510 couriers and messengers

5520 dispatchers

5530 meter readers, utilities

5540 postal service clerks

5550 postal service mail carriers

5560 postal service mail sorters, processors, and processing machine operators

5600 production, planning, and expediting clerks

5610 shipping, receiving, and traffic clerks

5620 stock clerks and order fillers

5630 weighers, measurers, checkers, and samplers, recordkeeping

5700 secretaries and administrative assistants

5800 computer operators

5810 data entry keyers

5820 word processors and typists

5830 desktop publishers

5840 insurance claims and policy processing clerks

5850 mail clerks and mail machine operators, except postal service

5860 office clerks, general

5900 office machine operators, except computer

5910 proofreaders and copy markers

5920 statistical assistants

5940 office and administrative support workers, all other

6005 first-line supervisors of farming, fishing, and forestry workers

6010 agricultural inspectors

6020 animal breeders

6040 graders and sorters, agricultural products

6050 miscellaneous agricultural workers

6100 fishers and related fishing workers

6110 hunters and trappers

6120 forest and conservation workers

6130 logging workers

6200 first-line supervisors of construction trades and extraction workers

6210 boilermakers

6220 brickmasons, blockmasons, and stonemasons

6230 carpenters

6240 carpet, floor, and tile installers and finishers

6250 cement masons, concrete finishers, and terrazzo workers

6260 construction laborers

6300 paving, surfacing, and tamping equipment operators

6310 pile-driver operators

6320 operating engineers and other construction equipment operators

6330 drywall installers, ceiling tile installers, and tapers

6355 electricians

6360 glaziers

6400 insulation workers

6420 painters, construction and maintenance

6430 paperhangers

6440 pipelayers, plumbers, pipefitters, and steamfitters

6460 plasterers and stucco masons

6500 reinforcing iron and rebar workers

6515 roofers

6520 sheet metal workers

6530 structural iron and steel workers

6540 solar photovoltaic installers

6600 helpers, construction trades

6660 construction and building inspectors

6700 elevator installers and repairers

6710 fence erectors

6720 hazardous materials removal workers

6730 highway maintenance workers

6740 rail-track laying and maintenance equipment operators

6750 septic tank servicers and sewer pipe cleaners

6765 miscellaneous construction and related workers

6800 derrick, rotary drill, and service unit operators, oil, gas, and mining

6820 earth drillers, except oil and gas

6830 explosives workers, ordnance handling experts, and blasters

6840 mining machine operators

6910 roof bolters, mining

6920 roustabouts, oil and gas

6930 helpers-extraction workers

6940 other extraction workers

7000 first-line supervisors of mechanics, installers, and repairers

7010 computer, automated teller, and office machine repairers

7020 radio and telecommunications equipment installers and repairers

7030 avionics technicians

7040 electric motor, power tool, and related repairers

7050 electrical and electronics installers and repairers, transportation

7100 electrical and electronics repairers, industrial and utility

7110 electronic equipment installers and repairers, motor vehicles

7120 electronic home entertainment equipment installers and repairers

7130 security and fire alarm systems installers

7140 aircraft mechanics and service technicians

7150 automotive body and related repairers

7160 automotive glass installers and repairers

7200 automotive service technicians and mechanics

7210 bus and truck mechanics and diesel engine specialists

7220 heavy vehicle and mobile equipment service technicians and mechanics

7240 small engine mechanics

7260 miscellaneous vehicle and mobile equipment mechanics, installers, and repairers

7300 control and valve installers and repairers

7315 heating, air conditioning, and refrigeration mechanics and installers

7320 home appliance repairers

7330 industrial and refractory machinery mechanics

7340 maintenance and repair workers, general

7350 maintenance workers, machinery

7360 millwrights

7410 electrical power-line installers and repairers

7420 telecommunications line installers and repairers

7430 precision instrument and equipment repairers

7440 wind turbine service technicians

7510 coin, vending, and amusement machine servicers and repairers

7520 commercial divers

7540 locksmiths and safe repairers

7550 manufactured building and mobile home installers

7560 riggers

7600 signal and track switch repairers

7610 helpers-installation, maintenance, and repair workers

7630 other installation, maintenance, and repair workers

7700 first-line supervisors of production and operating workers

7710 aircraft structure, surfaces, rigging, and systems assemblers

7720 electrical, electronics, and electromechanical assemblers

7730 engine and other machine assemblers

7740 structural metal fabricators and fitters

7750 miscellaneous assemblers and fabricators

7800 bakers

7810 butchers and other meat, poultry, and fish processing workers

7830 food and tobacco roasting, baking, and drying machine operators and tenders

7840 food batchmakers

7850 food cooking machine operators and tenders

7855 food processing workers, all other

7900 computer control programmers and operators

7920 extruding and drawing machine setters, operators, and tenders, metal and plastic

7930 forging machine setters, operators, and tenders, metal and plastic

7940 rolling machine setters, operators, and tenders, metal and plastic

7950 cutting, punching, and press machine setters, operators, and tenders, metal and plastic

7960 drilling and boring machine tool setters, operators, and tenders, metal and plastic

8000 grinding, lapping, polishing, and buffing machine tool setters, operators, and tenders, metal and plastic

8010 lathe and turning machine tool setters, operators, and tenders, metal and plastic

8020 milling and planing machine setters, operators, and tenders, metal and plastic

8030 machinists

8040 metal furnace operators, tenders, pourers, and casters

8060 model makers and patternmakers, metal and plastic

8100 molders and molding machine setters, operators, and tenders, metal and plastic

8120 multiple machine tool setters, operators, and tenders, metal and plastic

8130 tool and die makers

8140 welding, soldering, and brazing workers

8150 heat treating equipment setters, operators, and tenders, metal and plastic

8160 layout workers, metal and plastic

8200 plating and coating machine setters, operators, and tenders, metal and plastic

8210 tool grinders, filers, and sharpeners

8220 metal workers and plastic workers, all other

8250 prepress technicians and workers

8255 printing press operators

8256 print binding and finishing workers

8300 laundry and dry-cleaning workers

8310 pressers, textile, garment, and related materials

8320 sewing machine operators

8330 shoe and leather workers and repairers

8340 shoe machine operators and tenders

8350 tailors, dressmakers, and sewers

8360 textile bleaching and dyeing machine operators and tenders

8400 textile cutting machine setters, operators, and tenders

8410 textile knitting and weaving machine setters, operators, and tenders

8420 textile winding, twisting, and drawing out machine setters, operators, and tenders

8430 extruding and forming machine setters, operators, and tenders, synthetic and glass fibers

8440 fabric and apparel patternmakers

8450 upholsterers

8460 textile, apparel, and furnishings workers, all other

8500 cabinetmakers and bench carpenters

8510 furniture finishers

8520 model makers and patternmakers, wood

8530 sawing machine setters, operators, and tenders, wood

8540 woodworking machine setters, operators, and tenders, except sawing

8550 woodworkers, all other

8600 power plant operators, distributors, and dispatchers

8610 stationary engineers and boiler operators

8620 water and wastewater treatment plant and system operators

8630 miscellaneous plant and system operators

8640 chemical processing machine setters, operators, and tenders

8650 crushing, grinding, polishing, mixing, and blending workers

8710 cutting workers

8720 extruding, forming, pressing, and compacting machine setters,

8730 operators, and tenders furnace, kiln, oven, drier, and kettle operators and tenders

8740 inspectors, testers, sorters, samplers, and weighers

8750 jewelers and precious stone and metal workers

8760 medical, dental, and ophthalmic laboratory technicians

8800 packaging and filling machine operators and tenders

8810 painting workers

8830 photographic process workers and processing machine operators

8840 semiconductor processors

8850 adhesive bonding machine operators and tenders

8860 cleaning, washing, and metal pickling equipment operators and tenders

8900 cooling and freezing equipment operators and tenders

8910 etchers and engravers

8920 molders, shapers, and casters, except metal and plastic

8930 paper goods machine setters, operators, and tenders

8940 tire builders

8950 helpers-production workers

8965 production workers, all other

9000 supervisors of transportation and material moving workers

9030 aircraft pilots and flight engineers

9040 air traffic controllers and airfield operations specialists

9050 flight attendants

9110 ambulance drivers and attendants, except emergency medical technicians

9120 bus drivers

9130 driver/sales workers and truck drivers

9140 taxi drivers and chauffeurs

9150 motor vehicle operators, all other

9200 locomotive engineers and operators

9230 railroad brake, signal, and switch operators

9240 railroad conductors and yardmasters

9260 subway, streetcar, and other rail transportation workers

9300 sailors and marine oilers

9310 ship and boat captains and operators

9330 ship engineers

9340 bridge and lock tenders

9350 parking lot attendants

9360 automotive and watercraft service attendants

9410 transportation inspectors

9415 transportation attendants, except flight attendants

9420 other transportation workers

9500 conveyor operators and tenders

9510 crane and tower operators

9520 dredge, excavating, and loading machine operators

9560 hoist and winch operators

9600 industrial truck and tractor operators

9610 cleaners of vehicles and equipment

9620 laborers and freight, stock, and material movers, hand

9630 machine feeders and offbearers

9640 packers and packagers, hand

9650 pumping station operators

9720 refuse and recyclable material collectors

9730 mine shuttle car operators

9740 tank car, truck, and ship loaders

9750 material moving workers, all other

9800 military officer special and tactical operations leaders

9810 first-line enlisted military supervisors

9820 military enlisted tactical operations and air/weapons specialists and crew members

9830 military, rank not specified

9997 uncodeable

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(q) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook