Education:version y

nateducy

Source

General Social Survey https://gss.norc.org

Details

Question 69. We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. g. Education

Overview

For further details see the official GSS documentation.

Counts by year:

year	iap	about right	don't know	no answer	too little	too much	skipped on web	Total

1972	1613	-	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	-	1504

1974	1484	-	-	-	-	-	-	1484

1975	1490	-	-	-	-	-	-	1490

1976	1499	-	-	-	-	-	-	1499

1977	1530	-	-	-	-	-	-	1530

1978	1532	-	-	-	-	-	-	1532

1980	1468	-	-	-	-	-	-	1468

1982	1860	-	-	-	-	-	-	1860

1983	1599	-	-	-	-	-	-	1599

1984	974	133	8	11	311	36	-	1473

1985	751	202	15	2	510	54	-	1534

1986	730	197	12	8	486	37	-	1470

1987	603	307	28	14	811	56	-	1819

1988	718	191	12	8	517	35	-	1481

1989	768	175	8	7	551	28	-	1537

1990	674	149	16	14	497	22	-	1372

1991	756	175	17	15	523	31	-	1517

1993	799	178	15	13	562	39	-	1606

1994	1508	313	16	36	1048	71	-	2992

1996	1441	262	20	59	1053	69	-	2904

1998	1381	274	25	8	1058	86	-	2832

2000	1409	273	17	3	1053	62	-	2817

2002	1358	252	15	2	1072	66	-	2765

2004	1401	239	10	1	1089	72	-	2812

2006	3002	281	21	-	1111	95	-	4510

2008	1004	196	11	-	756	56	-	2023

2010	1014	190	6	1	780	53	-	2044

2012	998	167	11	-	727	71	-	1974

2014	1274	220	18	-	932	94	-	2538

2016	1437	257	12	-	1060	101	-	2867

2018	1188	181	9	-	904	66	-	2348

2021	1987	346	4	-	1532	153	10	4032

2022	1780	277	7	-	1357	121	2	3544

Total	44534	5435	333	202	20300	1574	12	72390

Values

1 too little

2 about right

3 too much

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(q) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook