Other protestant denominations

other

Source

General Social Survey https://gss.norc.org

Details

Question None

Overview

For further details see the official GSS documentation.

Counts by year:

year	7th day adventist	advent children	assembly of god	baptist (northern)	brethren church, brethren	christ adelphians	christian & missionary alliance	christian scientist	christian; central christian	church of christ	church of god in christ	churches of god (except with christ and holiness)	community church	congregationalist, 1st congreg	covenant	disciples of christ	dutch reform	evangelical reformed	evangelical, evangelist	first christian	four square gospel	friends	holiness; church of holiness	iap	jehovah's witnesses	lds--reorganized	mennonite	mind science	mormon	nazarene	other fundamentalist	pentecostal	quaker	reformed	reformed united church of christ	religious science	salvation army	unitarian, universalist	united church of christ	united church, unity church	wesleyan	apostolic faith	christ church unity	christian catholic	disciples of god	don't know	evangelist free church	first christian disciples of christ	full gospel	holiness (nazarene)	mission covenant	pentecostal assembly of god	pentecostal church of god	pentecostal holiness, holiness pentecostal	reformed church of christ	united brethren, united brethren in christ	united holiness	unity	wesleyan methodist--pilgrim	zion union	zion union apostolic--reformed	christian reform	church of the living god	first church	no answer	swedish mission	holiness church of god	missionary baptist	united church of christianity	christian disciples	eden evangelist	moravian	evangelical covenant	sanctified, sanctification	church of god, saint & christ	free will baptist	spiritualist	witness holiness	church of prophecy	evangelical congregational	ind bible, bible, bible fellowship	new testament christian	african methodist	brethren, plymouth	charismatic	church of god in christ holiness	federated church	holy roller	house of prayer	pentecostal apostolic	apostolic christian	bible missionary	calvary bible	christ cathedral of truth	christ in christian union	church universal and triumphant	evangelical methodist	free methodist	grace brethren	latvian lutheran	worldwide church of god	first reformed	mennonite brethren	missionary church	open bible	pilgrim holiness	the way ministry	zion union apostolic	church of christ, evangelical	independent	the church of god of prophecy	chapel of faith	christian calvary chapel	evangelical united brethren	faith gospel tabernacle	polish national church	church of daniel's band	brother of christ	church of the first born	chinese gospel church	independent fundamental church of america	primitive baptist	grace reformed	new age spirituality	new song	united church of canada	christian tabernacle	new birth christian	american reform	apostolic church	assyrian evangelist church	church of jesus christ of the restoration	laotian christian	spirit of christ	living word	zwinglian	course in miracles	world overcomer outreach ministry	spirit filled	unity of the brethren	amish	christian union	community christian fellowship	community of christ	new hope christian fellowship	cathedral of joy	cornerstone church	great faith ministries	greater new testament church	harvest church	life sanctuary	rain on us deliverance ministries	real life ministries	sanctuary	shephard's chapel	shield of faith ministries	the word church	united christian	vineyard church	word of faith church	alliance	born again	journeys	national progressive baptist	metropolitan community	new apostolic	empowerment temple	faith covenant	grace independent baptist church	true light church of christ	assembly of christ	new life	pathways christian church	the amana church	the legacy church	armenian apostolic church	calvary	disciple of jesus	ethiopian evangelical church	hebrew israelites	hebrew roots	scandinavian church	messianic jew	reformed church of jesus christ of latter day saints	skipped on web	good united	pilgrim church	Total

1972	6	1	9	1	7	1	1	2	10	20	3	9	1	20	1	2	1	2	4	3	1	1	8	1428	5	2	5	1	15	1	1	18	2	5	1	2	2	4	5	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1613

1973	6	1	4	-	-	1	-	3	13	11	2	16	-	7	-	1	2	-	2	3	1	1	10	1333	9	3	1	-	6	7	1	9	1	3	-	1	3	2	10	-	2	1	1	1	1	5	1	1	2	1	1	1	1	3	3	1	1	1	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1504

1974	9	-	5	-	-	-	-	2	21	20	1	8	1	10	2	1	1	-	3	2	-	1	4	1309	7	2	-	-	8	5	-	13	2	5	1	-	1	3	12	-	1	2	-	2	-	8	-	-	-	1	-	1	-	1	-	1	-	1	1	-	-	1	1	2	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1484

1975	4	-	3	-	3	-	1	5	18	19	2	9	-	12	-	3	1	1	3	1	-	1	9	1305	8	-	-	-	10	5	-	16	-	7	2	1	1	1	11	1	1	2	-	-	-	3	-	2	-	-	-	2	-	1	-	2	1	1	-	-	-	4	-	-	3	-	1	1	3	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1490

1976	12	-	11	-	2	-	-	6	24	17	1	7	-	15	-	2	1	-	4	2	-	-	7	1304	13	-	1	-	9	6	4	12	1	4	-	-	1	2	12	1	1	3	-	-	-	-	-	1	-	-	1	-	2	4	-	-	-	2	-	-	-	-	-	1	-	-	-	-	-	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1499

1977	6	-	10	-	1	1	-	2	15	22	1	16	-	13	-	1	-	-	1	4	-	1	5	1333	7	1	-	-	11	8	2	20	-	4	-	2	1	4	10	-	-	1	-	-	-	-	2	2	2	1	-	-	-	1	-	3	-	2	-	-	-	1	1	-	7	-	1	-	-	-	-	-	1	3	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1530

1978	8	-	13	-	1	-	1	2	17	21	-	10	1	13	-	5	-	1	4	3	-	3	4	1319	10	3	1	-	10	7	1	9	-	2	-	1	3	6	14	-	-	4	-	-	-	-	2	-	1	-	2	1	-	8	-	1	-	1	1	-	-	5	-	-	5	-	1	1	-	-	-	1	-	-	1	2	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1532

1980	4	-	11	-	-	-	1	4	14	25	2	15	1	9	-	1	-	-	2	-	-	-	5	1270	8	2	1	-	18	4	-	17	1	1	-	1	1	4	10	1	1	1	1	-	-	-	-	-	1	-	-	1	-	5	-	2	-	3	-	-	1	3	-	1	4	-	-	-	1	-	-	3	-	-	-	-	-	-	1	1	3	2	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1468

1982	4	2	11	-	4	-	-	7	21	15	6	12	-	15	-	4	1	-	2	5	3	-	14	1603	10	-	-	-	10	8	1	19	2	7	-	1	-	2	12	-	1	3	-	-	-	-	-	1	-	9	-	1	2	2	-	-	-	1	-	-	-	2	-	-	13	-	-	-	4	1	-	1	-	2	1	1	4	-	1	-	-	-	1	1	1	1	2	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1860

1983	9	-	9	-	1	-	2	2	17	17	2	8	-	17	1	-	2	2	2	2	1	-	3	1364	14	1	-	1	32	7	3	23	2	3	-	-	2	2	11	-	2	-	-	-	-	-	1	-	1	-	-	-	-	3	-	1	-	1	-	-	-	2	1	-	6	-	-	-	1	-	-	-	-	1	1	-	2	-	-	-	-	-	-	-	-	2	-	-	-	1	1	1	1	1	1	1	1	1	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1599

1984	7	-	7	-	-	-	-	3	21	25	1	11	1	6	-	2	1	-	-	4	1	-	7	1241	9	3	-	-	30	9	-	15	2	2	-	1	-	6	3	1	2	-	-	1	-	1	-	-	2	-	-	1	2	2	-	1	-	1	-	-	-	5	-	1	14	-	-	4	-	1	-	-	-	-	-	3	-	-	-	1	2	-	-	-	-	2	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	1	1	1	1	1	2	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1473

1985	2	-	8	-	1	-	-	2	26	18	1	10	-	15	-	9	1	-	3	3	-	-	7	1313	7	-	-	-	30	9	1	19	1	-	-	1	1	3	6	2	2	-	-	-	-	-	1	2	2	-	-	-	-	1	1	1	-	1	-	-	-	5	-	-	7	-	-	-	-	-	-	1	1	-	-	-	-	-	-	-	1	-	-	1	1	1	-	-	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	1	1	-	-	-	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1534

1986	4	-	8	-	2	-	1	3	18	28	1	5	-	12	-	2	-	-	-	4	1	-	3	1246	9	-	8	-	36	10	-	18	-	-	-	-	3	5	10	-	2	-	-	1	1	1	1	1	1	1	-	-	2	3	-	1	-	1	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	1	1	-	-	-	-	-	1	-	-	-	-	2	-	-	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	-	-	-	2	-	1	1	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1470

1987	7	-	11	-	3	-	1	3	26	22	7	13	3	11	-	4	-	1	2	4	1	-	10	1510	18	1	-	-	30	7	2	31	4	6	-	1	-	6	5	-	1	8	-	-	-	-	1	-	3	4	-	1	1	3	1	1	-	1	-	-	-	3	-	-	13	-	-	5	1	-	-	-	-	2	-	1	2	-	-	-	-	-	-	1	-	2	-	-	-	1	1	1	-	-	-	-	-	-	-	-	1	-	-	3	-	-	-	-	-	3	-	-	2	-	-	-	2	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1819

1988	4	-	21	-	2	-	-	1	5	20	-	9	-	9	1	5	1	-	2	6	2	2	6	1261	8	1	-	1	33	8	2	23	1	3	-	1	-	2	6	-	-	3	-	-	-	-	-	-	2	2	-	-	1	1	-	1	-	2	-	-	-	2	-	-	7	-	-	1	1	-	-	-	-	3	-	-	-	-	-	-	-	-	-	-	1	-	-	-	-	1	1	-	-	-	-	-	-	-	1	-	-	-	-	-	2	-	-	-	-	-	-	-	-	-	-	-	-	1	2	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1481

1989	6	-	11	-	4	-	1	2	5	16	1	8	-	6	-	2	2	-	3	-	-	-	6	1327	10	-	1	-	30	6	-	21	1	5	-	1	-	4	13	-	2	1	-	-	-	-	1	-	1	2	-	-	-	7	-	1	-	1	-	-	-	4	-	-	8	-	-	4	1	-	-	-	-	1	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	1	1	-	-	1	-	-	-	-	-	-	-	1	-	-	-	-	-	-	3	-	-	-	-	-	-	-	-	-	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1537

1990	7	1	13	-	1	-	2	1	20	16	2	4	-	13	-	8	1	-	-	1	2	-	5	1153	5	-	1	-	25	2	6	14	1	1	-	1	-	3	6	-	-	-	-	-	-	1	-	-	1	3	-	1	1	2	-	1	-	-	2	-	-	6	-	1	8	-	-	6	1	2	-	-	-	1	-	3	1	-	-	-	-	-	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	-	1	-	-	-	-	4	-	-	-	-	3	1	-	-	-	-	-	-	-	-	-	-	-	1	1	1	1	2	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1372

1991	4	-	9	-	2	-	1	-	27	22	3	14	1	5	-	5	3	-	3	6	1	-	10	1254	10	1	4	-	32	10	3	27	2	-	-	1	-	7	8	-	1	-	-	-	-	-	1	-	1	1	-	2	3	2	-	-	-	1	-	-	-	1	-	-	4	-	1	8	-	-	-	-	-	1	-	3	-	-	-	1	2	-	-	-	-	-	-	-	-	-	-	1	-	-	-	-	1	-	-	-	-	-	-	1	-	1	-	-	-	2	-	-	-	-	-	-	-	-	-	-	-	-	-	1	-	-	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1517

1993	7	-	10	-	4	-	1	1	18	26	2	12	-	7	-	6	2	-	2	3	-	-	9	1379	13	-	1	-	20	4	2	19	1	1	-	3	1	4	4	-	4	1	-	-	-	-	1	-	2	-	-	-	1	3	-	1	-	4	-	-	-	13	-	1	-	-	-	3	-	-	-	1	-	-	-	7	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1606

1994	13	-	20	-	4	-	3	5	25	35	4	15	-	20	1	8	1	-	3	5	2	1	2	2643	22	-	2	-	35	8	2	31	3	1	-	2	-	14	6	3	1	-	-	-	-	-	2	-	3	2	-	1	-	8	1	3	-	2	-	-	-	2	-	-	14	-	1	-	-	-	-	1	-	1	-	-	1	-	-	-	1	-	-	-	1	-	-	-	-	2	1	-	-	-	-	-	-	-	1	-	1	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2992

1996	13	-	22	2	8	-	2	-	11	23	1	12	1	13	-	3	1	2	4	7	2	-	2	2596	16	-	3	-	22	10	2	48	1	-	3	2	-	8	9	2	-	-	-	-	-	-	1	1	-	1	-	4	2	3	-	1	-	1	-	1	-	3	-	1	8	-	1	-	-	-	-	1	1	1	-	-	3	-	-	-	-	-	-	-	1	1	-	1	-	-	-	-	-	-	-	-	-	-	-	-	1	1	-	-	-	-	-	-	-	6	-	-	1	1	-	-	-	-	-	1	-	-	-	-	-	1	-	-	-	1	-	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2904

1998	13	-	11	-	1	-	2	2	12	18	-	18	1	12	1	4	-	-	3	4	1	-	3	2536	26	1	2	-	12	7	-	41	2	2	-	1	-	8	8	-	2	-	-	-	-	-	2	1	2	1	-	1	1	8	-	2	-	4	-	-	-	3	-	-	32	-	-	-	-	-	-	1	-	-	-	-	1	-	-	-	2	-	-	-	-	2	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	3	-	-	-	-	-	-	-	-	-	-	-	-	-	3	-	-	1	-	1	4	-	-	-	-	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2832

2000	12	-	18	-	3	-	2	2	11	31	5	11	1	16	1	2	-	-	2	1	1	-	3	2514	17	-	2	-	19	11	4	44	3	3	-	2	-	10	8	-	3	2	-	-	-	-	2	-	2	-	-	2	1	2	-	-	-	7	-	-	-	5	-	-	18	-	-	-	-	1	-	-	-	-	-	1	-	-	-	-	2	-	-	-	1	3	-	-	-	-	-	-	-	-	-	-	-	1	-	-	-	-	-	1	1	-	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2817

2002	2	-	15	-	-	2	-	4	19	24	2	11	-	11	1	3	-	-	14	-	-	-	7	2410	26	-	4	-	22	6	3	51	2	-	-	1	1	8	9	-	3	4	-	-	1	14	1	-	1	-	-	-	-	2	-	3	-	3	-	-	-	4	-	2	22	-	-	10	-	-	-	-	-	13	-	8	2	-	-	-	-	1	-	-	1	-	-	-	-	-	1	-	-	-	-	-	-	3	-	-	-	-	-	-	-	-	-	-	-	4	-	-	-	-	-	-	-	-	-	-	-	2	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2765

2004	9	3	12	2	-	-	-	-	26	20	2	6	1	11	-	2	-	-	5	-	4	-	6	2448	18	-	1	1	42	14	1	55	3	2	-	1	-	8	8	-	-	2	-	-	-	1	5	-	4	-	-	1	-	6	1	1	-	-	-	-	-	4	-	-	19	-	-	11	-	-	-	1	1	1	-	11	3	-	-	-	4	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	5	-	-	-	1	-	-	-	-	-	-	-	6	-	-	-	-	-	-	-	-	-	-	-	2	-	-	-	-	-	-	-	2	-	-	-	-	-	-	-	-	-	-	1	4	1	2	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2812

2006	17	3	25	-	2	-	2	2	26	48	9	8	5	15	2	6	1	-	13	-	5	1	14	3901	25	1	1	-	70	13	5	95	6	4	-	-	2	10	12	-	6	10	-	-	-	2	7	-	1	-	-	-	-	4	-	-	-	1	-	-	-	12	1	-	33	-	-	9	-	-	-	4	-	1	-	22	3	-	-	-	3	-	-	1	-	3	-	-	1	1	-	-	-	-	-	-	-	2	-	-	-	-	-	-	3	-	-	-	1	6	-	-	-	-	-	1	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	5	-	-	1	1	1	1	1	1	1	1	2	1	1	1	1	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	4510

2008	10	1	4	-	1	-	3	1	19	18	1	9	-	6	-	2	2	-	11	-	1	-	4	1725	19	-	1	-	34	2	2	44	9	3	-	2	-	9	3	-	2	4	-	1	-	10	-	-	2	2	-	1	-	1	-	-	-	1	-	-	-	3	-	-	11	-	-	7	-	-	-	2	1	-	-	8	1	-	-	-	3	-	-	-	1	-	-	-	-	1	-	-	-	-	-	-	-	2	-	-	-	-	-	4	-	1	-	-	-	3	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2023

2010	7	-	5	-	-	-	-	1	22	19	1	9	1	7	-	4	1	-	4	1	-	1	3	1822	13	-	3	1	15	4	-	37	2	4	-	-	-	3	4	1	1	4	-	-	-	1	1	-	1	-	-	-	-	-	-	-	-	-	-	-	-	2	-	-	16	-	-	3	-	-	-	-	-	1	-	6	3	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	1	-	-	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	-	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2044

2012	10	-	9	-	1	-	2	-	18	15	-	12	1	6	1	1	1	-	13	-	-	-	1	1740	14	-	2	-	16	4	-	40	5	3	1	1	-	5	9	-	-	2	-	-	-	-	2	-	3	-	-	-	-	6	-	-	-	-	-	-	-	2	-	-	4	-	-	3	-	-	-	-	-	-	-	8	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	3	-	-	-	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2	-	1	-	-	1	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1974

2014	5	-	5	-	2	-	2	2	13	17	2	1	1	7	1	4	-	-	6	-	1	-	6	2263	23	-	1	-	33	10	-	54	3	3	-	-	1	7	6	-	2	-	-	-	-	1	2	-	2	-	-	-	2	2	-	1	-	-	-	-	-	3	-	1	16	-	-	6	-	-	-	-	1	-	-	3	1	-	-	-	-	-	-	-	1	1	-	-	-	-	-	-	-	-	-	-	-	2	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	-	4	-	-	-	-	-	-	-	-	1	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2538

2016	7	1	11	-	6	-	1	-	27	22	2	8	-	12	-	3	-	-	16	1	-	-	5	2550	28	-	2	-	23	3	-	60	4	2	-	-	-	2	5	2	4	2	-	-	-	1	1	-	1	-	-	1	-	1	-	-	-	-	-	-	-	1	-	-	19	-	-	7	-	1	-	-	-	-	-	6	3	-	-	-	2	-	-	-	3	-	1	-	-	1	-	-	-	-	-	-	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2	-	-	-	-	-	-	-	-	1	-	-	1	1	-	-	-	-	-	-	-	-	-	-	-	-	2867

2018	14	-	10	-	1	-	1	2	21	13	3	5	-	6	1	3	-	1	10	-	-	-	7	2080	13	-	1	-	26	3	-	43	2	4	-	-	-	5	8	-	3	-	-	-	-	2	-	-	1	-	-	-	-	6	-	-	1	-	-	-	-	1	-	-	20	-	-	8	-	-	-	-	-	-	-	5	2	-	-	-	-	-	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	1	2	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	1	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	2	1	1	1	2	1	-	-	-	-	-	2348

2021	11	-	5	-	3	-	1	1	-	9	2	5	-	6	-	1	-	-	9	-	2	-	3	3734	22	-	1	-	39	7	-	34	2	6	-	-	-	8	9	-	4	5	-	-	-	-	-	-	-	-	-	1	-	2	-	1	-	1	1	-	-	-	-	-	46	-	-	5	-	-	-	-	1	-	-	-	-	-	-	-	1	-	-	-	1	1	-	-	-	-	-	-	-	-	-	-	-	2	-	-	-	-	1	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	1	32	-	-	4032

2022	11	-	6	-	-	-	-	-	-	15	3	4	-	4	-	3	-	1	12	-	-	-	-	3303	21	-	-	-	31	3	-	37	3	3	-	1	-	6	7	-	1	2	-	-	-	-	2	-	1	-	-	1	-	-	-	-	-	-	-	-	-	1	-	-	32	-	1	5	-	-	-	1	-	-	-	-	-	-	-	-	-	-	1	-	2	-	-	-	-	1	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	3	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	1	-	9	1	1	3544

Total	270	13	362	5	70	5	34	73	586	707	75	330	21	367	14	112	27	11	167	75	33	13	198	63517	483	22	50	5	834	228	48	1057	74	99	8	32	24	181	279	15	56	67	2	6	3	51	40	12	46	31	4	25	22	103	7	31	3	45	6	2	2	104	4	11	410	1	7	107	13	7	1	19	7	33	4	98	34	1	2	3	28	3	2	4	15	22	3	2	3	12	9	4	1	1	3	1	4	20	6	1	5	3	3	15	9	4	2	2	5	49	2	1	5	2	1	2	2	1	3	2	1	10	1	5	1	3	2	2	2	14	1	1	1	1	1	1	1	1	3	1	5	11	1	2	2	1	1	1	1	1	1	1	2	1	1	1	1	1	2	1	3	8	1	3	1	1	1	1	2	1	2	1	1	1	1	1	2	1	1	1	2	1	2	1	41	1	1	72390

Values

1 hungarian reformed

2 evangelical congregational

3 ind bible, bible, bible fellowship

5 church of prophecy

6 new testament christian

7 church of god, saint & christ

8 moravian

9 christian & missionary alliance

10 advent children

11 spiritualist

12 assembly of god

13 free methodist

14 apostolic faith

15 african methodist

16 free will baptist

17 eden evangelist

18 holiness (nazarene)

19 baptist (northern)

20 brethren church, brethren

21 witness holiness

22 brethren, plymouth

23 united brethren, united brethren in christ

24 independent

25 christian disciples

26 christ in christian union

27 open bible

28 christian catholic

29 christ church unity

30 christ adelphians

31 christian; central christian

32 christian reform

33 christian scientist

34 church of christ, evangelical

35 church of christ

36 churches of god (except with christ and holiness)

37 church of god in christ

38 church of god in christ holiness

39 church of the living god

40 congregationalist, 1st congreg

41 community church

42 covenant

43 dutch reform

44 disciples of christ

45 evangelical, evangelist

46 evangelical reformed

47 evangelist free church

48 first church

49 first christian disciples of christ

50 first reformed

51 first christian

52 full gospel

53 four square gospel

54 friends

55 holy roller

56 holiness; church of holiness

57 pilgrim holiness

58 jehovah's witnesses

59 lds

60 lds--mormon

61 lds--reorganized

62 lds--jesus christ; church of jesus lds

63 mennonite

64 mormon

65 nazarene

66 pentecostal assembly of god

67 pentecostal church of god

68 pentecostal

69 pentecostal holiness, holiness pentecostal

70 quaker

71 reformed

72 reformed united church of christ

73 reformed church of christ

74 religious science

75 mind science

76 salvation army

77 7th day adventist

78 sanctified, sanctification

79 united holiness

80 unitarian, universalist

81 united church of christ

82 united church, unity church

83 wesleyan

84 wesleyan methodist--pilgrim

85 zion union

86 zion union apostolic

87 zion union apostolic--reformed

88 disciples of god

89 grace reformed

90 holiness church of god

91 evangelical covenant

92 mission covenant

93 missionary baptist

94 swedish mission

95 unity

96 united church of christianity

97 other fundamentalist

98 federated church

99 american reform

100 grace brethren

101 christ in god

102 charismatic

103 pentecostal apostolic

104 house of prayer

105 latvian lutheran

106 triumph church of god

107 apostolic christian

108 christ cathedral of truth

109 bible missionary

110 calvary bible

111 amish

112 evangelical methodist

113 worldwide church of god

114 church universal and triumphant

115 mennonite brethren

116 church of the first born

117 missionary church

118 the way ministry

119 united church of canada

120 evangelical united brethren

121 the church of god of prophecy

122 chapel of faith

123 polish national church

124 faith gospel tabernacle

125 christian calvary chapel

126 carmelite

127 church of daniel's band

128 christian tabernacle

129 living word

130 true light church of christ

131 macedonia

132 brother of christ

133 primitive baptist

134 independent fundamental church of america

135 chinese gospel church

136 new age spirituality

137 new song

138 apostolic church

139 faith christian

140 people's church

141 new birth christian

142 unity school of christianity

143 assyrian evangelist church

144 spirit of christ

145 church of jesus christ of the restoration

146 laotian christian

148 schwenkfelder

149 polish catholic

150 zwinglian

151 world overcomer outreach ministry

152 course in miracles

153 unity of the brethren

154 spirit filled

155 christian union

156 church of living christ

157 community of christ

158 new hope christian fellowship

159 community christian fellowship

160 friends in christ

161 hawaiian ohana

162 reformed church of jesus christ of latter day saints

163 swedenurgian/churches of the new jerusalem

164 divine science

166 united christian

167 sanctuary

168 rain on us deliverance ministries

169 the word church

170 cornerstone church

171 life sanctuary

172 word of faith church

173 harvest church

174 shephard's chapel

175 greater new testament church

176 vineyard church

177 real life ministries

178 cathedral of joy

179 great faith ministries

180 shield of faith ministries

181 born again

182 alliance

183 jacobite apostolic

184 church of god of israel

185 journeys

186 national progressive baptist

187 new apostolic

188 metropolitan community

189 family life church

190 faith fellowship

191 faith covenant

192 new thought

193 free spirit ministry

194 jacobite syrian christian church

195 the ark church

196 empowerment temple

197 grace independent baptist church

198 new life

201 pathways christian church

204 universal church of the kingdom of god

205 assembly of christ

206 the amana church

207 the legacy church

208 calvary

210 ethiopian evangelical church

211 disciple of jesus

212 scandinavian church

213 hebrew roots

214 hebrew israelites

215 armenian apostolic church

216 word of jesus

217 old testament

218 apostolic church of jesus christ

219 messianic jew

220 pilgrim church

221 good united

222 hutterite

223 armenian evangelical church

224 3c usa church

225 israelite church of god

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(q) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook