Father's industry code (naics 2007)

paind10

Source

General Social Survey https://gss.norc.org

Details

Question None

Overview

For further details see the official GSS documentation.

Counts by year:

year	accounting, tax preparation, bookkeeping, and payroll services	administration of environmental quality and housing programs	administration of human resource programs	aerospace product and parts manufacturing	agricultural chemical manufacturing	air transportation	aircraft and parts manufacturing	alcoholic beverages, merchant wholesalers	animal food, grain, and oilseed milling	animal production	animal slaughtering and processing	apparel, fabrics, and notions, merchant wholesalers	architectural, engineering, and related services	auto parts, accessories, and tire stores	automobile dealers	automotive repair and maintenance	bakeries, except retail	banking and related activities	barber shops	beverage manufacturing	building material and supplies dealers	bus service and urban transit	carpet and rug mills	clothing stores	coal mining	colleges and universities, including junior colleges	commercial and industrial machinery and equipment repair and maintenance	commercial and service industry machinery manufacturing	communications, and audio and video equipment manufacturing	computer and peripheral equipment manufacturing	construction	construction, and mining and oil and gas field machinery manufacturing	crop production	cut and sew apparel manufacturing	cutlery and hand tool manufacturing	dairy product manufacturing	department stores and discount stores	don't know	drinking places, alcoholic beverages	drugs, sundries, and chemical and allied products, merchant wholesalers	drycleaning and laundry services	electric power generation, transmission, and distribution	electrical lighting and electrical equipment manufacturing, and other electrical component manufacturing, n.e.c.	electronic and precision equipment repair and maintenance	elementary and secondary schools	engines, turbines, and power transmission equipment manufacturing	executive offices and legislative bodies	fabric mills, except knitting mills	farm product raw materials, merchant wholesalers	fiber, yarn, and thread mills	fishing, hunting, and trapping	footwear and leather goods repair	footwear manufacturing	forestry, except logging	foundries	fuel dealers	funeral homes, and cemeteries and crematories	furniture and home furnishings stores	furniture and related product manufacturing	gasoline stations	glass and glass product manufacturing	groceries and related products, merchant wholesalers	grocery stores	hardware stores	hardware, plumbing and heating equipment, and supplies, merchant wholesalers	health and personal care, except drug, stores	hospitals	household appliance manufacturing	household appliance stores	iap	independent artists, performing arts, spectator sports, and related industries	individual and family services	industrial and miscellaneous chemicals	insurance carriers and related activities	investigation and security services	iron and steel mills and steel product manufacturing	jewelry, luggage, and leather goods stores	justice, public order, and safety activities	knitting fabric mills, and apparel knitting mills	labor unions	landscaping services	leather tanning and finishing and other allied products manufacturing	legal services	logging	lumber and other construction materials, merchant wholesalers	machinery manufacturing, n.e.c.	machinery, equipment, and supplies, merchant wholesalers	management, scientific, and technical consulting services	metal ore mining	metals and minerals, except petroleum, merchant wholesalers	metalworking machinery manufacturing	miscellaneous durable goods, merchant wholesalers	miscellaneous fabricated metal products manufacturing	miscellaneous general merchandise stores	miscellaneous manufacturing, n.e.c.	miscellaneous nonmetallic mineral product manufacturing	miscellaneous paper and pulp products	miscellaneous retail stores	miscellaneous wood products	motor vehicles and motor vehicle equipment manufacturing	museums, art galleries, historical sites, and similar institutions	national security and international affairs	natural gas distribution	navigational, measuring, electromedical, and control instruments manufacturing	newspaper publishers	no answer	non-depository credit and related activities	nonferrous metal (except aluminum) production and processing	nonmetallic mineral mining and quarrying	not specified food industries	not specified machinery manufacturing	not specified manufacturing industries	not specified metal industries	not specified retail trade	not specified type of mining	office supplies and stationery stores	offices of chiropractors	offices of dentists	offices of physicians	ordnance	other amusement, gambling, and recreation industries	other consumer goods rental	other direct selling establishments	other general government and support	other personal services	other professional, scientific, and technical services	paint, coating, and adhesive manufacturing	paper and paper products, merchant wholesalers	paperboard containers and boxes	periodical, book, and directory publishers	personal and household goods repair and maintenance	petroleum and petroleum products, merchant wholesalers	petroleum refining	pharmaceutical and medicine manufacturing	pharmacies and drug stores	plastics product manufacturing	postal service	pottery, ceramics, and plumbing fixture manufacturing	printing and related support activities	private households	pulp, paper, and paperboard mills	radio and television broadcasting and cable subscription programming	rail transportation	railroad rolling stock manufacturing	real estate	recyclable material, merchant wholesalers	religious organizations	restaurants and other food services	retail bakeries	rubber product, except tire, manufacturing	sawmills and wood preservation	scientific research and development services	securities, commodities, funds, trusts, and other financial investments	services incidental to transportation	services to buildings and dwellings	sewage treatment facilities	ship and boat building	soap, cleaning compound, and cosmetics manufacturing	specialized design services	specialty food stores	structural clay product manufacturing	structural metals, and boiler, tank, and shipping container manufacturing	taxi and limousine service	textile and fabric finishing and coating mills	textile product mills, except carpet and rug	tire manufacturing	tobacco manufacturing	traveler accommodation	truck transportation	u.s. air force	u.s. army	u.s. navy	uncodeable	warehousing and storage	waste management and remediation services	water transportation	water, steam, air-conditioning, and irrigation systems	wholesale electronic markets, agents and brokers	wired telecommunications carriers	active duty military personnel, all other	administration of economic programs and space research	advertising and related services	agricultural implement manufacturing	aluminum production and processing	apparel accessories and other apparel manufacturing	beer, wine, and liquor stores	business support services	cement, concrete, lime, and gypsum product manufacturing	civic, social, advocacy organizations, and grantmaking and giving services	electric and gas, and other combinations	electrical and electronic goods, merchant wholesalers	electronic component and product manufacturing, n.e.c.	fruit and vegetable preserving and specialty food manufacturing	lawn and garden equipment and supplies stores	machine shops; turned product; screw, nut, and bolt manufacturing	medical equipment and supplies manufacturing	miscellaneous nondurable goods, merchant wholesalers	motor vehicles, parts and supplies, merchant wholesalers	offices of optometrists	oil and gas extraction	other schools and instruction, and educational support services	radio, tv, and computer stores	residential care facilities, without nursing	resin, synthetic rubber and fibers, and filaments manufacturing	retail florists	savings institutions, including credit unions	seafood and other miscellaneous foods, n.e.c.	sewing, needlework, and piece goods stores	shoe stores	sporting and athletic goods, and doll, toy and game manufacturing	sporting goods, camera, and hobby and toy stores	sugar and confectionery products	support activities for agriculture and forestry	support activities for mining	veneer, plywood, and engineered wood products	veterinary services	bowling centers	business, professional, political, and similar organizations	farm supplies, merchant wholesalers	mail order houses	motion pictures and video industries	not specified utilities	not specified wholesale trade	other information services	pipeline transportation	public finance activities	u.s. armed forces/n.s.	u.s. marine corps	furniture and home furnishing, merchant wholesalers	metal forgings and stampings	other telecommunications services	prefabricated wood buildings and mobile homes	scenic and sightseeing transportation	used merchandise stores	vending machine operators	1976	automotive equipment rental and leasing	beauty salons	book stores and news dealers	coating, engraving, heat treating and allied activities	couriers and messengers	management of companies and enterprises	nursing care facilities	other transportation equipment manufacturing	professional and commercial equipment and supplies, merchant wholesalers	u.s. coast guard	business, technical, and trade schools and training	computer systems design and related services	employment services	commercial, industrial, and other intangible assets rental and leasing	miscellaneous petroleum and coal products	other administrative and other support services	other motor vehicle dealers	149	1720	music stores	data processing, hosting, and related services	offices of other health practitioners	other health care services	sound recording industries	travel arrangements and reservation services	child day care services	libraries and archives	recreational vehicle parks and camps, and rooming and boarding houses	outpatient care centers	vocational rehabilitation services	gift, novelty, and souvenir shops	nail salons and other personal care services	car washes	home health care services	community food and housing, and emergency services	2320	9720	software publishers	Total

1972	1	1	1	5	1	2	5	4	6	56	5	3	2	1	8	12	3	8	12	7	10	7	1	6	32	1	2	3	2	3	133	1	89	6	1	2	3	1	2	1	2	8	8	1	19	1	2	6	6	4	2	1	3	2	2	3	2	1	4	8	1	11	12	1	1	1	7	3	1	198	1	1	12	14	1	19	4	27	2	1	1	5	2	7	2	6	2	2	8	1	3	1	7	8	3	1	3	2	1	14	1	20	1	1	2	75	1	6	2	1	4	4	1	1	2	1	1	3	6	4	5	1	1	1	2	1	3	1	2	2	6	2	5	2	4	1	15	1	4	2	3	2	43	2	11	3	12	3	3	2	17	1	3	2	1	1	7	1	1	8	2	2	4	1	2	2	1	6	22	1	1	1	288	4	1	6	4	1	9	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1613

1973	4	-	-	-	-	2	6	1	3	98	3	1	5	-	9	6	3	8	6	1	8	4	-	4	42	5	6	2	2	-	136	3	163	4	1	5	1	-	4	2	3	9	3	-	8	2	3	4	1	-	1	2	2	-	5	3	2	4	7	5	5	15	21	5	1	1	5	5	6	151	3	-	7	12	1	27	3	19	2	1	3	-	2	3	4	10	2	-	8	1	3	-	8	9	3	1	1	2	1	20	1	12	-	-	6	28	2	3	2	-	2	6	3	3	2	-	1	2	4	2	1	-	1	1	1	3	1	-	1	1	2	2	3	2	7	1	9	-	11	-	13	-	64	2	7	4	13	6	4	3	21	-	1	4	1	1	8	1	-	2	-	4	8	2	1	2	1	3	21	2	1	2	111	2	1	2	5	2	4	1	4	1	4	4	1	1	1	6	3	1	3	1	1	2	2	2	2	1	1	1	1	1	1	2	1	1	1	1	1	2	1	2	1	4	2	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1504

1974	1	1	1	1	1	1	8	-	-	90	3	1	3	3	11	10	1	2	10	7	5	10	1	4	36	5	2	3	-	-	130	1	114	3	-	6	2	-	2	1	5	4	8	1	12	3	4	7	1	-	-	-	2	-	5	2	1	3	8	5	4	9	17	5	3	-	8	3	1	182	-	1	10	9	-	22	2	11	-	-	-	1	3	9	-	10	3	-	11	1	3	-	6	7	8	2	3	3	6	22	1	11	1	1	8	92	-	2	-	1	1	3	-	3	3	-	-	5	2	2	7	-	1	5	6	-	-	1	1	1	3	4	4	2	3	3	19	2	2	5	2	-	52	2	8	-	8	9	4	-	15	-	5	6	-	1	6	3	1	5	-	3	1	-	3	3	4	4	12	2	1	3	87	2	-	2	7	-	7	2	5	2	6	2	-	1	-	5	1	-	1	-	-	2	2	3	-	1	-	-	-	-	1	1	2	1	1	-	1	3	1	1	2	6	3	-	1	1	1	1	4	1	2	1	2	2	2	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1484

1975	2	4	1	2	-	2	15	-	4	99	6	2	3	1	14	13	1	5	10	7	5	6	-	4	34	8	2	2	2	2	131	4	121	3	2	11	6	-	1	2	5	10	4	-	10	1	6	6	2	1	5	2	2	3	2	4	-	4	10	10	6	4	23	2	3	-	4	5	1	153	1	-	7	13	1	25	2	19	2	1	1	3	1	3	1	11	2	1	15	-	2	-	2	3	5	1	2	-	3	25	1	23	3	-	3	18	2	6	2	-	-	1	-	2	3	1	1	2	7	-	4	-	1	-	1	-	-	1	-	1	5	3	3	2	6	-	16	3	5	3	3	2	43	3	12	-	8	6	-	1	20	1	-	10	1	-	10	1	-	5	3	4	2	2	2	2	1	6	26	3	1	3	119	2	4	3	4	-	2	-	3	2	2	4	-	-	-	5	2	1	1	1	1	3	1	2	1	3	1	1	-	1	1	2	-	3	3	-	2	1	-	6	-	2	-	1	-	-	-	-	-	2	-	1	-	-	1	1	1	2	1	1	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1490

1976	2	-	3	2	5	-	17	1	2	114	2	-	2	1	11	13	3	7	2	4	11	4	-	4	48	4	2	2	4	1	111	4	127	7	3	9	5	1	1	1	4	7	4	2	15	1	5	4	1	-	1	-	4	-	10	4	2	3	4	3	7	12	13	3	2	-	7	3	1	187	5	-	6	16	-	39	2	16	1	2	1	2	4	3	1	10	1	1	15	2	8	2	6	2	3	1	1	1	6	35	1	15	4	1	4	24	2	2	4	-	2	6	-	2	4	1	-	2	10	2	2	-	5	-	2	1	-	1	3	-	3	4	2	2	-	3	7	1	6	4	4	2	51	3	10	3	9	10	3	-	12	2	1	5	1	2	4	1	-	4	2	3	3	1	1	3	1	3	13	3	-	2	76	4	3	3	2	1	9	-	2	2	6	3	1	1	1	1	1	1	1	1	3	3	1	1	-	2	-	-	-	-	1	-	1	-	1	-	3	1	-	3	3	4	1	-	-	-	-	-	2	1	1	-	-	-	1	1	-	-	-	-	-	-	-	1	1	1	1	1	1	1	1	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1499

1977	2	1	-	2	-	4	10	2	3	58	-	-	4	3	4	13	2	7	3	8	4	4	2	5	56	7	5	3	4	2	123	5	108	-	1	7	3	-	3	1	9	12	5	-	12	1	2	3	2	-	2	-	2	-	10	2	-	3	7	5	2	10	16	3	3	-	2	3	1	193	3	-	3	9	-	28	2	20	-	-	2	5	8	6	2	13	1	1	17	-	7	-	4	2	5	3	1	3	3	29	1	26	-	2	6	25	3	1	2	-	-	1	-	5	2	-	-	3	4	1	1	-	3	-	1	-	1	1	-	2	5	1	6	-	2	2	18	3	1	5	7	1	45	2	5	3	8	4	2	2	18	2	4	9	2	1	10	1	-	3	4	8	-	1	-	2	1	8	22	-	4	2	186	1	-	5	6	1	5	-	1	2	6	3	-	-	-	4	2	2	-	1	1	3	3	1	-	1	-	2	-	-	-	-	1	1	1	-	1	3	-	1	2	3	-	-	-	-	-	-	1	-	-	-	1	-	1	-	-	-	-	-	-	-	-	-	1	-	1	-	-	-	-	1	1	-	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1530

1978	-	2	1	4	1	3	15	2	5	126	5	-	5	1	9	7	3	5	3	3	5	6	-	3	56	3	4	5	1	2	141	3	139	4	4	5	7	2	4	-	6	12	3	2	14	1	8	11	1	2	1	2	3	4	9	2	1	1	12	8	8	16	16	1	2	-	6	-	2	205	3	1	9	13	-	26	3	29	1	-	2	2	9	9	2	3	3	1	17	-	3	-	8	3	5	1	1	-	2	31	-	15	3	2	5	17	-	2	1	2	1	5	-	2	1	-	1	-	6	4	1	-	-	1	-	1	1	1	1	-	3	4	3	-	3	1	15	4	8	3	5	3	44	-	4	5	7	11	8	2	11	2	2	5	1	1	9	3	2	3	1	4	1	1	-	5	3	2	16	8	7	3	-	-	2	1	2	1	4	1	2	5	2	4	-	-	-	1	2	2	2	-	2	2	4	3	1	1	-	2	-	1	1	1	-	-	-	1	3	3	2	1	2	8	3	-	-	1	1	-	1	1	-	-	2	2	2	-	-	4	1	-	-	-	-	-	-	1	1	1	1	1	-	-	-	-	-	-	-	1	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1532

1980	2	4	-	4	2	2	6	2	3	96	4	3	4	3	11	12	3	5	9	4	6	3	1	4	55	10	3	4	2	-	111	3	120	3	1	2	3	-	-	1	4	10	1	-	19	3	5	7	-	1	1	1	3	-	4	2	2	5	8	8	7	10	23	2	1	-	7	1	2	230	-	-	9	10	-	28	3	21	2	2	3	3	7	7	2	8	-	-	16	1	2	1	4	5	6	2	1	1	3	16	2	15	4	1	9	9	-	2	4	3	-	8	-	5	1	-	2	-	8	-	3	-	-	2	2	1	2	2	3	1	6	3	2	2	2	2	9	1	9	4	7	-	44	4	9	1	6	8	2	2	15	3	3	3	2	1	4	2	2	1	-	4	3	-	2	3	-	7	18	5	4	3	46	1	-	-	1	1	10	1	3	1	4	3	-	1	-	-	2	2	-	2	1	3	4	2	-	1	-	-	1	-	-	2	-	-	-	-	1	-	2	2	2	8	2	1	-	-	1	-	-	-	1	-	-	5	1	1	-	3	1	-	-	1	-	-	-	-	1	-	-	1	-	1	-	1	-	-	-	-	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1468

1982	-	3	1	4	-	2	5	4	3	82	4	2	4	4	11	19	2	7	8	10	3	11	-	3	44	5	1	7	2	1	167	3	188	9	-	5	6	2	5	2	6	9	5	1	25	2	9	3	1	1	3	3	1	3	10	-	1	4	6	9	4	11	26	5	1	-	11	2	1	296	2	1	12	10	1	35	2	19	2	3	4	3	3	11	1	9	5	-	12	-	5	2	5	5	5	-	-	-	1	23	4	22	4	1	8	85	2	5	1	3	-	9	3	7	6	-	-	3	3	3	3	-	-	4	1	2	4	-	2	4	1	6	6	-	1	-	19	4	4	3	4	-	46	1	12	1	13	13	5	-	21	1	1	9	2	-	10	1	1	4	4	8	2	1	1	6	1	4	33	4	3	3	20	4	4	4	3	2	4	-	4	1	3	3	1	-	-	4	4	5	1	-	5	2	2	1	1	2	1	1	-	2	1	1	1	3	-	-	1	1	-	5	1	6	1	-	-	-	-	-	1	1	2	1	1	2	3	1	1	-	2	3	-	2	-	-	-	-	1	-	-	-	1	-	2	-	-	-	-	-	1	-	-	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1860

1983	2	1	2	8	-	2	8	1	3	71	11	1	6	-	16	16	4	10	10	1	7	6	-	7	18	11	4	2	6	2	170	3	106	7	1	10	10	-	1	1	6	15	4	3	23	2	6	10	2	1	3	1	5	-	8	4	1	7	15	7	2	6	20	2	5	-	6	2	3	207	-	1	12	12	-	56	1	21	6	4	5	2	9	6	1	3	3	1	18	3	7	-	5	2	2	-	2	2	1	27	-	21	1	6	4	18	1	-	6	1	-	2	-	2	3	1	-	6	8	3	-	-	-	3	-	1	2	2	-	1	5	-	4	2	1	3	19	-	4	3	6	1	35	1	12	2	15	8	3	2	13	4	7	7	3	1	2	-	-	5	2	11	2	-	1	4	3	8	26	3	4	-	23	5	3	5	5	2	8	-	3	1	8	5	-	-	1	4	-	1	-	2	1	3	5	-	-	3	3	4	1	3	-	3	-	-	2	-	1	1	1	4	1	5	-	1	1	-	1	2	4	-	1	-	-	-	-	1	1	3	-	1	-	-	-	-	-	1	-	1	-	1	1	1	-	1	-	-	-	-	1	-	1	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1599

1984	2	1	-	4	-	4	10	-	3	126	8	-	8	4	11	11	2	2	8	5	8	5	-	3	28	8	6	-	2	3	140	1	127	9	1	8	4	-	5	1	4	11	8	1	19	2	4	9	1	-	5	-	3	2	7	1	2	5	6	8	8	9	21	2	-	2	7	1	-	202	-	1	7	19	2	29	3	15	2	3	4	1	7	3	-	8	-	-	1	-	3	-	3	3	5	4	4	-	3	29	1	15	3	2	7	18	1	1	4	1	3	5	4	3	4	-	1	1	1	2	2	-	-	3	1	-	2	-	-	2	3	3	3	1	2	1	12	-	7	1	7	1	31	4	9	2	8	13	-	1	8	2	3	12	2	1	9	-	-	6	2	3	2	2	1	2	-	1	22	3	8	-	2	3	1	4	3	-	12	-	3	1	7	-	-	1	-	3	-	-	1	6	1	3	4	-	-	3	-	2	-	2	1	1	-	2	1	2	1	-	1	4	1	4	-	-	-	-	-	1	3	-	1	-	-	2	1	2	-	1	2	2	-	-	-	-	4	1	1	1	-	1	1	-	-	-	-	2	-	-	-	-	-	-	-	-	-	2	1	1	1	-	-	-	-	-	-	-	-	-	-	-	-	-	1473

1985	2	-	1	2	-	1	8	1	4	110	13	3	5	1	15	14	4	5	4	5	4	3	1	3	19	8	1	2	2	3	148	3	110	3	-	1	4	-	4	1	3	9	5	2	18	4	7	12	1	-	2	-	4	-	6	5	1	4	6	8	6	5	26	1	2	-	8	1	2	194	1	3	7	19	-	41	-	19	2	2	2	1	8	2	2	9	2	1	5	-	3	1	4	4	5	3	1	1	5	26	2	26	5	6	2	8	2	2	4	1	3	10	1	3	-	1	-	3	11	2	3	1	-	1	2	-	1	-	3	2	4	2	4	2	5	-	19	4	6	3	10	1	38	-	10	3	13	11	5	-	10	4	3	5	4	2	7	1	1	8	5	7	3	-	1	6	1	2	23	5	5	4	55	2	3	2	6	4	8	-	2	3	2	2	1	1	-	5	-	2	1	-	1	2	6	2	1	-	-	1	-	-	-	1	1	2	2	-	-	2	1	1	2	3	-	-	-	1	-	-	-	-	1	-	1	3	-	1	-	-	-	-	-	-	-	-	1	-	2	1	-	-	1	1	-	-	-	-	1	1	-	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1534

1986	1	1	1	2	-	2	8	1	5	124	10	3	2	3	15	10	1	4	4	9	9	8	1	-	10	10	6	2	2	1	124	4	105	11	-	2	9	-	2	3	7	10	7	2	11	-	6	8	2	2	-	5	1	1	9	6	3	5	12	5	7	10	15	3	1	-	10	2	1	222	6	-	10	12	-	39	1	28	3	6	2	1	5	3	4	11	4	1	6	-	2	-	3	3	7	5	4	2	2	28	1	13	5	1	5	21	1	3	5	1	1	10	2	3	2	-	-	3	3	2	2	-	1	1	1	-	-	1	5	2	-	3	3	1	3	-	12	-	6	5	5	3	32	-	13	1	10	9	2	-	7	3	1	3	1	-	5	1	3	-	2	5	3	1	2	2	1	4	23	2	2	3	10	4	3	4	2	-	6	-	4	4	3	4	-	2	-	5	1	1	3	1	2	1	6	-	-	1	-	-	1	1	1	1	1	1	-	-	2	-	1	4	3	2	-	-	-	-	-	-	-	-	3	-	1	-	3	-	-	1	-	1	-	1	2	-	-	-	-	1	1	-	1	-	1	1	-	-	-	-	1	-	1	-	-	2	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1470

1987	3	2	5	4	2	6	9	4	5	117	8	1	7	4	13	11	4	6	2	1	3	12	1	1	28	13	5	1	-	1	168	3	139	6	2	5	7	-	2	-	6	10	5	2	24	2	6	11	2	1	2	2	11	1	6	3	-	6	6	4	7	9	17	2	1	-	13	6	2	291	3	-	13	17	1	39	-	22	-	2	1	4	8	7	4	20	3	2	3	2	1	-	4	5	7	-	4	3	2	38	2	20	6	1	3	29	1	4	3	1	2	10	4	2	5	-	-	2	9	3	3	1	3	2	-	1	2	1	5	3	7	3	3	2	2	2	11	1	8	3	10	2	47	1	9	2	13	8	3	1	7	-	3	-	1	-	8	-	-	5	5	4	6	-	-	8	2	4	27	3	8	-	82	2	2	1	-	2	4	1	3	2	3	-	-	2	1	3	2	2	3	2	1	1	2	3	-	2	-	2	-	2	-	2	1	-	2	-	1	1	3	-	2	5	2	-	-	-	1	-	3	2	-	-	-	5	1	2	-	-	1	-	-	3	-	-	-	-	-	2	-	-	-	1	-	-	-	-	-	-	1	1	-	-	-	-	2	1	-	-	-	1	1	1	-	-	-	-	-	-	-	-	-	-	1819

1988	4	2	-	6	1	2	9	-	2	102	7	1	8	2	13	15	3	1	4	3	10	11	-	3	12	14	1	1	3	1	130	2	119	3	2	5	3	2	3	-	2	8	4	3	28	1	16	14	5	1	1	2	5	2	8	2	1	5	9	9	1	7	21	7	1	-	10	1	-	207	4	-	8	10	1	19	4	23	-	3	3	1	5	8	2	12	2	2	4	1	2	-	5	5	3	-	-	2	2	32	3	19	5	3	6	14	3	2	-	1	-	15	2	7	2	1	-	2	6	4	3	-	2	3	-	-	2	1	3	2	2	9	2	3	3	2	8	4	4	1	6	5	32	1	9	1	11	9	4	3	11	4	1	10	3	3	6	1	-	2	1	5	7	2	-	2	-	5	19	6	5	4	-	3	3	1	1	3	10	-	7	4	3	1	-	2	1	4	1	4	1	5	1	4	2	3	2	1	-	-	-	1	-	-	-	-	-	1	3	1	-	1	-	2	-	1	-	-	1	-	1	1	3	-	-	4	1	1	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	1	-	-	2	1	1	-	-	-	-	-	-	-	-	-	-	1	-	2	1	1	-	-	-	-	-	-	-	-	1481

1989	4	4	-	1	3	4	12	-	5	125	14	-	8	4	13	20	3	4	6	4	7	9	-	1	21	14	4	-	2	3	124	5	113	5	1	2	6	1	4	1	7	12	8	2	25	2	4	9	1	2	-	-	1	2	4	1	2	4	6	2	5	17	13	4	-	-	5	2	1	225	2	1	5	13	-	36	1	23	-	6	4	3	6	6	2	4	1	-	3	-	6	-	1	2	3	1	-	-	-	33	-	23	2	3	6	16	2	2	8	1	2	4	1	5	4	1	1	-	5	4	6	-	-	3	-	-	1	2	5	3	2	5	2	1	4	9	22	1	4	4	7	1	32	3	16	3	12	12	3	-	8	3	3	6	2	-	5	2	2	3	-	5	1	-	2	2	1	4	39	7	2	5	6	-	4	5	4	1	9	-	5	-	2	4	-	1	-	1	-	1	-	1	2	3	4	1	2	1	-	2	1	1	1	-	1	-	1	-	2	3	1	2	3	2	-	1	-	1	-	-	2	-	1	-	-	-	5	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1	2	-	2	1	-	2	-	-	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	1537

1990	2	1	4	3	1	4	15	-	4	88	9	3	3	3	10	15	2	6	3	5	10	7	-	4	9	8	3	1	2	1	122	4	89	5	2	2	6	1	1	2	11	10	7	1	18	2	10	4	4	-	-	2	4	-	4	5	2	4	11	6	6	15	15	2	5	1	6	3	-	217	4	1	9	14	-	37	1	11	2	2	5	-	-	3	1	5	3	1	5	1	5	1	2	1	2	-	1	-	-	39	-	13	6	3	7	16	2	-	-	1	1	6	2	6	4	-	1	3	4	-	-	-	1	5	-	1	1	1	2	-	3	3	7	1	1	3	15	1	4	-	7	2	32	1	4	4	8	9	2	-	6	1	3	6	3	-	5	-	-	6	-	4	1	-	-	3	-	2	16	1	6	3	-	3	2	5	4	1	15	1	5	5	5	5	2	-	-	4	1	1	1	4	5	5	3	-	1	1	1	1	1	1	-	2	1	1	-	-	2	1	-	3	2	2	-	1	-	-	-	-	-	2	1	-	2	1	3	1	2	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	3	1	-	-	-	-	1	-	-	-	-	1	-	2	-	-	1	-	-	2	-	-	-	-	-	-	-	-	-	-	1372

1991	2	2	5	6	-	2	10	-	4	110	9	2	8	4	12	16	3	3	6	4	5	9	2	2	23	7	4	2	-	2	132	2	105	7	-	4	7	3	2	-	5	10	2	-	24	3	6	5	2	1	1	-	3	-	7	2	3	7	6	6	4	8	18	3	1	-	12	1	1	220	3	1	7	16	-	29	-	7	2	4	4	1	2	4	2	10	-	-	-	1	5	1	1	5	5	1	-	1	3	33	1	26	3	4	6	44	2	1	1	3	1	5	-	6	7	1	1	4	8	2	3	1	1	1	1	-	1	4	-	2	5	2	3	4	3	-	21	3	7	2	9	1	30	2	6	-	12	10	4	2	8	3	4	8	3	1	6	1	3	3	-	8	2	-	-	3	3	4	27	7	9	1	1	4	4	1	5	3	8	-	7	3	2	-	-	-	1	3	1	2	1	4	2	3	4	2	4	-	-	2	-	1	-	1	2	-	2	-	1	2	-	1	1	5	1	1	-	1	-	-	2	1	4	-	-	1	1	-	1	-	-	1	-	-	1	-	-	1	1	-	-	-	1	-	-	1	3	3	-	-	1	-	-	-	-	2	-	-	-	-	-	-	-	-	-	-	1	-	-	-	-	-	-	-	1517

1993	1	4	4	11	1	6	6	4	3	106	4	-	7	4	19	20	2	9	3	1	6	7	-	4	10	10	7	5	3	5	130	4	126	7	1	4	5	-	2	3	4	20	2	3	23	-	12	9	2	2	2	1	4	1	10	1	-	5	12	8	5	6	16	3	5	2	11	1	-	197	1	2	12	16	-	15	3	28	2	3	8	3	10	11	2	9	3	-	2	-	5	1	3	2	7	1	3	1	1	36	2	25	2	4	4	44	-	3	6	2	-	11	2	5	-	1	-	4	6	3	3	1	2	-	-	1	3	3	4	1	2	2	4	4	2	2	16	-	10	1	8	2	29	2	5	2	12	16	4	2	9	1	1	12	2	4	8	4	2	4	3	5	3	3	-	3	-	5	24	5	6	1	4	3	2	3	6	1	13	-	5	5	4	3	-	3	1	2	-	-	2	3	1	4	5	-	1	1	1	2	-	-	-	3	2	1	2	1	1	2	1	-	3	2	1	1	-	1	-	-	-	2	2	-	-	5	1	1	-	-	2	1	-	-	2	-	-	1	-	-	1	-	-	-	3	1	-	-	2	-	-	-	1	-	-	-	-	-	-	-	1	-	-	-	1	-	-	-	-	-	-	-	-	-	1606

1994	10	2	7	19	2	5	25	3	9	170	10	2	15	2	20	25	6	18	10	8	20	11	3	8	14	20	8	8	1	10	222	13	176	12	2	13	9	-	4	1	3	22	10	1	58	3	20	20	-	6	6	2	5	2	15	8	4	4	15	14	11	20	29	3	7	2	19	8	6	508	6	-	29	33	-	52	4	47	4	6	10	1	16	13	1	17	4	1	5	2	13	-	4	4	12	1	1	2	3	47	9	33	3	3	10	41	2	2	10	2	2	25	1	9	7	-	4	6	7	1	5	-	4	1	-	-	1	3	3	3	2	6	12	4	9	3	28	1	14	3	15	2	82	1	14	1	21	25	5	5	17	4	4	18	8	2	18	1	2	5	3	12	6	1	5	6	-	9	47	16	7	10	5	5	5	2	5	3	22	-	11	2	11	4	-	-	1	5	-	3	4	7	5	10	9	6	1	6	-	4	1	1	-	7	1	2	1	-	3	1	-	6	6	16	1	2	-	2	1	-	4	3	2	-	2	3	5	2	3	2	1	-	1	2	1	-	4	-	-	3	5	2	-	1	4	2	1	2	1	1	2	-	-	-	-	-	1	-	-	-	1	-	-	-	-	-	1	-	-	-	-	-	-	-	2992

1996	10	6	2	14	2	9	21	2	5	152	10	1	14	4	25	27	3	13	8	6	12	12	1	7	18	17	5	6	4	3	261	3	142	10	1	7	16	-	4	2	7	34	17	2	59	-	10	16	2	5	6	1	6	2	11	5	4	13	12	11	6	16	36	4	4	-	20	4	2	502	3	1	17	26	3	40	3	46	5	6	8	4	12	10	1	13	4	1	6	2	5	-	7	5	9	8	3	2	2	61	3	53	9	5	5	63	1	3	7	3	5	16	5	11	2	1	-	4	9	4	8	-	2	7	2	4	2	1	5	4	2	9	7	1	4	7	25	2	11	3	9	5	59	1	18	2	25	18	4	2	21	2	5	21	3	2	6	2	4	4	1	10	6	-	3	4	1	5	54	13	14	9	2	5	7	6	6	4	16	-	10	9	9	9	1	1	1	13	6	1	3	5	6	5	8	1	2	6	3	7	1	5	3	4	-	1	2	1	-	3	3	6	2	14	1	-	-	1	4	-	2	2	5	-	1	3	6	3	1	-	-	-	-	-	2	-	-	4	-	2	2	2	3	-	1	1	-	5	3	-	1	1	1	-	-	2	-	-	2	-	-	-	-	1	2	-	-	-	-	-	-	-	-	-	2904

1998	7	12	4	9	2	4	20	-	3	135	14	1	11	8	24	32	8	12	4	7	12	11	1	6	19	31	11	1	5	11	263	4	147	4	2	15	9	-	1	3	6	19	10	8	50	4	19	16	3	2	5	5	5	2	13	9	3	6	17	15	2	10	31	3	1	-	19	7	3	503	7	4	14	31	5	36	2	40	3	7	11	4	15	10	4	17	6	1	8	1	6	-	8	3	10	3	3	4	2	51	2	48	7	-	5	66	-	2	5	4	3	22	2	12	1	1	-	5	15	3	2	-	1	5	3	1	1	-	2	6	4	4	16	5	7	3	26	1	11	-	8	1	53	1	16	2	19	39	1	-	14	5	8	10	3	2	13	2	2	7	2	12	2	1	2	4	1	8	34	16	17	11	3	3	6	3	6	5	18	-	10	1	8	4	1	2	1	6	2	1	-	6	2	6	11	7	1	6	-	3	2	3	1	1	1	7	2	1	4	2	3	3	4	12	1	2	-	2	2	-	4	2	4	-	2	5	5	1	1	2	-	-	-	1	1	-	-	1	-	1	1	1	2	-	3	-	1	4	-	-	2	-	2	-	-	-	-	2	-	2	-	-	2	-	1	2	-	-	-	-	-	-	-	-	2832

2000	9	7	3	10	-	14	19	5	7	171	13	2	11	4	28	33	9	6	5	12	10	10	3	5	21	20	6	3	4	5	248	6	139	14	1	13	8	-	4	5	9	28	4	2	43	-	18	20	3	2	7	-	5	-	5	7	2	10	10	16	6	12	24	8	4	-	18	9	2	518	5	3	20	28	6	54	4	44	1	2	6	1	11	6	6	12	9	1	1	-	7	2	3	-	3	2	2	-	-	61	4	51	9	1	8	60	4	3	9	1	2	20	1	4	3	3	-	4	9	7	8	-	1	3	1	1	2	2	4	3	9	7	11	3	5	1	23	-	13	3	18	4	48	2	19	2	23	19	4	4	11	5	11	23	2	4	11	2	-	6	1	6	5	2	5	5	2	7	45	11	7	11	-	11	5	4	4	1	11	1	8	2	4	7	-	-	1	8	1	3	5	9	3	6	9	2	-	4	-	6	-	3	2	5	2	1	1	1	2	1	1	4	5	13	2	-	-	1	-	-	3	4	2	-	1	5	2	4	1	1	1	-	-	-	-	-	-	2	-	-	4	1	-	1	2	-	-	6	1	-	2	-	1	-	-	-	-	-	1	-	-	-	-	-	-	-	-	-	-	-	-	-	-	-	2817

2002	10	4	9	10	3	5	20	3	4	131	9	1	14	4	15	44	8	9	5	11	17	8	1	4	16	25	5	4	4	10	249	-	127	5	1	11	8	1	5	2	3	28	7	2	55	5	17	29	4	2	4	1	4	-	7	3	2	6	16	16	5	11	25	1	3	1	21	6	-	526	4	3	25	19	2	50	-	53	2	8	7	4	10	11	2	9	7	3	4	2	9	1	3	2	7	4	-	4	8	53	5	41	6	-	10	63	1	2	4	1	3	16	2	5	3	4	1	5	8	1	5	1	2	1	2	2	-	-	4	2	6	8	13	3	7	7	31	-	14	2	21	2	44	1	15	2	16	28	4	2	11	3	9	13	7	3	11	1	-	5	1	4	7	-	2	6	4	9	57	5	16	13	-	11	7	1	5	1	24	1	8	3	1	12	-	3	1	6	1	-	-	4	4	6	6	1	-	3	2	5	2	5	-	6	-	-	1	1	2	2	2	4	-	8	-	1	1	1	-	-	1	2	1	1	1	5	4	2	-	3	1	-	-	2	-	-	1	-	1	1	2	1	2	1	2	2	-	5	1	1	1	3	-	-	-	-	-	1	1	-	1	-	-	-	1	1	-	-	-	-	-	-	-	-	2765

2004	9	3	10	7	2	9	21	3	4	110	12	3	16	5	31	32	5	15	3	7	15	10	-	5	23	37	5	5	-	8	268	7	131	6	2	7	8	-	4	2	5	21	9	5	78	4	18	10	2	3	7	2	4	1	10	6	5	6	16	13	8	5	24	7	-	1	25	4	7	514	1	2	22	26	3	39	2	38	-	4	7	-	13	18	1	14	6	1	2	2	8	-	10	4	7	1	3	4	4	51	2	39	6	2	10	86	5	-	3	4	3	33	1	6	3	1	-	3	11	2	6	-	3	8	2	1	2	-	4	5	3	6	7	3	5	3	24	4	16	1	19	1	30	2	16	-	16	22	6	4	16	7	3	15	5	1	10	2	1	6	1	13	5	-	-	3	4	3	46	12	14	15	-	10	12	1	6	2	19	-	13	7	4	8	-	1	2	5	4	-	-	6	2	9	10	3	2	-	1	3	2	2	1	6	-	1	-	1	-	1	1	2	2	8	1	1	1	-	2	1	2	3	1	-	3	6	6	6	1	1	3	2	-	-	1	-	1	1	2	-	2	1	1	1	4	-	-	3	-	-	2	1	-	-	-	2	-	2	1	2	2	-	-	2	-	-	-	-	-	-	-	-	-	-	2812

2006	5	4	11	9	2	4	9	-	3	109	11	1	18	11	25	40	6	14	5	10	16	16	-	8	23	26	7	3	2	7	306	9	164	16	4	8	5	-	11	-	7	35	13	3	67	1	23	10	3	3	7	1	3	2	6	5	2	12	29	14	6	12	28	4	2	2	32	2	5	2049	3	-	20	35	4	40	4	49	3	6	11	1	16	25	3	16	3	1	6	1	9	-	4	2	6	1	5	3	6	53	6	36	4	3	8	55	4	2	3	4	1	41	-	7	1	-	-	5	14	2	9	1	1	-	-	-	1	3	1	3	7	9	12	5	7	6	33	2	7	6	17	2	37	3	13	-	23	27	3	2	11	9	7	19	4	1	13	2	2	8	4	6	3	-	1	9	1	4	59	16	11	5	-	8	9	7	6	2	24	1	10	3	1	8	-	2	1	7	2	2	4	4	1	4	8	3	-	5	4	1	-	2	-	7	-	6	2	-	-	2	-	6	2	4	3	2	2	-	2	-	2	1	3	-	1	8	8	-	1	-	-	-	-	-	-	-	-	1	-	-	1	-	4	1	3	2	-	6	1	-	-	1	4	-	-	2	-	1	4	-	1	1	-	-	1	-	1	1	-	-	-	-	-	-	4510

2008	6	6	2	7	3	9	18	2	4	71	8	1	17	5	21	25	6	9	4	6	8	11	-	4	14	13	4	5	2	7	184	4	91	5	1	6	6	-	-	2	3	13	5	1	35	-	19	6	1	-	7	-	5	4	9	3	4	5	8	12	10	8	23	-	2	2	14	3	2	354	3	2	13	28	-	27	-	38	3	14	11	-	9	9	3	10	4	2	5	1	6	1	5	1	7	3	2	2	1	42	3	30	10	2	4	44	5	3	2	1	2	12	2	1	1	2	1	2	6	2	2	-	4	2	1	-	1	1	3	3	2	4	6	1	2	5	16	1	9	2	11	1	28	3	17	1	12	27	4	1	14	4	8	11	4	3	1	2	4	3	-	6	3	1	3	2	2	6	34	12	10	8	-	7	7	4	6	-	22	2	11	-	5	2	-	1	-	4	-	4	2	4	2	4	9	1	-	2	1	3	-	-	1	3	-	2	4	-	-	2	2	2	2	6	2	1	-	-	-	-	3	1	-	-	-	4	4	1	-	2	1	2	1	-	-	-	-	1	-	1	1	2	1	-	1	-	-	1	-	1	2	-	1	-	-	-	-	-	2	1	1	1	-	-	1	-	-	1	1	1	-	-	-	-	2023

2010	2	4	4	10	2	9	15	-	3	81	9	3	11	4	30	32	6	11	5	4	12	10	-	2	9	34	1	4	2	9	185	2	79	10	-	5	6	-	1	3	6	24	5	6	35	3	8	11	2	1	8	1	1	-	6	3	1	8	17	8	6	14	15	4	2	2	18	1	1	354	4	1	12	15	3	23	2	25	2	1	9	1	6	8	2	15	1	2	4	1	5	1	6	1	4	3	3	1	4	43	3	33	4	1	9	38	6	1	3	3	2	11	-	5	1	1	-	7	12	-	8	1	2	3	-	3	-	-	4	5	5	4	6	3	3	5	13	-	9	2	15	1	32	2	21	-	8	27	1	2	18	4	5	16	10	-	6	4	3	1	2	5	7	-	1	3	2	7	39	7	6	3	-	5	6	3	3	-	14	-	12	1	1	4	-	2	-	8	2	-	1	5	1	1	1	1	2	1	-	3	-	2	3	3	1	3	5	-	-	2	2	3	2	9	4	1	-	-	2	-	3	2	3	-	1	8	2	1	1	1	2	-	-	-	1	-	1	2	1	-	1	-	2	-	5	-	1	3	1	-	1	-	-	-	-	1	-	-	1	-	3	-	1	-	-	-	-	1	-	2	1	-	-	-	2044

2012	4	4	4	4	1	4	22	3	2	48	12	1	10	6	16	25	4	11	3	12	5	2	-	4	7	29	1	5	1	3	200	3	119	4	1	3	8	-	3	2	3	13	7	4	43	4	22	3	-	5	-	-	4	-	6	2	3	7	5	7	5	12	19	1	2	-	22	2	4	341	5	3	16	22	4	27	2	35	1	4	6	2	4	2	3	12	1	5	4	1	7	1	7	1	2	1	1	1	3	41	5	6	7	2	7	42	1	-	3	3	2	31	5	4	2	-	-	3	10	1	6	1	1	1	-	2	-	2	5	-	3	3	6	1	3	6	19	1	7	3	9	4	28	2	13	2	16	17	3	2	4	2	3	9	3	1	2	3	1	-	1	4	3	-	1	3	-	4	29	-	-	-	-	7	3	3	2	1	5	-	9	1	3	5	-	-	-	4	-	3	3	6	1	3	7	3	-	1	-	3	1	3	2	2	-	-	1	-	4	2	1	2	3	10	-	1	1	1	-	-	-	-	-	-	1	1	75	-	-	2	1	-	1	2	-	-	1	2	-	3	4	1	4	-	-	-	1	7	3	-	2	3	-	-	-	-	-	1	2	-	1	-	1	-	5	-	-	3	1	-	-	-	-	-	1974

2014	5	3	3	6	2	8	18	5	7	75	10	1	13	6	28	38	11	15	3	9	13	9	1	2	15	27	10	4	7	6	242	4	125	7	1	8	7	-	2	4	4	22	12	4	42	2	14	4	2	3	2	-	1	2	3	4	1	7	13	9	5	10	27	2	3	2	18	5	4	459	11	4	14	27	7	31	3	49	-	6	14	4	13	7	2	9	5	7	3	1	12	-	8	1	5	3	1	1	2	66	5	28	5	1	3	67	4	4	3	3	1	14	1	15	3	-	-	3	14	-	5	-	2	2	-	1	-	1	3	2	5	3	11	2	2	4	19	4	20	3	9	1	26	1	15	-	20	31	4	1	7	4	5	14	8	3	3	3	3	7	1	15	8	1	2	4	1	4	53	15	16	13	-	5	8	4	3	1	14	3	10	8	5	-	1	2	2	1	-	7	2	5	5	7	5	1	3	2	-	4	-	5	-	3	-	4	4	-	2	2	3	2	7	9	1	-	1	-	2	-	3	1	2	-	-	4	5	3	1	4	-	-	1	1	1	-	1	1	-	1	-	1	3	2	3	1	1	4	3	1	-	3	1	-	-	1	-	1	-	-	-	-	-	2	4	-	-	-	2	-	-	-	-	-	2538

2016	9	8	3	11	-	10	9	1	8	93	10	-	13	8	32	37	11	19	8	10	13	8	1	4	4	27	11	4	4	4	266	7	99	7	2	11	5	-	3	4	4	36	5	8	66	2	27	3	5	-	2	-	3	1	5	1	3	6	15	12	11	13	22	2	2	2	32	4	2	548	6	3	20	21	6	37	-	51	-	11	15	-	11	3	5	16	5	7	6	1	6	1	9	4	7	-	-	2	3	68	3	53	8	2	4	71	3	4	7	2	1	37	6	14	4	-	-	3	10	1	5	-	3	4	1	2	1	2	4	1	8	5	10	6	9	3	18	6	21	-	7	4	37	2	21	2	17	33	6	2	8	6	10	22	6	2	13	2	2	5	1	13	6	-	-	5	-	9	56	15	19	5	-	4	8	5	10	1	19	4	8	5	4	2	-	-	1	5	3	3	-	7	2	3	4	5	2	3	-	1	3	3	2	1	-	1	3	-	2	3	2	3	8	11	-	2	-	1	3	1	3	1	7	-	1	4	3	1	-	2	3	-	2	-	-	-	2	1	1	2	5	-	-	2	4	1	-	17	1	3	1	4	-	-	-	-	-	1	5	-	-	-	1	-	1	-	-	-	2	-	-	-	-	-	2867

2018	9	7	7	9	4	8	16	2	4	44	7	1	13	6	30	39	4	12	3	5	7	12	1	3	5	20	8	3	2	5	208	3	115	1	2	6	8	-	2	3	9	14	7	5	45	1	4	4	1	1	1	1	1	1	5	3	4	8	6	14	4	14	18	4	2	2	20	4	6	464	2	1	12	26	2	32	2	46	2	9	14	-	8	7	5	14	3	10	3	-	8	3	4	1	5	1	2	1	1	25	3	16	3	3	5	38	3	2	3	-	-	12	2	20	2	-	1	6	10	3	9	-	1	2	2	-	1	1	2	3	6	6	4	6	4	3	26	-	4	-	4	4	31	2	25	2	10	23	4	1	13	6	9	22	3	7	6	2	1	3	1	7	3	2	1	6	1	10	51	20	28	16	-	12	9	2	3	5	16	-	9	5	5	3	1	1	-	3	1	3	5	6	3	4	7	7	3	4	1	1	-	6	2	3	2	-	-	-	1	2	3	7	4	16	-	3	-	1	1	-	-	-	2	-	2	7	10	3	-	2	-	-	1	1	-	-	1	1	-	1	2	3	2	-	6	1	1	7	-	4	1	2	-	-	-	-	-	2	5	-	1	-	-	-	4	-	-	1	1	-	-	1	1	-	2348

2021	13	6	12	18	4	10	33	2	4	86	12	4	30	10	41	55	9	36	12	9	11	18	-	8	11	59	16	8	4	8	436	15	96	7	4	10	10	-	15	6	15	35	11	10	83	6	25	13	3	2	6	1	7	1	7	4	5	7	27	13	7	27	46	3	2	5	47	5	4	563	12	6	19	45	4	39	7	75	2	5	22	2	23	11	5	16	9	7	7	4	12	3	13	5	15	5	5	7	4	101	4	38	12	6	20	144	5	3	8	6	11	17	2	8	2	6	1	13	14	4	10	-	3	1	2	6	4	-	7	6	7	8	16	5	6	12	40	3	15	2	14	4	27	5	30	4	28	51	2	5	12	13	11	28	19	4	12	3	6	12	2	18	13	1	1	7	4	12	81	-	-	-	-	8	7	6	8	3	24	81	15	14	8	2	-	2	4	9	6	6	8	12	3	15	15	13	5	3	1	5	7	9	2	5	2	1	6	1	4	1	3	5	4	29	2	3	-	-	1	-	1	1	2	-	2	8	-	-	-	2	5	3	1	2	3	-	2	2	2	3	5	5	4	-	9	-	1	26	5	1	-	3	3	-	-	4	1	1	11	-	2	1	1	1	4	-	1	-	1	-	-	-	-	1	4032

2022	12	4	13	13	1	7	18	2	3	64	22	2	31	11	25	63	7	32	6	8	13	19	1	3	4	41	10	6	4	3	394	11	106	7	-	11	11	-	-	1	6	25	15	8	87	-	29	5	-	-	1	-	2	4	4	2	4	10	17	9	4	21	32	5	2	3	44	4	5	623	8	8	22	34	8	42	3	78	1	6	28	1	18	7	8	15	4	5	7	3	9	2	7	4	4	3	2	4	4	64	8	42	5	3	8	157	3	-	6	3	2	15	5	7	4	2	1	4	17	4	9	1	1	3	1	2	2	1	4	3	11	3	11	4	10	5	32	3	18	1	14	4	27	2	22	2	19	48	3	5	9	9	13	26	10	2	9	3	-	6	1	4	4	3	1	9	3	13	82	-	-	-	-	15	16	6	7	2	18	77	11	5	7	5	-	2	3	3	8	2	6	12	6	6	4	3	3	4	1	4	1	6	1	8	-	1	6	1	2	1	2	2	3	15	1	2	1	1	-	-	2	2	1	1	2	8	-	-	2	3	4	-	2	2	1	-	-	-	1	1	6	5	5	2	5	-	1	29	1	-	-	1	1	-	-	1	-	1	10	-	3	1	-	1	1	-	-	-	2	2	1	-	-	-	3544

Total	157	114	124	227	48	166	466	60	137	3438	294	50	331	140	616	800	152	336	197	213	310	299	23	138	784	577	176	116	87	132	6503	149	4164	217	47	229	221	14	107	63	189	565	230	95	1193	68	394	312	68	54	101	39	119	43	238	117	73	195	378	307	191	395	748	107	76	31	514	113	79	12803	122	55	440	670	65	1158	75	1092	60	140	234	66	284	258	86	376	111	68	237	36	197	25	179	114	192	66	65	62	92	1363	89	907	156	76	217	1636	74	78	131	63	63	443	60	200	94	30	19	121	277	78	149	10	53	79	38	38	45	40	95	79	147	146	217	85	137	108	655	61	299	83	313	69	1359	64	446	62	471	610	115	61	423	120	160	389	130	57	263	58	49	155	58	232	135	29	47	139	50	192	1196	228	242	162	1126	165	157	114	150	57	413	177	223	110	148	125	10	35	25	147	59	66	65	132	77	138	173	83	40	75	22	76	26	72	29	87	24	46	57	13	53	54	43	98	85	253	35	30	9	17	26	6	57	39	58	5	30	111	162	41	19	41	32	17	13	21	17	1	21	25	17	27	45	30	40	17	64	20	12	133	28	15	28	25	18	1	1	19	5	19	48	6	19	6	7	12	27	4	4	7	10	5	2	1	1	1	72390

Values

170 crop production

180 animal production

190 forestry, except logging

270 logging

280 fishing, hunting, and trapping

290 support activities for agriculture and forestry

370 oil and gas extraction

380 coal mining

390 metal ore mining

470 nonmetallic mineral mining and quarrying

480 not specified type of mining

490 support activities for mining

570 electric power generation, transmission, and distribution

580 natural gas distribution

590 electric and gas, and other combinations

670 water, steam, air-conditioning, and irrigation systems

680 sewage treatment facilities

690 not specified utilities

770 construction

1070 animal food, grain, and oilseed milling

1080 sugar and confectionery products

1090 fruit and vegetable preserving and specialty food manufacturing

1170 dairy product manufacturing

1180 animal slaughtering and processing

1190 retail bakeries

1270 bakeries, except retail

1280 seafood and other miscellaneous foods, n.e.c.

1290 not specified food industries

1370 beverage manufacturing

1390 tobacco manufacturing

1470 fiber, yarn, and thread mills

1480 fabric mills, except knitting mills

1490 textile and fabric finishing and coating mills

1570 carpet and rug mills

1590 textile product mills, except carpet and rug

1670 knitting fabric mills, and apparel knitting mills

1680 cut and sew apparel manufacturing

1690 apparel accessories and other apparel manufacturing

1770 footwear manufacturing

1790 leather tanning and finishing and other allied products manufacturing

1870 pulp, paper, and paperboard mills

1880 paperboard containers and boxes

1890 miscellaneous paper and pulp products

1990 printing and related support activities

2070 petroleum refining

2090 miscellaneous petroleum and coal products

2170 resin, synthetic rubber and fibers, and filaments manufacturing

2180 agricultural chemical manufacturing

2190 pharmaceutical and medicine manufacturing

2270 paint, coating, and adhesive manufacturing

2280 soap, cleaning compound, and cosmetics manufacturing

2290 industrial and miscellaneous chemicals

2370 plastics product manufacturing

2380 tire manufacturing

2390 rubber product, except tire, manufacturing

2470 pottery, ceramics, and plumbing fixture manufacturing

2480 structural clay product manufacturing

2490 glass and glass product manufacturing

2570 cement, concrete, lime, and gypsum product manufacturing

2590 miscellaneous nonmetallic mineral product manufacturing

2670 iron and steel mills and steel product manufacturing

2680 aluminum production and processing

2690 nonferrous metal (except aluminum) production and processing

2770 foundries

2780 metal forgings and stampings

2790 cutlery and hand tool manufacturing

2870 structural metals, and boiler, tank, and shipping container manufacturing

2880 machine shops; turned product; screw, nut, and bolt manufacturing

2890 coating, engraving, heat treating and allied activities

2970 ordnance

2980 miscellaneous fabricated metal products manufacturing

2990 not specified metal industries

3070 agricultural implement manufacturing

3080 construction, and mining and oil and gas field machinery manufacturing

3090 commercial and service industry machinery manufacturing

3170 metalworking machinery manufacturing

3180 engines, turbines, and power transmission equipment manufacturing

3190 machinery manufacturing, n.e.c.

3290 not specified machinery manufacturing

3360 computer and peripheral equipment manufacturing

3370 communications, and audio and video equipment manufacturing

3380 navigational, measuring, electromedical, and control instruments manufacturing

3390 electronic component and product manufacturing, n.e.c.

3470 household appliance manufacturing

3490 electrical lighting and electrical equipment manufacturing, and other electrical component manufacturing, n.e.c.

3570 motor vehicles and motor vehicle equipment manufacturing

3580 aircraft and parts manufacturing

3590 aerospace product and parts manufacturing

3670 railroad rolling stock manufacturing

3680 ship and boat building

3690 other transportation equipment manufacturing

3770 sawmills and wood preservation

3780 veneer, plywood, and engineered wood products

3790 prefabricated wood buildings and mobile homes

3870 miscellaneous wood products

3890 furniture and related product manufacturing

3960 medical equipment and supplies manufacturing

3970 sporting and athletic goods, and doll, toy and game manufacturing

3980 miscellaneous manufacturing, n.e.c.

3990 not specified manufacturing industries

4070 motor vehicles, parts and supplies, merchant wholesalers

4080 furniture and home furnishing, merchant wholesalers

4090 lumber and other construction materials, merchant wholesalers

4170 professional and commercial equipment and supplies, merchant wholesalers

4180 metals and minerals, except petroleum, merchant wholesalers

4190 electrical and electronic goods, merchant wholesalers

4260 hardware, plumbing and heating equipment, and supplies, merchant wholesalers

4270 machinery, equipment, and supplies, merchant wholesalers

4280 recyclable material, merchant wholesalers

4290 miscellaneous durable goods, merchant wholesalers

4370 paper and paper products, merchant wholesalers

4380 drugs, sundries, and chemical and allied products, merchant wholesalers

4390 apparel, fabrics, and notions, merchant wholesalers

4470 groceries and related products, merchant wholesalers

4480 farm product raw materials, merchant wholesalers

4490 petroleum and petroleum products, merchant wholesalers

4560 alcoholic beverages, merchant wholesalers

4570 farm supplies, merchant wholesalers

4580 miscellaneous nondurable goods, merchant wholesalers

4585 wholesale electronic markets, agents and brokers

4590 not specified wholesale trade

4670 automobile dealers

4680 other motor vehicle dealers

4690 auto parts, accessories, and tire stores

4770 furniture and home furnishings stores

4780 household appliance stores

4790 radio, tv, and computer stores

4870 building material and supplies dealers

4880 hardware stores

4890 lawn and garden equipment and supplies stores

4970 grocery stores

4980 specialty food stores

4990 beer, wine, and liquor stores

5070 pharmacies and drug stores

5080 health and personal care, except drug, stores

5090 gasoline stations

5170 clothing stores

5180 shoe stores

5190 jewelry, luggage, and leather goods stores

5270 sporting goods, camera, and hobby and toy stores

5280 sewing, needlework, and piece goods stores

5290 music stores

5370 book stores and news dealers

5380 department stores and discount stores

5390 miscellaneous general merchandise stores

5470 retail florists

5480 office supplies and stationery stores

5490 used merchandise stores

5570 gift, novelty, and souvenir shops

5580 miscellaneous retail stores

5590 electronic shopping

5591 electronic auctions

5592 mail order houses

5670 vending machine operators

5680 fuel dealers

5690 other direct selling establishments

5790 not specified retail trade

6070 air transportation

6080 rail transportation

6090 water transportation

6170 truck transportation

6180 bus service and urban transit

6190 taxi and limousine service

6270 pipeline transportation

6280 scenic and sightseeing transportation

6290 services incidental to transportation

6370 postal service

6380 couriers and messengers

6390 warehousing and storage

6470 newspaper publishers

6480 periodical, book, and directory publishers

6490 software publishers

6570 motion pictures and video industries

6590 sound recording industries

6670 radio and television broadcasting and cable subscription programming

6672 internet publishing and broadcasting and web search portals

6680 wired telecommunications carriers

6690 other telecommunications services

6695 data processing, hosting, and related services

6770 libraries and archives

6780 other information services

6870 banking and related activities

6880 savings institutions, including credit unions

6890 non-depository credit and related activities

6970 securities, commodities, funds, trusts, and other financial investments

6990 insurance carriers and related activities

7070 real estate

7080 automotive equipment rental and leasing

7170 video tape and disk rental

7180 other consumer goods rental

7190 commercial, industrial, and other intangible assets rental and leasing

7270 legal services

7280 accounting, tax preparation, bookkeeping, and payroll services

7290 architectural, engineering, and related services

7370 specialized design services

7380 computer systems design and related services

7390 management, scientific, and technical consulting services

7460 scientific research and development services

7470 advertising and related services

7480 veterinary services

7490 other professional, scientific, and technical services

7570 management of companies and enterprises

7580 employment services

7590 business support services

7670 travel arrangements and reservation services

7680 investigation and security services

7690 services to buildings and dwellings

7770 landscaping services

7780 other administrative and other support services

7790 waste management and remediation services

7860 elementary and secondary schools

7870 colleges and universities, including junior colleges

7880 business, technical, and trade schools and training

7890 other schools and instruction, and educational support services

7970 offices of physicians

7980 offices of dentists

7990 offices of chiropractors

8070 offices of optometrists

8080 offices of other health practitioners

8090 outpatient care centers

8170 home health care services

8180 other health care services

8190 hospitals

8270 nursing care facilities

8290 residential care facilities, without nursing

8370 individual and family services

8380 community food and housing, and emergency services

8390 vocational rehabilitation services

8470 child day care services

8560 independent artists, performing arts, spectator sports, and related industries

8570 museums, art galleries, historical sites, and similar institutions

8580 bowling centers

8590 other amusement, gambling, and recreation industries

8660 traveler accommodation

8670 recreational vehicle parks and camps, and rooming and boarding houses

8680 restaurants and other food services

8690 drinking places, alcoholic beverages

8770 automotive repair and maintenance

8780 car washes

8790 electronic and precision equipment repair and maintenance

8870 commercial and industrial machinery and equipment repair and maintenance

8880 personal and household goods repair and maintenance

8890 footwear and leather goods repair

8970 barber shops

8980 beauty salons

8990 nail salons and other personal care services

9070 drycleaning and laundry services

9080 funeral homes, and cemeteries and crematories

9090 other personal services

9160 religious organizations

9170 civic, social, advocacy organizations, and grantmaking and giving services

9180 labor unions

9190 business, professional, political, and similar organizations

9290 private households

9370 executive offices and legislative bodies

9380 public finance activities

9390 other general government and support

9470 justice, public order, and safety activities

9480 administration of human resource programs

9490 administration of environmental quality and housing programs

9570 administration of economic programs and space research

9590 national security and international affairs

9670 u.s. army

9680 u.s. air force

9690 u.s. navy

9770 u.s. marine corps

9780 u.s. coast guard

9790 u.s. armed forces/n.s.

9870 active duty military personnel, all other

9890 uncodeable

9997 uncodeable

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(q) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook