R's father's occupation, 2010 census & 2008 isco code

paisco08

Source

General Social Survey https://gss.norc.org

Details

Question None

Values

110 commissioned armed forces officers

210 non-commissioned armed forces officers

310 armed forces occupations, other ranks

1000 managers

1100 chief executives, senior officials and legislators

1110 legislators and senior officials

1111 legislators

1112 senior government officials

1113 traditional chiefs and heads of village

1114 senior officials of special-interest organizations

1120 managing directors and chief executives

1200 administrative and commercial managers

1210 business services and administration managers

1211 finance managers

1212 human resource managers

1213 policy and planning managers

1219 business services and administration managers not elsewhere classified

1220 sales, marketing and development managers

1221 sales and marketing managers

1222 advertising and public relations managers

1223 research and development managers

1300 production and specialised services managers

1310 production managers in agriculture, forestry and fisheries

1311 agricultural and forestry production managers

1312 aquaculture and fisheries production managers

1320 manufacturing, mining, construction, and distribution managers

1321 manufacturing managers

1322 mining managers

1323 construction managers

1324 supply, distribution and related managers

1330 information and communications technology service managers

1340 professional services managers

1341 child care services managers

1342 health services managers

1343 aged care services managers

1344 social welfare managers

1345 education managers

1346 financial and insurance services branch managers

1349 professional services managers not elsewhere classified

1400 hospitality, retail and other services managers

1410 hotel and restaurant managers

1411 hotel managers

1412 restaurant managers

1420 retail and wholesale trade managers

1430 other services managers

1431 sports, recreation and cultural centre managers

1439 services managers not elsewhere classified

2000 professionals

2100 science and engineering professionals

2110 physical and earth science professionals

2111 physicists and astronomers

2112 meteorologists

2113 chemists

2114 geologists and geophysicists

2120 mathematicians, actuaries and statisticians

2130 life science professionals

2131 biologists, botanists, zoologists and related professionals

2132 farming, forestry and fisheries advisers

2133 environmental protection professionals

2140 engineering professionals (excluding electrotechnology)

2141 industrial and production engineers

2142 civil engineers

2143 environmental engineers

2144 mechanical engineers

2145 chemical engineers

2146 mining engineers, metallurgists and related professionals

2149 engineering professionals not elsewhere classified

2150 electrotechnology engineers

2151 electrical engineers

2152 electronics engineers

2153 telecommunications engineers

2160 architects, planners, surveyors and designers

2161 building architects

2162 landscape architects

2163 product and garment designers

2164 town and traffic planners

2165 cartographers and surveyors

2166 graphic and multimedia designers

2200 health professionals

2210 medical doctors

2211 generalist medical practitioners

2212 specialist medical practitioners

2220 nursing and midwifery professionals

2221 nursing professionals

2222 midwifery professionals

2230 traditional and complementary medicine professionals

2240 paramedical practitioners

2250 veterinarians

2260 other health professionals

2261 dentists

2262 pharmacists

2263 environmental and occupational health and hygiene professionals

2264 physiotherapists

2265 dieticians and nutritionists

2266 audiologists and speech therapists

2267 optometrists and ophthalmic opticians

2269 health professionals not elsewhere classified

2300 teaching professionals

2310 university and higher education teachers

2320 vocational education teachers

2330 secondary education teachers

2340 primary school and early childhood teachers

2341 primary school teachers

2342 early childhood educators

2350 other teaching professionals

2351 education methods specialists

2352 special needs teachers

2353 other language teachers

2354 other music teachers

2355 other arts teachers

2356 information technology trainers

2359 teaching professionals not elsewhere classified

2400 business and administration professionals

2410 finance professionals

2411 accountants

2412 financial and investment advisers

2413 financial analysts

2420 administration professionals

2421 management and organization analysts

2422 policy administration professionals

2423 personnel and careers professionals

2424 training and staff development professionals

2430 sales, marketing and public relations professionals

2431 advertising and marketing professionals

2432 public relations professionals

2433 technical and medical sales professionals (excluding ict)

2434 information and communications technology sales professionals

2500 information and communications technology professionals

2510 software and applications developers and analysts

2511 systems analysts

2512 software developers

2513 web and multimedia developers

2514 applications programmers

2519 software and applications developers and analysts not elsewhere classified

2520 database and network professionals

2521 database designers and administrators

2522 systems administrators

2523 computer network professionals

2529 database and network professionals not elsewhere classified

2600 legal, social and cultural professionals

2610 legal professionals

2611 lawyers

2612 judges

2619 legal professionals not elsewhere classified

2620 librarians, archivists and curators

2621 archivists and curators

2622 librarians and related information professionals

2630 social and religious professionals

2631 economists

2632 sociologists, anthropologists and related professionals

2633 philosophers, historians and political scientists

2634 psychologists

2635 social work and counselling professionals

2636 religious professionals

2640 authors, journalists and linguists

2641 authors and related writers

2642 journalists

2643 translators, interpreters and other linguists

2650 creative and performing artists

2651 visual artists

2652 musicians, singers and composers

2653 dancers and choreographers

2654 film, stage and related directors and producers

2655 actors

2656 announcers on radio, television and other media

2659 creative and performing artists not elsewhere classified

3000 technicians and associate professionals

3100 science and engineering associate professionals

3110 physical and engineering science technicians

3111 chemical and physical science technicians

3112 civil engineering technicians

3113 electrical engineering technicians

3114 electronics engineering technicians

3115 mechanical engineering technicians

3116 chemical engineering technicians

3117 mining and metallurgical technicians

3118 draughtspersons

3119 physical and engineering science technicians not elsewhere classified

3120 mining, manufacturing and construction supervisors

3121 mining supervisors

3122 manufacturing supervisors

3123 construction supervisors

3130 process control technicians

3131 power production plant operators

3132 incinerator and water treatment plant operators

3133 chemical processing plant controllers

3134 petroleum and natural gas refining plant operators

3135 metal production process controllers

3139 process control technicians not elsewhere classified

3140 life science technicians and related associate professionals

3141 life science technicians (excluding medical)

3142 agricultural technicians

3143 forestry technicians

3150 ship and aircraft controllers and technicians

3151 ships' engineers

3152 ships' deck officers and pilots

3153 aircraft pilots and related associate professionals

3154 air traffic controllers

3155 air traffic safety electronics technicians

3200 health associate professionals

3210 medical and pharmaceutical technicians

3211 medical imaging and therapeutic equipment technicians

3212 medical and pathology laboratory technicians

3213 pharmaceutical technicians and assistants

3214 medical and dental prosthetic technicians

3220 nursing and midwifery associate professionals

3221 nursing associate professionals

3222 midwifery associate professionals

3230 traditional and complementary medicine associate professionals

3240 veterinary technicians and assistants

3250 other health associate professionals

3251 dental assistants and therapists

3252 medical records and health information technicians

3253 community health workers

3254 dispensing opticians

3255 physiotherapy technicians and assistants

3256 medical assistants

3257 environmental and occupational health inspectors and associates

3258 ambulance workers

3259 health associate professionals not elsewhere classified

3300 business and administration associate professionals

3310 financial and mathematical associate professionals

3311 securities and finance dealers and brokers

3312 credit and loans officers

3313 accounting associate professionals

3314 statistical, mathematical and related associate professionals

3315 valuers and loss assessors

3320 sales and purchasing agents and brokers

3321 insurance representatives

3322 commercial sales representatives

3323 buyers

3324 trade brokers

3330 business services agents

3331 clearing and forwarding agents

3332 conference and event planners

3333 employment agents and contractors

3334 real estate agents and property managers

3339 business services agents not elsewhere classified

3340 administrative and specialised secretaries

3341 office supervisors

3342 legal secretaries

3343 administrative and executive secretaries

3344 medical secretaries

3350 regulatory government associate professionals

3351 customs and border inspectors

3352 government tax and excise officials

3353 government social benefits officials

3354 government licensing officials

3355 police inspectors and detectives

3359 regulatory government associate professionals not elsewhere classified

3400 legal, social, cultural and related associate professionals

3410 legal, social and religious associate professionals

3411 police inspectors and detectives

3412 social work associate professionals

3413 religious associate professionals

3420 sports and fitness workers

3421 athletes and sports players

3422 sports coaches, instructors and officials

3423 fitness and recreation instructors and program leaders

3430 artistic, cultural and culinary associate professionals

3431 photographers

3432 interior designers and decorators

3433 gallery, museum and library technicians

3434 chefs

3435 other artistic and cultural associate professionals

3500 information and communications technicians

3510 information and communications technology operations and user support technicians

3511 information and communications technology operations technicians

3512 information and communications technology user support technicians

3513 computer network and systems technicians

3514 web technicians

3520 telecommunications and broadcasting technicians

3521 broadcasting and audio-visual technicians

3522 telecommunications engineering technicians

4000 clerical support workers

4100 general and keyboard clerks

4110 general office clerks

4120 secretaries (general)

4130 keyboard operators

4131 typists and word processing operators

4132 data entry clerks

4200 customer services clerks

4210 tellers, money collectors and related clerks

4211 bank tellers and related clerks

4212 bookmakers, croupiers and related gaming workers

4213 pawnbrokers and money-lenders

4214 debt-collectors and related workers

4220 client information workers

4221 travel consultants and clerks

4222 contact centre information clerks

4223 telephone switchboard operators

4224 hotel receptionists

4225 enquiry clerks

4226 receptionists (general)

4227 survey and market research interviewers

4229 client information workers not elsewhere classified

4300 numerical and material recording clerks

4310 numerical clerks

4311 accounting and bookkeeping clerks

4312 statistical, finance and insurance clerks

4313 payroll clerks

4320 material-recording and transport clerks

4321 stock clerks

4322 production clerks

4323 transport clerks

4400 other clerical support workers

4410 other clerical support workers

4411 library clerks

4412 mail carriers and sorting clerks

4413 coding, proof-reading and related clerks

4414 scribes and related workers

4415 filing and copying clerks

4416 personnel clerks

4419 clerical support workers not elsewhere classified

5000 service and sales workers

5100 personal service workers

5110 travel attendants, conductors and guides

5111 travel attendants and travel stewards

5112 transport conductors

5113 travel guides

5120 cooks

5130 waiters and bartenders

5131 waiters

5132 bartenders

5140 hairdressers, beauticians and related workers

5141 hairdressers

5142 beauticians and related workers

5150 building and housekeeping supervisors

5151 cleaning and housekeeping supervisors in offices, hotels and other establishments

5152 domestic housekeepers

5153 building caretakers

5160 other personal services workers

5161 astrologers, fortune-tellers and related workers

5162 companions and valets

5163 undertakers and embalmers

5164 pet groomers and animal care workers

5165 driving instructors

5169 personal services workers not elsewhere classified

5200 sales workers

5210 street and market salespersons

5211 stall and market salespersons

5212 street food salespersons

5220 shop salespersons

5221 shop keepers

5222 shop supervisors

5223 shop sales assistants

5230 cashiers and ticket clerks

5240 other sales workers

5241 fashion and other models

5242 sales demonstrators

5243 door to door salespersons

5244 contact centre salespersons

5245 service station attendants

5246 food service counter attendants

5249 sales workers not elsewhere classified

5300 personal care workers

5310 child care workers and teachers' aides

5311 child care workers

5312 teachers' aides

5320 personal care workers in health services

5321 health care assistants

5322 home-based personal care workers

5329 personal care workers in health services not elsewhere classified

5400 protective services workers

5410 protective services workers

5411 fire-fighters

5412 police officers

5413 prison guards

5414 security guards

5419 protective services workers not elsewhere classified

6000 skilled agricultural, forestry and fishery workers

6100 market-oriented skilled agricultural workers

6110 market gardeners and crop growers

6111 field crop and vegetable growers

6112 tree and shrub crop growers

6113 gardeners, horticultural and nursery growers

6114 mixed crop growers

6120 animal producers

6121 livestock and dairy producers

6122 poultry producers

6123 apiarists and sericulturists

6129 animal producers not elsewhere classified

6130 mixed crop and animal producers

6200 market-oriented skilled forestry, fishery and hunting workers

6210 forestry and related workers

6220 fishery workers, hunters and trappers

6221 aquaculture workers

6222 inland and coastal waters fishery workers

6223 deep-sea fishery workers

6224 hunters and trappers

6300 subsistence farmers, fishers, hunters and gatherers

6310 subsistence crop farmers

6320 subsistence livestock farmers

6330 subsistence mixed crop and livestock farmers

6340 subsistence fishers, hunters, trappers and gatherers

7000 craft and related trades workers

7100 building and related trades workers, excluding electricians

7110 building frame and related trades workers

7111 house builders

7112 bricklayers and related workers

7113 stonemasons, stone cutters, splitters and carvers

7114 concrete placers, concrete finishers and related workers

7115 carpenters and joiners

7119 building frame and related trades workers not elsewhere classified

7120 building finishers and related trades workers

7121 roofers

7122 floor layers and tile setters

7123 plasterers

7124 insulation workers

7125 glaziers

7126 plumbers and pipe fitters

7127 air conditioning and refrigeration mechanics

7130 painters, building structure cleaners and related trades workers

7131 painters and related workers

7132 spray painters and varnishers

7133 building structure cleaners

7200 metal, machinery and related trades workers

7210 sheet and structural metal workers, moulders and welders, and related workers

7211 metal moulders and coremakers

7212 welders and flamecutters

7213 sheet-metal workers

7214 structural-metal preparers and erectors

7215 riggers and cable splicers

7220 blacksmiths, toolmakers and related trades workers

7221 blacksmiths, hammersmiths and forging press workers

7222 toolmakers and related workers

7223 metal working machine tool setters and operators

7224 metal polishers, wheel grinders and tool sharpeners

7230 machinery mechanics and repairers

7231 motor vehicle mechanics and repairers

7232 aircraft engine mechanics and repairers

7233 agricultural and industrial machinery mechanics and repairers

7234 bicycle and related repairers

7300 handicraft and printing workers

7310 handicraft workers

7311 precision-instrument makers and repairers

7312 musical instrument makers and tuners

7313 jewellery and precious-metal workers

7314 potters and related workers

7315 glass makers, cutters, grinders and finishers

7316 sign writers, decorative painters, engravers and etchers

7317 handicraft workers in wood, basketry and related materials

7318 handicraft workers in textile, leather and related materials

7319 handicraft workers not elsewhere classified

7320 printing trades workers

7321 pre-press technicians

7322 printers

7323 print finishing and binding workers

7400 electrical and electronic trades workers

7410 electrical equipment installers and repairers

7411 building and related electricians

7412 electrical mechanics and fitters

7413 electrical line installers and repairers

7420 electronics and telecommunications installers and repairers

7421 electronics mechanics and servicers

7422 information and communications technology installers and servicers

7500 food processing, wood working, garment and other craft and related trades workers

7510 food processing and related trades workers

7511 butchers, fishmongers and related food preparers

7512 bakers, pastry-cooks and confectionery makers

7513 dairy-products makers

7514 fruit, vegetable and related preservers

7515 food and beverage tasters and graders

7516 tobacco preparers and tobacco products makers

7520 wood treaters, cabinet-makers and related trades workers

7521 wood treaters

7522 cabinet-makers and related workers

7523 woodworking-machine tool setters and operators

7530 garment and related trades workers

7531 tailors, dressmakers, furriers and hatters

7532 garment and related pattern-makers and cutters

7533 sewing, embroidery and related workers

7534 upholsterers and related workers

7535 pelt dressers, tanners and fellmongers

7536 shoemakers and related workers

7540 other craft and related workers

7541 underwater divers

7542 shotfirers and blasters

7543 product graders and testers (excluding foods and beverages)

7544 fumigators and other pest and weed controllers

7549 craft and related workers not elsewhere classified

8000 plant and machine operators, and assemblers

8100 stationary plant and machine operators

8110 mining and mineral processing plant operators

8111 miners and quarriers

8112 mineral and stone processing plant operators

8113 well drillers and borers and related workers

8114 cement, stone and other mineral products machine operators

8120 metal processing and finishing plant operators

8121 metal processing plant operators

8122 metal finishing, plating and coating machine operators

8130 chemical and photographic products plant and machine operators

8131 chemical products plant and machine operators

8132 photographic products machine operators

8140 rubber, plastic and paper products machine operators

8141 rubber products machine operators

8142 plastic products machine operators

8143 paper products machine operators

8150 textile, fur and leather products machine operators

8151 fibre preparing, spinning and winding machine operators

8152 weaving and knitting machine operators

8153 sewing machine operators

8154 bleaching, dyeing and fabric cleaning machine operators

8155 fur and leather preparing machine operators

8156 shoemaking and related machine operators

8157 laundry machine operators

8159 textile, fur and leather products machine operators not elsewhere classified

8160 food and related products machine operators

8170 wood processing and papermaking plant operators

8171 pulp and papermaking plant operators

8172 wood processing plant operators

8180 other stationary plant and machine operators

8181 glass and ceramics plant operators

8182 steam engine and boiler operators

8183 packing, bottling and labelling machine operators

8189 stationary plant and machine operators not elsewhere classified

8200 assemblers

8210 assemblers

8211 mechanical machinery assemblers

8212 electrical and electronic equipment assemblers

8219 assemblers not elsewhere classified

8300 drivers and mobile plant operators

8310 locomotive engine drivers and related workers

8311 locomotive engine drivers

8312 railway brake, signal and switch operators

8320 car, van and motorcycle drivers

8321 motorcycle drivers

8322 car, taxi and van drivers

8330 heavy truck and bus drivers

8331 bus and tram drivers

8332 heavy truck and lorry drivers

8340 mobile plant operators

8341 mobile farm and forestry plant operators

8342 earthmoving and related plant operators

8343 crane, hoist and related plant operators

8344 lifting truck operators

8350 ships' deck crews and related workers

9000 elementary occupations

9100 cleaners and helpers

9110 domestic, hotel and office cleaners and helpers

9111 domestic cleaners and helpers

9112 cleaners and helpers in offices, hotels and other establishments

9120 vehicle, window, laundry and other hand cleaning workers

9121 hand launderers and pressers

9122 vehicle cleaners

9123 window cleaners

9129 other cleaning workers

9200 agricultural, forestry and fishery labourers

9210 agricultural, forestry and fishery labourers

9211 crop farm labourers

9212 livestock farm labourers

9213 mixed crop and livestock farm labourers

9214 garden and horticultural labourers

9215 forestry labourers

9216 fishery and aquaculture labourers

9300 labourers in mining, construction, manufacturing and transport

9310 mining and construction labourers

9311 mining and quarrying labourers

9312 civil engineering labourers

9313 building construction labourers

9320 manufacturing labourers

9321 hand packers

9329 manufacturing labourers not elsewhere classified

9330 transport and storage labourers

9331 hand and pedal vehicle drivers

9332 drivers of animal-drawn vehicles and machinery

9333 freight handlers

9334 shelf fillers

9400 food preparation assistants

9410 food preparation assistants

9411 fast food preparers

9412 kitchen helpers

9500 street and related sales and service workers

9510 street and related service workers

9520 street vendors (excluding food)

9600 refuse workers and other elementary workers

9610 refuse workers

9611 garbage and recycling collectors

9612 refuse sorters

9613 sweepers and related labourers

9620 other elementary workers

9621 messengers, package deliverers and luggage porters

9622 odd job persons

9623 meter readers and vending-machine collectors

9624 water and firewood collectors

9629 elementary workers not elsewhere classified

9997 uncodeable

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(q) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook