Non-working activities, hobbies

sathobby

Source

General Social Survey https://gss.norc.org

Details

Question 164. For each area of life I am going to name, tell me the number that shows how much satisfaction you get from that area. b. Your non-working activities -- hobbies and so on.

Overview

For further details see the official GSS documentation.

Counts by year:

year	iap	a fair amount	a little	don't know	great deal	no answer	none	quite a bit	some	very great deal	Total

1972	1613	-	-	-	-	-	-	-	-	-	1613

1973	-	192	87	11	430	6	66	251	78	383	1504

1974	-	196	71	5	467	-	74	239	88	344	1484

1975	-	182	70	9	482	4	61	240	83	359	1490

1976	-	163	74	6	524	4	45	267	84	332	1499

1977	-	206	66	9	447	2	47	292	75	386	1530

1978	-	190	54	6	520	6	54	278	59	365	1532

1980	-	161	60	7	500	1	38	230	72	399	1468

1982	-	276	97	12	536	5	78	348	122	386	1860

1983	-	219	52	13	507	3	57	319	103	326	1599

1984	-	168	45	6	522	9	35	253	65	370	1473

1985	1534	-	-	-	-	-	-	-	-	-	1534

1986	-	191	71	9	497	13	71	274	76	268	1470

1987	-	218	76	13	619	11	47	319	92	424	1819

1988	484	117	35	4	359	2	30	181	56	213	1481

1989	502	110	40	5	368	4	34	178	53	243	1537

1990	473	105	27	5	317	2	32	154	52	205	1372

1991	500	133	39	3	350	2	31	178	58	223	1517

1993	549	125	40	4	345	7	28	193	63	252	1606

1994	2481	58	14	-	175	2	21	93	33	115	2992

1996	2904	-	-	-	-	-	-	-	-	-	2904

1998	2832	-	-	-	-	-	-	-	-	-	2832

2000	2817	-	-	-	-	-	-	-	-	-	2817

2002	2765	-	-	-	-	-	-	-	-	-	2765

2004	2812	-	-	-	-	-	-	-	-	-	2812

2006	4510	-	-	-	-	-	-	-	-	-	4510

2008	2023	-	-	-	-	-	-	-	-	-	2023

2010	2044	-	-	-	-	-	-	-	-	-	2044

2012	1974	-	-	-	-	-	-	-	-	-	1974

2014	2538	-	-	-	-	-	-	-	-	-	2538

2016	2867	-	-	-	-	-	-	-	-	-	2867

2018	2348	-	-	-	-	-	-	-	-	-	2348

2021	4032	-	-	-	-	-	-	-	-	-	4032

2022	3544	-	-	-	-	-	-	-	-	-	3544

Total	48146	3010	1018	127	7965	83	849	4287	1312	5593	72390

Values

1 very great deal

2 great deal

3 quite a bit

4 a fair amount

5 some

6 a little

7 none

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(q) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook