Spend evening with neighbor

socommun

Source

General Social Survey https://gss.norc.org

Details

Question 173. Would you use this card and tell me which answer comes closest to how often you do the following things: b. Spend a social evening with someone who lives in your neighborhood.

Overview

For further details see the official GSS documentation.

Counts by year:

year	iap	about once a month	about once a year	almost daily	don't know	never	no answer	once or twice a week	several times a month	several times a year	skipped on web	Total

1972	1613	-	-	-	-	-	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	-	-	-	-	-	1504

1974	-	249	91	110	6	322	2	347	189	168	-	1484

1975	-	218	117	89	3	350	2	334	196	181	-	1490

1976	1499	-	-	-	-	-	-	-	-	-	-	1499

1977	-	239	123	86	2	370	4	331	184	191	-	1530

1978	-	189	130	92	6	379	4	359	167	206	-	1532

1980	1468	-	-	-	-	-	-	-	-	-	-	1468

1982	-	305	182	122	1	407	8	347	234	254	-	1860

1983	-	230	148	91	3	415	4	316	194	198	-	1599

1984	1473	-	-	-	-	-	-	-	-	-	-	1473

1985	-	211	161	85	3	395	4	276	200	199	-	1534

1986	-	215	136	104	-	378	3	308	165	161	-	1470

1987	1819	-	-	-	-	-	-	-	-	-	-	1819

1988	493	149	69	46	2	279	2	203	110	128	-	1481

1989	531	149	106	51	5	305	2	170	96	122	-	1537

1990	444	134	88	52	5	274	2	165	105	103	-	1372

1991	493	140	78	55	2	311	3	191	119	125	-	1517

1993	526	156	102	49	2	320	2	181	130	138	-	1606

1994	1015	278	188	103	3	606	9	315	227	248	-	2992

1996	944	307	182	101	6	561	2	330	223	248	-	2904

1998	961	292	193	98	6	523	3	299	196	261	-	2832

2000	940	255	197	97	9	573	7	313	206	220	-	2817

2002	1857	108	94	82	3	272	-	133	113	103	-	2765

2004	1906	140	81	59	2	252	-	138	109	125	-	2812

2006	2518	288	210	95	-	588	2	326	208	275	-	4510

2008	694	183	143	83	2	398	-	209	140	171	-	2023

2010	614	198	138	97	3	400	1	254	151	188	-	2044

2012	672	202	127	67	3	428	1	191	148	135	-	1974

2014	863	243	163	82	-	522	-	272	189	204	-	2538

2016	979	295	164	102	1	611	1	281	216	217	-	2867

2018	789	243	145	70	-	510	2	219	174	196	-	2348

2021	1318	322	323	39	-	1076	-	235	272	438	9	4032

2022	1168	303	226	83	2	864	1	320	264	312	1	3544

Total	29101	6241	4105	2290	80	12689	71	7363	4925	5515	10	72390

Values

1 almost daily

2 once or twice a week

3 several times a month

4 about once a month

5 several times a year

6 about once a year

7 never

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(q) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook