Spend evening with relatives

socrel

Source

General Social Survey https://gss.norc.org

Details

Question 173. Would you use this card and tell me which answer comes closest to how often you do the following things: a. Spend a social evening with relatives?

Overview

For further details see the official GSS documentation.

Counts by year:

year	iap	about once a month	about once a year	almost daily	never	no answer	once or twice a week	several times a month	several times a year	don't know	skipped on web	Total

1972	1613	-	-	-	-	-	-	-	-	-	-	1613

1973	1504	-	-	-	-	-	-	-	-	-	-	1504

1974	-	237	116	135	50	2	424	286	234	-	-	1484

1975	-	245	103	107	67	2	474	249	243	-	-	1490

1976	1499	-	-	-	-	-	-	-	-	-	-	1499

1977	-	266	93	128	75	3	442	271	251	1	-	1530

1978	-	220	125	107	57	4	439	289	289	2	-	1532

1980	1468	-	-	-	-	-	-	-	-	-	-	1468

1982	-	300	148	176	82	8	488	316	339	3	-	1860

1983	-	254	116	100	83	3	421	320	300	2	-	1599

1984	1473	-	-	-	-	-	-	-	-	-	-	1473

1985	-	251	104	107	61	4	426	295	282	4	-	1534

1986	-	244	114	143	76	5	392	224	271	1	-	1470

1987	1819	-	-	-	-	-	-	-	-	-	-	1819

1988	493	144	74	94	45	3	263	182	182	1	-	1481

1989	531	162	72	74	58	3	265	164	206	2	-	1537

1990	444	143	65	73	32	2	250	179	183	1	-	1372

1991	493	154	66	100	45	3	258	202	195	1	-	1517

1993	526	211	83	84	40	3	263	209	185	2	-	1606

1994	1015	350	140	155	80	6	507	363	373	3	-	2992

1996	944	338	131	179	93	1	526	323	367	2	-	2904

1998	961	326	139	150	76	4	505	316	352	3	-	2832

2000	940	291	106	180	92	10	492	369	334	3	-	2817

2002	1857	150	56	103	45	1	223	175	153	2	-	2765

2004	1906	148	43	112	29	-	247	167	159	1	-	2812

2006	2518	308	126	206	77	3	554	349	369	-	-	4510

2008	694	201	87	165	63	-	330	263	219	1	-	2023

2010	614	239	92	189	53	2	374	249	229	3	-	2044

2012	672	182	94	156	70	1	306	266	224	3	-	1974

2014	863	260	119	211	79	1	395	308	301	1	-	2538

2016	979	321	122	260	81	2	446	357	297	2	-	2867

2018	789	259	105	217	66	1	382	287	240	2	-	2348

2021	1318	481	275	191	148	1	438	478	694	1	7	4032

2022	1168	377	173	282	157	-	486	433	462	3	3	3544

Total	29101	7062	3087	4184	1980	78	11016	7889	7933	50	10	72390

Values

1 almost daily

2 once or twice a week

3 several times a month

4 about once a month

5 several times a year

6 about once a year

7 never

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(q) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook