wordi

Word i

Source

General Social Survey https://gss.norc.org

Details

Question 1612. We would like to know something about how people go about guessing words they do not know. On this card are listed some words--you may know some of them, and you may not know quite a few of them. On each line the first word is in capital letters like BEAST. Then there are five other words. Tell me the number of the word that comes closest to the meaning of the word in capital letters. For example, if the word in capital letters is BEAST, you would say "4" since "animal" come closer to BEAST than any of the other words. If you wish, I will read the words to you. These words are difficult for almost everyone just give me your best guess if you are not sure of the answer.

Overview

For further details see the official GSS documentation.

Counts by year:

year	iap	correct answer	incorrect answer	no answer	not available in this year	Total

1972	1613	-	-	-	-	1613

1973	1504	-	-	-	-	1504

1974	-	1118	296	70	-	1484

1975	1490	-	-	-	-	1490

1976	-	1158	248	93	-	1499

1977	1530	-	-	-	-	1530

1978	-	1201	252	79	-	1532

1980	1468	-	-	-	-	1468

1982	-	1251	433	176	-	1860

1983	1599	-	-	-	-	1599

1984	-	1087	287	99	-	1473

1985	1534	-	-	-	-	1534

1986	1470	-	-	-	-	1470

1987	-	1228	391	200	-	1819

1988	493	672	224	92	-	1481

1989	531	730	206	70	-	1537

1990	444	642	185	101	-	1372

1991	493	747	187	90	-	1517

1993	526	786	211	83	-	1606

1994	1015	1414	390	173	-	2992

1996	944	1425	401	134	-	2904

1998	1445	1018	244	125	-	2832

2000	1398	969	289	161	-	2817

2002	2765	-	-	-	-	2765

2004	1340	1094	327	51	-	2812

2006	3041	1018	337	114	-	4510

2008	845	848	289	41	-	2023

2010	614	1033	334	63	-	2044

2012	672	953	301	48	-	1974

2014	863	1230	404	41	-	2538

2016	979	1381	464	43	-	2867

2018	789	1115	417	27	-	2348

2021	-	-	-	-	4032	4032

2022	2378	861	261	44	-	3544

Total	33783	24979	7378	2218	4032	72390

Values

0 incorrect answer

1 correct answer

NA(d) don't know

NA(i) iap

NA(j) I don't have a job

NA(m) dk, na, iap

NA(n) no answer

NA(p) not imputable

NA(q) not imputable

NA(r) refused

NA(s) skipped on web

NA(u) uncodeable

NA(x) not available in this release

NA(y) not available in this year

NA(z) see codebook