General Social Survey Survey 2020 Panel Data

A tibble containing the General Social Survey 2020 Panel Data File, in wide format.

Usage

data(gss_panel20)

Format

A tibble with 5,215 rows and 4,296 columns. Variables are encoded as labelled vectors. The GSS Codebook is the authoritative source for the variables in this dataset. It is available at http://gss.norc.org/Get-Documentation. Due to the COVID-19 pandemic, in 2020 the conducted the GSS was conducted as two studies: (1) a panel re-interview of past respondents from the 2016 and 2018 cross sectional GSS studies (referred to as the 2016-2020 GSS Panel), and (2) an independent fresh cross-sectional address-based sampling push to web study (referred to as 2020 cross-sectional survey). This data object is for the first study; namely, the study empaneling former 2016 and 2018 GSS respondents to answer a GSS questionnaire in 2020 (i.e., the 2016-2020 GSS panel).

This data focuses on Wave 2 of the 2016-2020 GSS Panel – i.e. the panel reinterviews with 2018 GSS
respondents and a randomly selected subset of 2016 GSS respondents. The GSS has used a panel format
previously, as parts of the 2006-2014 GSS. In the 2016-2020 GSS Panel, variables only
contain data from one of the three years. To differentiate between versions of each variable,
they have been appended with suffixes. Variables from 2016 (Wave 1a) have _1a
appended, variables from 2018 (Wave 1b) have _1b appended, and variables from 2020 (Wave 2) have _2
appended. Users can also track cases from 2016 and 2018, and reinterviews from 2020 with the variable
`samptype`.

Because of its relatively complex nature, users are strongly encouraged to consult the
official [GSS documentation for this dataset](https://gss.norc.org/Documents/codebook/2016-2020%20GSS%20Panel%20Codebook%20-%20R1a.pdf).

Source

National Opinion Research Center, http://gss.norc.org.